% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sanitize.R
\name{add_ignore}
\alias{add_ignore}
\title{Exclude version control of the drake project cache files.}
\usage{
add_ignore(path = drake::find_project())
}
\arguments{
\item{path}{path to \code{.gitignore}. Typically below the root directory.}
}
\description{
Add drake cache directory (hidden folder named \code{.drake})
to \code{.gitignore}, when if there is not registered.
}
\examples{
\dontrun{
add_ignore(path = '.gitignore')
}
}
