% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manipulate.R
\name{analyses}
\alias{analyses}
\title{Function \code{analyses}}
\usage{
analyses(plan, datasets)
}
\arguments{
\item{plan}{data frame with unevaluated wildcards}

\item{datasets}{dataframe with the drake workflow for datasets}
}
\value{
data frame of instructions for analyses (wildcards unevaluated)
}
\description{
Generate a portion of a \code{drake} workflow for
analyzing multiple datasets with multiple methods.
Use the \code{\link{help_drake}} function to get more help.
}
\details{
Use the \code{\link{help_drake}} function to get more help.
}
\seealso{
\code{\link{help_drake}}, \code{\link{summaries}},
 \code{\link{make}}, \code{\link{plan}}, \code{\link{example_plan}}
}

