% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api-wildcards.R
\name{analysis_wildcard_}
\alias{analysis_wildcard_}
\title{Show the analysis wildcard
used in \code{\link[=plan_summaries]{plan_summaries()}}.}
\usage{
analysis_wildcard_()
}
\value{
The analysis wildcard used in \code{\link[=plan_summaries]{plan_summaries()}}.
}
\description{
Used to generate workflow plan data frames.
}
\note{
When the now-deprecated \code{\link[=analysis_wildcard]{analysis_wildcard()}} is removed,
this function's name should have the final "_" stripped off.
}
\examples{
# See ?plan_analyses for examples
}
\seealso{
\code{\link[=plan_summaries]{plan_summaries()}}
}
