% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/defunct.R
\name{as_drake_filename}
\alias{as_drake_filename}
\title{as_drake_filename \ifelse{html}{\figure{lifecycle-defunct.svg}{options: alt='Defunct lifecycle'}}{\strong{Defunct}}}
\usage{
as_drake_filename(...)
}
\arguments{
\item{...}{Arguments}
}
\description{
2019-02-15
}
\keyword{internal}
