% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api-plan.R
\name{as_drake_lang}
\alias{as_drake_lang}
\title{Internal formatting function.}
\usage{
as_drake_lang(x)
}
\arguments{
\item{x}{S3 \code{drake_lang} character vector.}
}
\description{
Internal.
}
\keyword{internal}
