% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drake_plan_class.R
\name{as_drake_plan}
\alias{as_drake_plan}
\alias{as_drake_plan.data.frame}
\alias{as_drake_plan.list}
\alias{as_drake_plan.tbl_df}
\title{Mark a data frame as a \code{drake} workflow plan}
\usage{
as_drake_plan(x, ...)

\method{as_drake_plan}{data.frame}(x, ...)

\method{as_drake_plan}{list}(x, ...)

\method{as_drake_plan}{tbl_df}(x, ...)
}
\arguments{
\item{x}{object to mark as a \code{drake} plan}

\item{...}{other arguments to the method}
}
\description{
Used for pretty printing only (coming soon).
You do not actually have to mark plans as such.
You can keep them as ordinary data frames.
}
\examples{
plan <- list(target = "x", command = "get_data()")
class(plan)
plan <- as_drake_plan(plan)
class(plan)
}
\keyword{internal}
