% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/defunct.R
\name{as_file}
\alias{as_file}
\title{as_file \ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#defunct}{\figure{lifecycle-defunct.svg}{options: alt='[Defunct]'}}}{\strong{[Defunct]}}}
\usage{
as_file(...)
}
\arguments{
\item{...}{Arguments}
}
\description{
2019-02-15
}
\keyword{internal}
