% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{available_hash_algos}
\alias{available_hash_algos}
\title{List the available hash algorithms for drake caches.
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}}
\usage{
available_hash_algos()
}
\value{
A character vector of names of available hash algorithms.
}
\description{
Deprecated on 2018-12-12.
}
\keyword{internal}
