% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workplan.R
\name{bind_plans}
\alias{bind_plans}
\title{Row-bind together drake plans}
\usage{
bind_plans(...)
}
\arguments{
\item{...}{workflow plan data frames (see \code{\link[=drake_plan]{drake_plan()}})}
}
\description{
combine drake plans together in a way that
correctly fills in missing entries.
}
\examples{
# You might need to refresh your data regularly (see ?triggers).
download_plan <- drake_plan(
  data = target(
    command = download_data(),
    trigger = "always"
  ),
  strings_in_dots = "literals"
)
# But if the data don't change, the analyses don't need to change.
analysis_plan <- drake_plan(
  usage = get_usage_metrics(data),
  topline = scrape_topline_table(data)
)
your_plan <- bind_plans(download_plan, analysis_plan)
your_plan
# make(your_plan) # nolint
}
\seealso{
drake_plan, make
}
