% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blitz.R
\name{blitz_build}
\alias{blitz_build}
\title{Build a target using "blitz" parallelism}
\usage{
blitz_build(target, deps = NULL, config)
}
\arguments{
\item{target}{name of the target}

\item{deps}{named list of dependencies}

\item{config}{internal configuration list}
}
\description{
For internal use only
}
\keyword{internal}
