% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makefile.R
\name{build}
\alias{build}
\title{Internal function \command{build}}
\usage{
build(target, verbose, force_rehash)
}
\arguments{
\item{target}{name of target to make}

\item{verbose}{logical, same as in \command{link{run}()}}

\item{force_rehash}{logical, same as with \command{\link{run}()}}
}
\description{
Builds an individual target 
inside the \command{Makefiles} created by 
\command{\link{run}(..., makefile = TRUE)}.  
Not meant to be called by the user.
}

