% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/staged.R
\name{build_clustermq}
\alias{build_clustermq}
\title{Build a target using the clustermq backend}
\usage{
build_clustermq(target, config)
}
\arguments{
\item{target}{name of the target}

\item{config}{internal configuration list}
}
\description{
For internal use only
}
\keyword{internal}
