% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/staged.R
\name{build_clustermq_staged}
\alias{build_clustermq_staged}
\title{Build a target using the clustermq backend}
\usage{
build_clustermq_staged(target, meta_list, config)
}
\arguments{
\item{target}{name of the target}

\item{meta_list}{list of metadata}

\item{config}{internal configuration list}
}
\description{
For internal use only
}
\keyword{internal}
