% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distributed.R
\name{build_distributed}
\alias{build_distributed}
\title{Internal function build_distributed}
\usage{
build_distributed(target, meta_list, cache_path)
}
\arguments{
\item{target}{name of the tartet to build}

\item{meta_list}{list of hash metadata
output from \code{meta_list()}.}

\item{config}{Internal configuration list from
\code{\link{drake_config}()}.}
}
\value{
Nothing.
}
\description{
Build a target for a distributed parallel
backend such as "Makefile" or "future".
}
\details{
This function is not meant to be called by the user.
It is only exposed so that \code{future::future()} works
as expected for \code{"future"} parallelism.
}
\examples{
# This is not actually a user-side function.
# See the body of mk() to see how build_distributed
# is used in "Makefile" parallelism.
}
\keyword{internal}
