% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph.R
\name{build_graph}
\alias{build_graph}
\title{Function \code{plot_graph}}
\usage{
build_graph(plan, targets = plan$target, envir = parent.frame())
}
\arguments{
\item{plan}{workflow plan data frame, same as for function
\code{\link{run}()}.}

\item{targets}{names of targets to bulid, same as for function
\code{\link{run}()}.}

\item{envir}{environment to import from, same as for function
\code{\link{run}()}.}
}
\description{
Make a graph of the dependency structure of your workflow.
}
\details{
This function returns an igraph object representing how
the targets in your workflow depend on each other. 
(\code{help(package = "igraph")}). To plot the graph, call
to \code{\link{plot.igraph}()} on your graph, or just use 
\code{\link{plot_graph}()} from the start.
}

