% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api-plan.R
\name{c.drake_lang}
\alias{c.drake_lang}
\title{Internal formatting function.}
\usage{
\method{c}{drake_lang}(x, ...)
}
\arguments{
\item{x}{S3 \code{drake_lang} character vector.}

\item{...}{Other args}
}
\description{
Internal.
}
\keyword{internal}
