% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api-plan.R
\name{c.langs}
\alias{c.langs}
\title{Internal formatting function.}
\usage{
\method{c}{langs}(x, ...)
}
\arguments{
\item{x}{S3 \code{langs} character vector.}

\item{...}{Other args}
}
\description{
Internal.
}
\keyword{internal}
