% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-cache.R
\name{cache_namespaces_}
\alias{cache_namespaces_}
\title{List all the \code{storr} cache namespaces used by drake.}
\usage{

  cache_namespaces_(default = storr::storr_environment()$default_namespace)
}
\arguments{
\item{default}{name of the default \code{storr} namespace}
}
\value{
A character vector of \code{storr} namespaces used for drake.
}
\description{
Ordinary users do not need to worry about this function.
It is just another window into \code{drake}'s internals.
}
\note{
When the now-deprecated \code{\link[=cache_namespaces]{cache_namespaces()}} is removed,
this function's name should have the final "_" stripped off.
}
\examples{
cache_namespaces_()
}
\seealso{
\code{\link[=make]{make()}}
}
