% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache_types.R
\name{cache_types}
\alias{cache_types}
\title{Function cache_types}
\usage{
cache_types()
}
\description{
List the supported drake cache types.
}
\examples{
cache_types()
}
\seealso{
\code{\link{new_cache}}, \code{\link{get_cache}},
\code{\link{default_cache_type}}
}
