% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache.R
\name{cached_planned}
\alias{cached_planned}
\title{List targets in both the plan and the cache.
\lifecycle{maturing}}
\usage{
cached_planned(
  plan,
  path = NULL,
  cache = drake::drake_cache(path = path),
  namespace = NULL,
  jobs = 1
)
}
\arguments{
\item{plan}{A drake plan.}

\item{path}{Path to a \code{drake} cache
(usually a hidden \verb{.drake/} folder) or \code{NULL}.}

\item{cache}{drake cache. See \code{\link[=new_cache]{new_cache()}}.
If supplied, \code{path} is ignored.}

\item{namespace}{Character scalar, name of the storr namespace
to use for listing objects.}

\item{jobs}{Number of jobs/workers for parallel processing.}
}
\value{
A character vector of target and sub-target names.
}
\description{
Includes dynamic sub-targets as well.
See examples for details.
}
\seealso{
\code{\link[=cached]{cached()}}, \link{cached_unplanned}
}
