% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make.R
\name{check}
\alias{check}
\title{Function \code{check}}
\usage{
check(plan, output = plan$output, envir = parent.frame())
}
\arguments{
\item{plan}{workflow plan}

\item{output}{output object(s) to make}

\item{envir}{environment containing user-defined functions}
}
\description{
Check a workflow plan and outputs
and return an error if there are problems such as
conflicting objects or circular dependencies.
}
\details{
Run \code{\link{help_drake}} to see helpful links.
}
\seealso{
\code{\link{help_drake}}, \code{\link{plan}},
\code{\link{make}}, \code{\link{cached}}, \code{\link{readd}}
}

