% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{cleaned_namespaces}
\alias{cleaned_namespaces}
\title{Auxiliary storr namespaces
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}}
\usage{
cleaned_namespaces(default = storr::storr_environment()$default_namespace)
}
\arguments{
\item{default}{Name of the default \code{storr} namespace.}
}
\value{
A character vector of \code{storr} namespaces
that are cleaned during \code{\link[=clean]{clean()}}.
}
\description{
2019-02-13
}
\keyword{internal}
