% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backend_clustermq.R
\name{cmq_build}
\alias{cmq_build}
\title{Build a target using the clustermq backend
\lifecycle{stable}}
\usage{
cmq_build(target, meta, deps, layout, config)
}
\arguments{
\item{target}{Target name.}

\item{meta}{List of metadata.}

\item{deps}{Named list of target dependencies.}

\item{layout}{Internal, part of the full \code{config$layout}.}

\item{config}{A \code{\link[=drake_config]{drake_config()}} list.}
}
\description{
For internal use only
}
\keyword{internal}
