% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/staged.R
\name{cmq_staged_build}
\alias{cmq_staged_build}
\title{Build a target using the clustermq backend}
\usage{
cmq_staged_build(target, meta_list, config)
}
\arguments{
\item{target}{name of the target}

\item{meta_list}{list of metadata}

\item{config}{internal configuration list}
}
\description{
For internal use only
}
\keyword{internal}
