% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/examples.R
\name{debug_cleanup}
\alias{debug_cleanup}
\title{Function \code{debug_cleanup}}
\usage{
debug_cleanup()
}
\description{
For the \code{"debug"} example
(\code{example_plan('debug')}),
remove the files and cache
}
\seealso{
\code{\link{example_plan}}, \code{\link{make}},
\code{\link{help_drake}}
}

