% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/examples.R
\name{debug_setup}
\alias{debug_setup}
\title{Function \code{debug_setup}}
\usage{
debug_setup()
}
\description{
For the \code{"debug"} example 
(\code{example_plan('debug')}),
write the required input files and load the required functions
into the calling environment.
}
\seealso{
\code{\link{example_plan}}, \code{\link{make}},
\code{\link{help_drake}}
}

