% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache.R
\name{default_cache_path}
\alias{default_cache_path}
\title{Return the default file path of the drake/storr cache.}
\usage{
default_cache_path()
}
\value{
Default file path of the drake/storr cache.
}
\description{
Applies to file system caches only.
}
\examples{
default_cache_path()
}
