% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache_types.R
\name{default_cache_type}
\alias{default_cache_type}
\title{Function default_cache_type}
\usage{
default_cache_type()
}
\description{
Name the default type of drake cache.
}
\examples{
default_cache_type()
}
\seealso{
\code{\link{new_cache}}, \code{\link{get_cache}},
\code{\link{default_cache_type}}
}
