% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drake_graph_info.R
\name{default_graph_title}
\alias{default_graph_title}
\title{Return the default title for graph visualizations
\ifelse{html}{\figure{lifecycle-soft-deprecated.svg}{options: alt='Soft-deprecated lifecycle'}}{\strong{Soft-deprecated}}}
\usage{
default_graph_title()
}
\value{
A character scalar with the default graph title.
}
\description{
For internal use only.
}
\examples{
default_graph_title()
}
\keyword{internal}
