% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hasty.R
\name{default_hasty_build}
\alias{default_hasty_build}
\title{Build a target using "hasty" parallelism}
\usage{
default_hasty_build(target, config)
}
\arguments{
\item{target}{name of the target}

\item{config}{internal configuration list}
}
\description{
For internal use only
}
\keyword{internal}
