% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecate.R
\name{default_hook}
\alias{default_hook}
\title{Deprecated.}
\usage{
default_hook(code)
}
\arguments{
\item{code}{Placeholder for the code to build a target/import.}
}
\value{
Evaluated code.
}
\description{
Deprecated on 2018-10-24.
}
\examples{
default_hook(NULL)
}
