% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-deprecate.R
\name{default_Makefile_args}
\alias{default_Makefile_args}
\title{Deprecated}
\usage{
default_Makefile_args(jobs, verbose)
}
\arguments{
\item{jobs}{Number of jobs.}

\item{verbose}{Integer, control printing to the console/terminal.
\itemize{
\item \code{0}: print nothing.
\item \code{1}: print targets, retries, and failures.
\item \code{2}: also show a spinner when preprocessing tasks are underway.
}}
}
\value{
\code{args} for \code{system2(command, args)}
}
\description{
2019-01-03
}
\keyword{internal}
