% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{default_Makefile_command}
\alias{default_Makefile_command}
\title{Default Makefile command
\ifelse{html}{\figure{lifecycle-deprecated.svg}{options: alt='Deprecated lifecycle'}}{\strong{Deprecated}}}
\usage{
default_Makefile_command()
}
\value{
A character scalar
}
\description{
2019-01-03
}
\keyword{internal}
