% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/defunct.R
\name{default_system2_args}
\alias{default_system2_args}
\title{default_system2_args \ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#defunct}{\figure{lifecycle-defunct.svg}{options: alt='[Defunct]'}}}{\strong{[Defunct]}}}
\usage{
default_system2_args(...)
}
\arguments{
\item{...}{Arguments}
}
\description{
2019-02-15
}
\keyword{internal}
