% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/triggers.R
\name{default_trigger}
\alias{default_trigger}
\title{Function default_trigger}
\usage{
default_trigger()
}
\value{
A character scalar naming the default trigger.
}
\description{
Return the default trigger.
}
\examples{
default_trigger()
# See ?triggers for more examples.
}
\seealso{
\code{\link{triggers}}, \code{\link{make}}
}
