% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{default_verbose}
\alias{default_verbose}
\title{Default verbosity
\ifelse{html}{\figure{lifecycle-deprecated.svg}{options: alt='Deprecated lifecycle'}}{\strong{Deprecated}}}
\usage{
default_verbose()
}
\value{
1
}
\description{
Deprecated on 2019-01-01
}
\keyword{internal}
