% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{dependency_profile}
\alias{dependency_profile}
\title{States of the dependencies of a target
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}}
\usage{
dependency_profile(target, config, character_only = FALSE)
}
\arguments{
\item{target}{Name of the target.}

\item{config}{Deprecated.}

\item{character_only}{Logical, whether to assume \code{target}
is a character string rather than a symbol.}
}
\value{
A data frame of the old hashes and
new hashes of the data frame, along with
an indication of which hashes changed since
the last \code{\link[=make]{make()}}.
}
\description{
Deprecated on 2019-02-14.
}
\keyword{internal}
