% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/defunct.R
\name{deprecate_wildcard}
\alias{deprecate_wildcard}
\title{deprecate_wildcard \lifecycle{defunct}}
\usage{
deprecate_wildcard(...)
}
\arguments{
\item{...}{Arguments}
}
\description{
2019-02-15
}
\keyword{internal}
