% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/defunct.R
\name{deps}
\alias{deps}
\title{deps \ifelse{html}{\figure{lifecycle-defunct.svg}{options: alt='Defunct lifecycle'}}{\strong{Defunct}}}
\usage{
deps(...)
}
\arguments{
\item{...}{Arguments}
}
\description{
2019-05-16
}
\keyword{internal}
