% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knitr.R
\name{deps_in_document}
\alias{deps_in_document}
\title{Function deps_in_document}
\usage{
deps_in_document(target)
}
\arguments{
\item{target}{file path to the file or name of the file target,
source text of the document.}
}
\description{
Find the dependencies of a dynamic report. To
enable drake to watch for these dependencies, your command
to compile this report must be use \code{\link{dknit}}:
that is, it must look something like \code{dknit('your_report.Rmd')}
in your workflow plan data frame.
}
\details{
Drake looks for dependencies in the document by
analyzing evaluated code chunks for other targets/imports
mentioned in \code{\link{loadd}()} and \code{\link{readd}()}.
}
\examples{
\dontrun{
load_basic_example()
deps_in_document("'report.Rmd'")
deps_in_document("report.md")
}
}
\seealso{
\code{\link{deps_in_document}},
\code{\link{deps}}, \code{\link{deps_in_document}},
\code{\link{make}}, \code{\link{load_basic_example}}
}
