% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deps.R
\name{deps_profile_impl}
\alias{deps_profile_impl}
\title{Internal function with a drake_config() argument}
\usage{
deps_profile_impl(target, config, character_only = FALSE)
}
\arguments{
\item{target}{Name of a target.}

\item{config}{A \code{\link[=drake_config]{drake_config()}} object.}

\item{character_only}{Logical, whether to interpret
\code{target} as a character (\code{TRUE}) or a symbol (\code{FALSE}).}
}
\description{
Not a user-side function.
}
\keyword{internal}
