% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deps.R
\name{deps_target_impl}
\alias{deps_target_impl}
\title{Internal function with a drake_config() argument}
\usage{
deps_target_impl(target, config, character_only = FALSE)
}
\arguments{
\item{config}{A \code{\link[=drake_config]{drake_config()}} object.}
}
\description{
Not a user-side function.
}
\keyword{internal}
