% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knitr.R
\name{dknit}
\alias{dknit}
\title{Function dknit}
\usage{
dknit(input, ...)
}
\arguments{
\item{input}{charadcter scalar, name of an imported dynamic document
source file. The \code{input} is passed as the \code{input}
argument to \code{\link{knitr::knit}}, but it must be
enclosed in single quotes just like any other file target/import
in drake. The file must also already exist beforehand, so it
must be an imported file, not a built target.}
}
\description{
In a drake workflow, use the command
\code{dknit("'your_document.Rmd'")} to knit the file
\code{your_document.Rmd} while automatically accounting for the
dependencies in \code{your_document.Rmd}. In the document,
for any targerts/imports specifically invoked in calls to
\code{\link{readd()}} and \code{\link{loadd()}} inside
evaluated code chunks, drake will watch these items for changes.
When any one of them changes, \code{\link{drake::make}()} will
recompile your document.
}
\examples{
\dontrun{
load_basic_example
print(my_plan) # Notice: dknit('report.Rmd') makes 'report.md'.
plot_graph(my_plan) # 'report.md' depends on mentions in 'report.Rmd'.
make(my_plan) # First build.
make(my_plan) # Everything up to date.
reg2 <- function(d){
  d$x3 <- d$x ^ 3
  lm(y ~ x3, data = d)
}
make(my_plan) # 'report.md' should rebuild.
}
}
\seealso{
\code{\link{deps_in_document}},
\code{\link{deps}}, \code{\link{deps_of_document}},
\code{\link{make}}, \code{\link{load_basic_example}}
}
