% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/defunct.R
\name{doc_of_function_call}
\alias{doc_of_function_call}
\title{doc_of_function_call \ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#defunct}{\figure{lifecycle-defunct.svg}{options: alt='[Defunct]'}}}{\strong{[Defunct]}}}
\usage{
doc_of_function_call(...)
}
\arguments{
\item{...}{Arguments}
}
\description{
2019-02-15
}
\keyword{internal}
