% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{drake_batchtools_tmpl_file}
\alias{drake_batchtools_tmpl_file}
\title{Get a template file for execution on a cluster.
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}}
\usage{
drake_batchtools_tmpl_file(
  example = drake::drake_hpc_template_files(),
  to = getwd(),
  overwrite = FALSE
)
}
\arguments{
\item{example}{Name of template file.}

\item{to}{Character vector, where to write the file.}

\item{overwrite}{Logical, whether to overwrite an existing file of the
same name.}
}
\description{
Deprecated. Use \code{\link[=drake_hpc_template_file]{drake_hpc_template_file()}} instead.
}
\details{
Deprecated on 2018-06-27.
}
\keyword{internal}
