% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3_drake_deps.R
\name{drake_deps}
\alias{drake_deps}
\title{\code{drake_deps} helper}
\usage{
drake_deps(expr, exclude = character(0), restrict = NULL)
}
\arguments{
\item{expr}{An R expression}

\item{exclude}{Character vector of the names of symbols to exclude
from the code analysis.}

\item{restrict}{Optional character vector of allowable names of globals.
If \code{NULL}, all global symbols are detectable. If a character vector,
only the variables in \code{restrict} will count as global variables.}
}
\value{
A \code{drake_deps} object.
}
\description{
Static code analysis.
}
\examples{
if (FALSE) { # stronger than roxygen dontrun
expr <- quote({
  a <- base::list(1)
  b <- seq_len(10)
  file_out("abc")
  file_in("xyz")
  x <- "123"
  loadd(abc)
  readd(xyz)
})
drake_deps(expr)
}
}
\keyword{internal}
