% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/future.R
\name{drake_future_task}
\alias{drake_future_task}
\title{Task passed to individual futures in the \code{"future"} backend}
\usage{
drake_future_task(target, meta, config, protect)
}
\arguments{
\item{target}{name of the target}

\item{meta}{list of metadata}

\item{config}{\code{\link[=drake_config]{drake_config()}} list}

\item{protect}{Names of targets that still need their
dependencies available in \code{config$envir}.}
}
\value{
Either the target value or a list of build results.
}
\description{
For internal use only. Only exported to make available
to futures.
}
\keyword{internal}
