% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drake_ggraph.R
\name{drake_ggraph_impl}
\alias{drake_ggraph_impl}
\title{Internal function with a drake_config() argument}
\usage{
drake_ggraph_impl(
  config,
  build_times = "build",
  digits = 3,
  targets_only = FALSE,
  main = NULL,
  from = NULL,
  mode = c("out", "in", "all"),
  order = NULL,
  subset = NULL,
  make_imports = TRUE,
  from_scratch = FALSE,
  full_legend = FALSE,
  group = NULL,
  clusters = NULL,
  show_output_files = TRUE,
  label_nodes = FALSE,
  transparency = TRUE
)
}
\arguments{
\item{config}{A \code{\link[=drake_config]{drake_config()}} object.}

\item{make_imports}{Logical, whether to make the imports first.
Set to \code{FALSE} to save some time and risk obsolete output.}
}
\description{
Not a user-side function.
}
\keyword{internal}
