% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api-history.R
\name{drake_history}
\alias{drake_history}
\title{Show history}
\usage{
drake_history(cache = NULL, history = NULL, show_args = FALSE,
  make_imports = TRUE, do_prework = TRUE, verbose = TRUE)
}
\arguments{
\item{cache}{An optional
\code{drake}/\href{https://github.com/richfitz/storr}{storr} cache.
If not supplied, \code{drake} will try to find the default cache:
a folder called \code{.drake/} in the current working directory
or one of its ancestor directories.}

\item{history}{An optional \href{https://github.com/wlandau/txtq}{txtq}
produced by \code{make(history = TRUE)}. If not supplied,
\code{drake} will try to find it next to the cache in a folder
called \code{.drake_history/}.}

\item{show_args}{Logical, whether to search for named atomic arguments
to function calls inside \code{\link[=drake_plan]{drake_plan()}} commands.
Could be slow because this requires parsing and analyzing
lots of R code.}

\item{make_imports}{Logical, whether to make the imports first.
Set to \code{FALSE} to save some time and risk obsolete output.}

\item{do_prework}{Whether to do the \code{prework}
normally supplied to \code{\link[=make]{make()}}.}

\item{verbose}{Integer, control printing to the console/terminal.
\itemize{
\item \code{0}: print nothing.
\item \code{1}: print targets, retries, and failures.
\item \code{2}: also show a spinner when preprocessing tasks are underway.
}}
}
\value{
A data frame of target history.
}
\description{
See the history of your targets:
what you ran, when you ran it, and where
it lives. Optionally show the settings
under which you ran things (could be slow)
and whether each target is up to date (also could be slow).
}
\examples{
\dontrun{
isolate_example({
load_mtcars_example(
make(my_plan, history = TRUE)
reg2 <- function(d) {
  d$x2 <- d$x ^ 3
  lm(y ~ x2, data = d)
}
Sys.sleep(5)
make(my_plan, history = TRUE)
drake_history(show_args = TRUE)
})
}
}
