% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{drake_quotes}
\alias{drake_quotes}
\title{Put quotes around each element of a character vector.
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}}
\usage{
drake_quotes(x = NULL, single = FALSE)
}
\arguments{
\item{x}{Character vector or object to be coerced to character.}

\item{single}{Add single quotes if \code{TRUE}
and double quotes otherwise.}
}
\value{
Character vector with quotes around it.
}
\description{
Deprecated on 2019-01-01
}
\keyword{internal}
