% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{drake_tip}
\alias{drake_tip}
\title{Output a random tip about drake.
\lifecycle{deprecated}}
\usage{
drake_tip()
}
\value{
A character scalar with a tip on how to use drake.
}
\description{
Deprecated on 2019-01-12.
}
\details{
Tips are usually related to news and usage.
}
\keyword{internal}
