% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{drake_unquote}
\alias{drake_unquote}
\title{Remove leading and trailing
escaped quotes from character strings.
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}}
\usage{
drake_unquote(x = NULL)
}
\arguments{
\item{x}{Character vector.}
}
\value{
Character vector without leading
or trailing escaped quotes around
the elements.
}
\description{
Deprecated on 2019-01-01
}
\keyword{internal}
