% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drake_deps.R
\name{drake_validate.drake_deps}
\alias{drake_validate.drake_deps}
\title{\code{drake_deps} validator}
\usage{
\method{drake_validate}{drake_deps}(x)
}
\arguments{
\item{x}{A \code{drake_deps} object.}
}
\description{
\code{drake_deps} validator.
}
\examples{
drake_validate(new_drake_deps())
}
\keyword{internal}
