% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api-read.R
\name{eager_load_target}
\alias{eager_load_target}
\title{Load a target right away (internal function)}
\usage{
eager_load_target(target, cache, namespace, envir, verbose)
}
\arguments{
\item{target}{If \code{character_only} is \code{TRUE}, then
\code{target} is a character string naming the object to read.
Otherwise, \code{target} is an unquoted symbol with the name of the
object.}

\item{cache}{drake cache. See \code{\link[=new_cache]{new_cache()}}.
If supplied, \code{path} and \code{search} are ignored.}

\item{namespace}{Optional character string,
name of the \code{storr} namespace to read from.}

\item{envir}{Environment to load objects into. Defaults to the
calling environment (current workspace).}

\item{verbose}{Integer, control printing to the console/terminal.
\itemize{
\item \code{0}: print nothing.
\item \code{1}: print targets, retries, and failures.
\item \code{2}: also show a spinner when preprocessing tasks are underway.
}}
}
\description{
This function is only exported
to make active bindings work safely.
It is not actually a user-side function.
}
\keyword{internal}
