% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/examples.R
\name{example_plan}
\alias{example_plan}
\title{Function \code{example_plan}}
\usage{
example_plan(x = drake::example_plans())
}
\arguments{
\item{x}{name of workflow plan (character scalar)}
}
\value{
example workflow plan for \code{\link{make}}
}
\description{
Return an example workflow plan for \code{\link{make}}.
}
\details{
The \code{"debug"} example (\code{x <- example_plan("debug")})
requires you to run \code{\link{debug_setup}()} before 
running \code{\link{make}(x)}. That way, the required inputs
are generated.
Run \code{\link{help_drake}} to see helpful links.
}
\seealso{
\code{\link{example_plans}}, \code{\link{make}}, 
\code{\link{help_drake}}
}

