% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/examples.R
\name{example_plans}
\alias{example_plans}
\title{Function \code{example_plans}}
\usage{
example_plans()
}
\value{
names of example drake workflow plans
}
\description{
List the names of example drake workflow plans.
}
\details{
Run \code{\link{help_drake}} to see helpful links.
}
\seealso{
\code{\link{example_plan}}, \code{\link{make}}, 
\code{\link{help_drake}}
}

