% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/examples.R
\name{examples_drake}
\alias{examples_drake}
\title{Function \code{examples_drake}}
\usage{
examples_drake()
}
\value{
names of all the drake examples.
}
\description{
List the names of all the drake examples.
The \code{"basic"} example is the one from the
quickstart vignette: \code{vignette("quickstart")}.
}
\examples{
examples_drake()
}
\seealso{
\code{\link{example_drake}}, \code{\link{make}}
}
