% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/defunct.R
\name{expand}
\alias{expand}
\title{expand \ifelse{html}{\figure{lifecycle-defunct.svg}{options: alt='Defunct lifecycle'}}{\strong{Defunct}}}
\usage{
expand(...)
}
\arguments{
\item{...}{Arguments}
}
\description{
2019-02-15
}
\keyword{internal}
