% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{expand_plan}
\alias{expand_plan}
\title{Deprecated: create replicates of targets.
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}}
\usage{
expand_plan(plan, values = NULL, rename = TRUE, sep = "_", sanitize = TRUE)
}
\arguments{
\item{plan}{Workflow plan data frame.}

\item{values}{Values to expand over. These will be appended to
the names of the new targets.}

\item{rename}{Logical, whether to rename the targets
based on the \code{values}. See the examples for a demo.}

\item{sep}{Character scalar, delimiter between the original
target names and the values to append to create the new
target names. Only relevant when \code{rename} is \code{TRUE}.}

\item{sanitize}{Logical, whether to sanitize the plan.}
}
\value{
An expanded workflow plan data frame (with replicated targets).
}
\description{
Deprecated on 2019-05-16. Use \code{\link[=drake_plan]{drake_plan()}}
transformations instead. See
\verb{https://books.ropensci.org/drake/plans.html#large-plans}
for the details.
}
\details{
Duplicates the rows of a workflow plan data frame.
Prefixes are appended to the new target names
so targets still have unique names.
}
\seealso{
\code{\link[=drake_plan]{drake_plan()}}
}
\keyword{internal}
