% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dependencies.R
\name{file_input}
\alias{file_input}
\title{Declare the file inputs of a workflow plan command.}
\usage{
file_input(...)
}
\arguments{
\item{...}{Symbols or character vectors denoting file inputs.}
}
\value{
A character vector of declared files.
}
\description{
Use this function to help write the commands
in your workflow plan data frame. Treat the output as a character vector.
}
\examples{
file_input(my_file.txt)
file_input(data.csv, 'spreadsheet.xlsx', "serialized.rds")
}
\seealso{
file_output knitr_input
}
