% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-deprecate.R
\name{find_project}
\alias{find_project}
\title{Deprecated. Search up the file system
for the nearest root path of a drake project.}
\usage{
find_project(path = getwd())
}
\arguments{
\item{path}{Starting path for search back for the project.
Should be a subdirectory of the drake project.}
}
\value{
File path of the nearest drake project or \code{NULL}
if no drake project is found.
}
\description{
Deprecated on 2019-01-08.
}
\details{
Only works if the cache is a file system
in a folder named \code{.drake} (default).
}
\seealso{
\code{\link[=drake_plan]{drake_plan()}}, \code{\link[=make]{make()}}
}
