% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/future_lapply.R
\name{fl_master}
\alias{fl_master}
\title{Run the master process of the "future_lapply" backend.}
\usage{
fl_master(cache_path)
}
\arguments{
\item{cache_path}{path to the drake cache}
}
\value{
\code{NULL}
}
\description{
Users should not need to call this function directly.
}
\examples{
# No examples here. This function is not for end users.
}
\keyword{internal}
