% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-defunct.R
\name{from_plan}
\alias{from_plan}
\title{Defunct function.}
\usage{
from_plan(column)
}
\arguments{
\item{column}{Character, name of a column in your \code{drake} plan.}
}
\description{
The \code{from_plan()} function is now defunct
in order to reduce the demands on memory usage.
}
\details{
2019-03-28
}
\seealso{
\code{\link[=drake_envir]{drake_envir()}}
}
\keyword{internal}
