% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backend_future.R
\name{future_build}
\alias{future_build}
\title{Task passed to individual futures in the \code{"future"} backend
\lifecycle{stable}}
\usage{
future_build(target, meta, config, spec, ht_is_subtarget, protect)
}
\arguments{
\item{target}{Name of the target.}

\item{meta}{A list of metadata.}

\item{config}{A \code{\link[=drake_config]{drake_config()}} list.}

\item{ht_is_subtarget}{Internal, part of \code{config}.}

\item{protect}{Names of targets that still need their
dependencies available in memory.}
}
\value{
Either the target value or a list of build results.
}
\description{
For internal use only. Only exported to make available
to futures.
}
\keyword{internal}
