% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manipulate.R
\name{gather}
\alias{gather}
\title{Function \code{gather}}
\usage{
gather(x = NULL, output = "output", gather = "list")
}
\arguments{
\item{x}{argument data frame}

\item{output}{name of aggregated output object}

\item{gather}{function used to gather the outputs}
}
\value{
data frame with a code to gather the outputs in \code{x}
}
\description{
Aggregate/gather the outputs of a previous set of drake instructions.
Use the \code{\link{help_drake}} function to get more help.
}
\details{
Use the \code{\link{help_drake}} function to get more help.
}
\seealso{
\code{\link{help_drake}}
}

