% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{gather_plan}
\alias{gather_plan}
\title{Combine targets
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}}
\usage{
gather_plan(plan = NULL, target = "target", gather = "list", append = FALSE)
}
\arguments{
\item{plan}{Workflow plan data frame of prespecified targets.}

\item{target}{Name of the new aggregated target.}

\item{gather}{Function used to gather the targets. Should be
one of \code{list(...)}, \code{c(...)}, \code{rbind(...)}, or similar.}

\item{append}{Logical. If \code{TRUE}, the output will include the
original rows in the \code{plan} argument.
If \code{FALSE}, the output will only include the new
targets and commands.}
}
\value{
A workflow plan data frame that aggregates multiple
prespecified targets into one additional target downstream.
}
\description{
Deprecated on 2019-05-16. Use \code{\link[=drake_plan]{drake_plan()}}
transformations instead. See
\verb{https://books.ropensci.org/drake/plans.html#large-plans}
for the details.
}
\details{
Creates a new workflow plan to aggregate
existing targets in the supplied plan.
}
\seealso{
\code{\link[=drake_plan]{drake_plan()}}
}
\keyword{internal}
