% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{get_config}
\alias{get_config}
\title{Function get_config}
\usage{
get_config(plan, targets = drake::possible_targets(plan),
  envir = parent.frame(), verbose = TRUE,
  parallelism = drake::default_parallelism(), jobs = 1,
  packages = (.packages()), prework = character(0))
}
\arguments{
\item{plan}{same as for \code{\link{make}}}

\item{targets}{same as for \code{\link{make}}}

\item{envir}{same as for \code{\link{make}}}

\item{verbose}{same as for \code{\link{make}}}

\item{parallelism}{same as for \code{\link{make}}}

\item{jobs}{same as for \code{\link{make}}}

\item{packages}{same as for \code{\link{make}}}

\item{prework}{same as for \code{\link{make}}}
}
\description{
Compute the internal runtime parameter list of 
\code{\link{make}()}. This could save time if you are planning
multiple function calls of functions like \code{\link{outdated}()}
or \code{\link{plot_graph}()}.
}
\examples{
\dontrun{
load_basic_example()
config = get_config(my_plan)
outdated(my_plan, config = config)
max_useful_jobs(my_plan, config = config)
plot_graph(my_plan, config = config)
dataframes_graph(my_plan, config = config)
}
}
\seealso{
\code{\link{plan}}, \code{\link{make}}, \code{\link{plot_graph}}
}
