% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph.R
\name{graph}
\alias{graph}
\title{Function \command{graph}}
\usage{
graph(plan, targets = plan$target, envir = parent.frame())
}
\arguments{
\item{plan}{workflow plan data frame}

\item{targets}{targets to bulid}

\item{envir}{environment to use}
}
\description{
Graph the dependency structure of your workflow
}

