% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/help.R
\name{help_drake}
\alias{help_drake}
\title{Function \code{help_drake}}
\usage{
help_drake()
}
\value{
useful links
}
\description{
Prints links for tutorials, troubleshooting, bug reports, etc.
}
\seealso{
\code{\link{plan}}, \code{\link{make}}
}

