% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{imported}
\alias{imported}
\title{List all the imports in the drake cache.
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}}
\usage{
imported(
  files_only = FALSE,
  path = getwd(),
  search = TRUE,
  cache = drake::get_cache(path = path, search = search, verbose = verbose),
  verbose = 1L,
  jobs = 1
)
}
\arguments{
\item{files_only}{Logical, whether to show imported files only
and ignore imported objects. Since all your functions and
all their global variables are imported, the full list of
imported objects could get really cumbersome.}

\item{path}{Path to a \code{drake} cache
(usually a hidden \verb{.drake/} folder) or \code{NULL}.}

\item{search}{Deprecated.}

\item{cache}{drake cache. See \code{\link[=new_cache]{new_cache()}}.
If supplied, \code{path} is ignored.}

\item{verbose}{Deprecated on 2019-09-11.}

\item{jobs}{Number of jobs/workers for parallel processing.}
}
\value{
Character vector naming the imports in the cache.
}
\description{
Deprecated on 2019-01-08.
}
\details{
An import is a non-target object processed
by \code{\link[=make]{make()}}. Targets in the workflow
plan data frame (see \code{\link[=drake_config]{drake_config()}}
may depend on imports.
}
\seealso{
\code{\link[=cached]{cached()}}, \code{\link[=loadd]{loadd()}}
}
\keyword{internal}
