% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build.R
\name{just_build}
\alias{just_build}
\title{Build/process a single target or import.}
\usage{
just_build(target, meta, config)
}
\arguments{
\item{target}{name of the target}

\item{meta}{list of metadata that tell which
targets are up to date (from \code{\link[=drake_meta]{drake_meta()}}).}

\item{config}{internal configuration list}
}
\value{
a list containing the target's value and metadata
}
\description{
For internal use only.
the only reason this function is exported
is to allow \code{make(parallelism = "clustermq_staged")}.
}
\keyword{internal}
