% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{knitr_deps}
\alias{knitr_deps}
\title{Dependencies of a knitr report
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}}
\usage{
knitr_deps(target)
}
\arguments{
\item{target}{Encoded file path}
}
\value{
Data frame of dependencies
}
\description{
Deprecated on 2019-02-14
\code{knit("your_report.Rmd")} or
\code{knit("your_report.Rmd", quiet = TRUE)}.
}
\keyword{internal}
