% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dependencies.R
\name{knitr_input}
\alias{knitr_input}
\title{Declare the \code{knitr}/\code{rmarkdown} source files
of a workflow plan command.}
\usage{
knitr_input(...)
}
\arguments{
\item{...}{Symbols or character vectors naming source files
for \code{knitr}/\code{rmarkdown} dynamic reports.
knitr_input(my_report.Rmd)
knitr_input(report1.Rmd, 'report2.Rmd', "report3.Rmd")}
}
\value{
A character vector of declared files.
}
\description{
Use this function to help write the commands
in your workflow plan data frame.
}
\seealso{
file_input file_output
}
