% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run.R
\name{load_if_missing}
\alias{load_if_missing}
\title{Function \code{load_if_missing}}
\usage{
load_if_missing(packages)
}
\arguments{
\item{packages}{character vector of package names.}
}
\description{
loads and attaches packages 
if they are not already loaded.
}

