% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make.R
\name{make_impl}
\alias{make_impl}
\title{Internal function with a drake_config() argument}
\usage{
make_impl(config)
}
\arguments{
\item{config}{a \code{\link[=drake_config]{drake_config()}} object.}
}
\description{
Not a user-side function.
}
\keyword{internal}
