% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make.R
\name{make_session}
\alias{make_session}
\title{Internal function to be called by \code{\link[=make_with_config]{make_with_config()}}}
\usage{
make_session(config)
}
\arguments{
\item{config}{An input internal configuration list}
}
\description{
For internal use only. Not for the API.
}
\keyword{internal}
