% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{make_with_config}
\alias{make_with_config}
\title{Apply make() with a pre-computed config object
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}}
\usage{
make_with_config(config)
}
\arguments{
\item{config}{A configuration list returned by \code{\link[=drake_config]{drake_config()}}.}
}
\value{
nothing
}
\description{
Deprecated on 2019-01-04
}
\seealso{
\code{\link[=make]{make()}}, \code{\link[=drake_config]{drake_config()}}
}
\keyword{internal}
