% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{Makefile_recipe}
\alias{Makefile_recipe}
\title{Default Makefile recipe
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}}
\usage{
Makefile_recipe(
  recipe_command = drake::default_recipe_command(),
  target = "your_target",
  cache_path = NULL
)
}
\arguments{
\item{recipe_command}{Character scalar.}

\item{target}{Character scalar.}

\item{cache_path}{Character scalar.}
}
\value{
A character scalar
}
\description{
2019-01-03
}
\keyword{internal}
