% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manage_memory.R
\name{manage_memory}
\alias{manage_memory}
\title{Manage the in-memory dependencies of a target.
\ifelse{html}{\out{<a href='https://www.tidyverse.org/lifecycle/#stable'><img src='figures/lifecycle-stable.svg' alt='Stable lifecycle'></a>}}{\strong{Stable}}}
\usage{
manage_memory(target, config, downstream = NULL, jobs = 1)
}
\arguments{
\item{target}{Character, name of the target.}

\item{config}{\code{\link[=drake_config]{drake_config()}} list.}

\item{downstream}{Optional, character vector of any targets
assumed to be downstream.}

\item{jobs}{Number of jobs for local parallel computing}
}
\value{
Nothing.
}
\description{
Load/unload a target's dependencies.
Not a user-side function.
}
\keyword{internal}
