% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mclapply.R
\name{mc_master}
\alias{mc_master}
\title{Internal function to launch
a master process.}
\usage{
mc_master(config)
}
\arguments{
\item{config}{\code{drake_config()} list}
}
\value{
nothing important
}
\description{
For internal use only.
Exported only for the purpose of
using persistent workers in
\code{make(paralellism = "future_lapply", jobs = n)},
where \code{n > 1}.
}
\keyword{internal}
