% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mclapply.R
\name{mc_process}
\alias{mc_process}
\title{Internal function to launch
a master process or persistent worker.}
\usage{
mc_process(id, config)
}
\arguments{
\item{id}{character scalar with the job id}

\item{config}{\code{drake_config()} list}
}
\value{
nothing important
}
\description{
For internal use only.
Exported only for the purpose of
using persistent workers in
\code{make(paralellism = "parLapply", jobs = n)},
where \code{n > 1}.
}
\keyword{internal}
