% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/migrate.R
\name{migrate}
\alias{migrate}
\title{Function migrate}
\usage{
migrate(path)
}
\arguments{
\item{path}{Full path to the cache}
}
\description{
Migrate a project/cache from drake 4.4.0 or earlier
to be compatible with the version of drake on your system.
}
\details{
Versions after drake have a different internal structure for the cache.
This means projects built with drake 4.4.0 or before are not compatible
with projects built with a later version of drake. migrate() converts
an old cache to a format compatible with the version of drake
installed on your system.
}
\seealso{
\code{\link{rescue_cache}}, \code{\link{make}}
}
