% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outdated.R
\name{missed_impl}
\alias{missed_impl}
\title{Internal function with a drake_config() argument}
\usage{
missed_impl(config)
}
\arguments{
\item{config}{A \code{\link[=drake_config]{drake_config()}} object.}
}
\description{
Not a user-side function.
}
\keyword{internal}
