% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3_drake_deps_ht.R
\name{new_drake_deps_ht}
\alias{new_drake_deps_ht}
\title{\code{drake_deps_ht} constructor}
\usage{
new_drake_deps_ht(
  globals = ht_new(hash = TRUE),
  namespaced = ht_new(hash = FALSE),
  strings = ht_new(hash = FALSE),
  loadd = ht_new(hash = FALSE),
  readd = ht_new(hash = FALSE),
  file_in = ht_new(hash = FALSE),
  file_out = ht_new(hash = FALSE),
  knitr_in = ht_new(hash = FALSE)
)
}
\value{
A \code{drake_deps_ht} object.
}
\description{
List of class \code{drake_deps_ht}.
}
\examples{
if (FALSE) { # stronger than roxygen dontrun
new_drake_deps_ht()
}
}
\keyword{internal}
