% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3_drake_settings.R
\name{new_drake_settings}
\alias{new_drake_settings}
\title{\code{drake_settings} constructor}
\usage{
new_drake_settings(
  curl_handles = NULL,
  garbage_collection = NULL,
  jobs = NULL,
  jobs_preprocess = NULL,
  keep_going = NULL,
  lazy_load = NULL,
  lib_loc = NULL,
  lock_envir = NULL,
  lock_cache = NULL,
  memory_strategy = NULL,
  parallelism = NULL,
  recover = NULL,
  recoverable = NULL,
  session_info = NULL,
  skip_imports = NULL,
  skip_safety_checks = NULL,
  skip_targets = NULL,
  sleep = NULL,
  template = NULL
)
}
\value{
A \code{drake_settings} object.
}
\description{
List of class \code{drake_settings}.
}
\examples{
if (FALSE) { # stronger than roxygen dontrun
new_drake_settings()
}
}
\keyword{internal}
