% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3_drake_triggers.R
\name{new_drake_triggers}
\alias{new_drake_triggers}
\title{\code{drake_triggers} constructor}
\usage{
new_drake_triggers(
  command = TRUE,
  depend = TRUE,
  file = TRUE,
  seed = TRUE,
  format = TRUE,
  condition = FALSE,
  change = NULL,
  mode = "whitelist"
)
}
\arguments{
\item{command}{Logical, command trigger.}

\item{depend}{Logical, depend trigger.}

\item{file}{Logical, file trigger.}

\item{seed}{Logical, seed trigger.}

\item{format}{Logical, format trigger.}

\item{condition}{Language object or object coercible to logical,
condition trigger.}

\item{change}{Language object or literal value, change trigger.}

\item{mode}{Character, mode of condition trigger.}
}
\value{
A \code{drake_triggers} object.
}
\description{
List of class \code{drake_triggers}.
}
\examples{
if (FALSE) { # stronger than roxygen dontrun
new_drake_triggers()
}
}
\keyword{internal}
