% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package_dependencies.R
\name{pacakge_dependency_version}
\alias{pacakge_dependency_version}
\title{Function pacakge_dependency_version}
\usage{
pacakge_dependency_version(x)
}
\arguments{
\item{x}{name of the package. Could have \code{"package:"} in front
of it or not.}
}
\description{
Packages are formally imported into drake workflows,
and they serve as dependencies of package functions.
Specifically, the version of each package is tracked.
For base packages like \code{base} and \code{stats},
the version of R itself is tracked. Here, the version string
is used, so different snapshots of R-devel have different versions.
}
\examples{
pacakge_dependency_version("base")
pacakge_dependency_version("stats")
pacakge_dependency_version("package:stats")
\dontrun{
load_basic_example()
plot_graph(my_plan) # Hover over the package nodes on the left.
}
}
