% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package_dependencies.R
\name{package_version}
\alias{package_version}
\title{Function package_version}
\usage{
package_version(x)
}
\description{
Packages are formally imported into drake workflows,
and they serve as dependencies of package functions.
Specifically, the version of each package is tracked.
For base packages like \code{base} and \code{stats},
the version of R itself is tracked. Here, the version string
is used, so different snapshots of R-devel have different versions.
}
\examples{
packageVersion("base")
packageVersion("drake")
packageVersion("package:drake")
\dontrun{
load_basic_example()
plot_graph(my_plan) # Hover over the package nodes on the left.
}
}
