% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/defunct.R
\name{parallel_stages}
\alias{parallel_stages}
\title{parallel_stages \ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#defunct}{\figure{lifecycle-defunct.svg}{options: alt='[Defunct]'}}}{\strong{[Defunct]}}}
\usage{
parallel_stages(...)
}
\arguments{
\item{...}{Arguments}
}
\description{
2019-02-15
}
\keyword{internal}
