% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/defunct.R
\name{plan_drake}
\alias{plan_drake}
\title{plan_drake \ifelse{html}{\figure{lifecycle-defunct.svg}{options: alt='Defunct lifecycle'}}{\strong{Defunct}}}
\usage{
plan_drake(...)
}
\arguments{
\item{...}{Arguments}
}
\description{
2019-02-15
}
\keyword{internal}
