% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{plan_summaries}
\alias{plan_summaries}
\title{Specialized wildcard for summaries
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}}
\usage{
plan_summaries(
  plan,
  analyses,
  datasets,
  gather = rep("list", nrow(plan)),
  sep = "_"
)
}
\arguments{
\item{plan}{Workflow plan data frame with commands for the summaries.
Use the \code{analysis__} and \code{dataset__} wildcards
just like the \code{dataset__} wildcard in \code{\link[=plan_analyses]{plan_analyses()}}.}

\item{analyses}{Workflow plan data frame of analysis instructions.}

\item{datasets}{Workflow plan data frame with instructions to make
or import the datasets.}

\item{gather}{Character vector, names of functions to gather the
summaries. If not \code{NULL}, the length must be the number of
rows in the \code{plan}. See the \code{\link[=gather_plan]{gather_plan()}} function
for more.}

\item{sep}{Character scalar, delimiter for creating the
new target names.}
}
\value{
An evaluated workflow plan data frame of instructions
for computing summaries of analyses and datasets.
analyses of multiple datasets in multiple ways.
}
\description{
Use \code{\link[=drake_plan]{drake_plan()}} with transformations instead. See
\verb{https://books.ropensci.org/drake/plans.html#large-plans}
for details.
}
\details{
2019-01-13
}
\keyword{internal}
