% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api-plan.R
\name{print.drake_plan}
\alias{print.drake_plan}
\title{Internal formatting function.}
\usage{
\method{print}{drake_plan}(x, ...)
}
\arguments{
\item{x}{A \code{drake} plan.}

\item{...}{Other args.}
}
\description{
Internal.
}
\keyword{internal}
