% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_imports.R
\name{process_import}
\alias{process_import}
\title{Process an imported data object
\ifelse{html}{\out{<a href='https://www.tidyverse.org/lifecycle/#stable'><img src='figures/lifecycle-stable.svg' alt='Stable lifecycle'></a>}}{\strong{Stable}}}
\usage{
process_import(import, config)
}
\arguments{
\item{import}{Character, name of an import to process}

\item{config}{\code{\link[=drake_config]{drake_config()}} object}
}
\description{
For internal use only. Not a user-side function.
}
\keyword{internal}
