% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean.R
\name{prune}
\alias{prune}
\title{Function \code{prune}}
\usage{
prune(plan)
}
\arguments{
\item{plan}{data frame, drake workflow like the one generated
by \code{\link{example_plan}}.}
}
\description{
Removes any cached output objects and drake-generated 
files not listed in \code{plan$output$}. 
Your working directory (\code{\link{getwd}()}) must be the
root directory of your drake project.
WARNING: this removes files.
Only do this if you're sure you won't lose any important work.
}
\seealso{
\code{\link{clean}}, \code{\link{make}},
\code{\link{help_drake}}
}

