% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-igraph.R
\name{prune_drake_graph}
\alias{prune_drake_graph}
\title{Prune an igraph. Not a user-side function.}
\usage{
prune_drake_graph(graph, to = igraph::V(graph)$name, jobs = 1)
}
\arguments{
\item{graph}{An igraph object to be pruned.}

\item{to}{Character vector, of destination vertices.}

\item{jobs}{deprecated}
}
\value{
An \code{igraph} object
}
\description{
Similar to \code{igraph::subcomponent(mode = "in")}
but able to support multiple destination vertices.
For the internals of \code{drake} only.
Not a user-side function.
}
\details{
Removes igraph attributes, so be careful.
}
\examples{
# Not a user-side function.
}
\keyword{internal}
