% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{read_drake_graph}
\alias{read_drake_graph}
\title{Read a workflow graph from the cache
\lifecycle{deprecated}}
\usage{
read_drake_graph(path = getwd(), search = TRUE, cache = NULL,
  verbose = 1L)
}
\arguments{
\item{path}{Path to a \code{drake} cache
(usually a hidden \code{.drake/} folder) or \code{NULL}.}

\item{search}{Deprecated.}

\item{cache}{drake cache. See \code{\link[=new_cache]{new_cache()}}.
If supplied, \code{path} is ignored.}

\item{verbose}{Integer, control printing to the console/terminal.
\itemize{
\item \code{0}: print nothing.
\item \code{1}: print targets, retries, and failures.
\item \code{2}: also show a spinner when preprocessing tasks are underway.
}}
}
\description{
drake no longer stores the config object,
the plan, etc. in the cache during \code{make()}. This change
improves speed.
}
\details{
2019-01-06
}
\keyword{internal}
