% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{read_drake_plan}
\alias{read_drake_plan}
\title{Read the plan from the cache
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}}
\usage{
read_drake_plan(path = getwd(), search = TRUE, cache = NULL, verbose = 1L)
}
\arguments{
\item{path}{Path to a \code{drake} cache
(usually a hidden \verb{.drake/} folder) or \code{NULL}.}

\item{search}{Deprecated.}

\item{cache}{drake cache. See \code{\link[=new_cache]{new_cache()}}.
If supplied, \code{path} is ignored.}

\item{verbose}{Deprecated on 2019-09-11.}
}
\description{
drake no longer stores the config object,
the plan, etc. in the cache during \code{make()}. This change
improves speed.
}
\details{
2019-01-06
}
\keyword{internal}
