% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/defunct.R
\name{read_graph}
\alias{read_graph}
\title{read_graph \ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#defunct}{\figure{lifecycle-defunct.svg}{options: alt='[Defunct]'}}}{\strong{[Defunct]}}}
\usage{
read_graph(...)
}
\arguments{
\item{...}{Arguments}
}
\description{
2019-02-15
}
\keyword{internal}
