% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api-cache.R
\name{recover_cache_}
\alias{recover_cache_}
\title{or create a new one otherwise.}
\usage{
recover_cache_(path = NULL, hash_algorithm = NULL,
  short_hash_algo = NULL, long_hash_algo = NULL, force = FALSE,
  verbose = 1L, fetch_cache = NULL, console_log_file = NULL)
}
\description{
or create a new one otherwise.
}
