% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api-outdated.R
\name{recoverable}
\alias{recoverable}
\title{List the \emph{recoverable} targets that are out of date.}
\usage{
recoverable(config, make_imports = TRUE, do_prework = TRUE)
}
\arguments{
\item{config}{Optional internal runtime parameter list
produced with \code{\link[=drake_config]{drake_config()}}.
You must use a fresh \code{config} argument with an up-to-date
dependency graph that was never modified by hand.
If needed, rerun \code{\link[=drake_config]{drake_config()}} early and often.
See the details in the help file for \code{\link[=drake_config]{drake_config()}}.}

\item{make_imports}{Logical, whether to make the imports first.
Set to \code{FALSE} to save some time and risk obsolete output.}

\item{do_prework}{Whether to do the \code{prework}
normally supplied to \code{\link[=make]{make()}}.}
}
\value{
Character vector of the names of recoverable targets.
}
\description{
Some outdated targets can be recovered using old data.
For details, see the \code{recover} argument in the help file of \code{\link[=make]{make()}}.
The \code{\link[=recoverable]{recoverable()}} function lists the outdated targets that
can be recovered with \code{make(recover = TRUE)}. Instead of rerunning
the commands, \code{\link[=make]{make()}} will simply assign the old data to the new
targets as appropriate.
}
\examples{
\dontrun{
isolate_example("Quarantine side effects.", {
if (suppressWarnings(require("knitr"))) {
load_mtcars_example() # Get the code with drake_example("mtcars").
# Recopute the config list early and often to have the
# most current information. Do not modify the config list by hand.
config <- drake_config(my_plan)
outdated(config = config) # Which targets are out of date?
recoverable(config = config) # Which of these are recoverable?
}
})
}
}
\seealso{
\code{\link[=r_recoverable]{r_recoverable()}}, \code{\link[=r_outdated]{r_outdated()}}, \code{\link[=drake_config]{drake_config()}}, \code{\link[=missed]{missed()}},
\code{\link[=drake_plan]{drake_plan()}}, \code{\link[=make]{make()}}
}
