% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-deprecate.R
\name{reduce_plan}
\alias{reduce_plan}
\title{Deprecated:
write commands to reduce several targets down to one.}
\usage{
reduce_plan(plan = NULL, target = "target", begin = "", op = " + ",
  end = "", pairwise = TRUE, append = FALSE, sep = "_")
}
\arguments{
\item{plan}{Workflow plan data frame of prespecified targets.}

\item{target}{Name of the new reduced target.}

\item{begin}{Character, code to place at the beginning
of each step in the reduction.}

\item{op}{Binary operator to apply in the reduction}

\item{end}{Character, code to place at the end
of each step in the reduction.}

\item{pairwise}{Logical, whether to create multiple
new targets, one for each pair/step in the reduction (\code{TRUE}),
or to do the reduction all in one command.}

\item{append}{Logical. If \code{TRUE}, the output will include the
original rows in the \code{plan} argument.
If \code{FALSE}, the output will only include the new
targets and commands.}

\item{sep}{Character scalar, delimiter for creating new target names.}
}
\value{
A workflow plan data frame that aggregates multiple
prespecified targets into one additional target downstream.
}
\description{
Deprecated on 2019-05-16. Use \code{\link[=drake_plan]{drake_plan()}}
transformations instead. See
\url{https://ropenscilabs.github.io/drake-manual/plans.html#large-plans}
for the details.
}
\details{
Creates a new workflow plan data frame with the
commands to do a reduction (i.e. to repeatedly apply a binary
operator to pairs of targets to produce one target).
}
\seealso{
\code{\link[=drake_plan]{drake_plan()}}
}
\keyword{internal}
