% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hasty.R
\name{remote_hasty_build}
\alias{remote_hasty_build}
\title{Build a target using "hasty" parallelism}
\usage{
remote_hasty_build(target, deps = NULL, config)
}
\arguments{
\item{target}{name of the target}

\item{deps}{named list of dependencies}

\item{config}{internal configuration list}
}
\description{
For internal use only
}
\keyword{internal}
