% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/future_lapply.R
\name{remote_worker}
\alias{remote_worker}
\title{Run a remote persistent worker}
\usage{
remote_worker(worker, cache_path)
}
\arguments{
\item{worker}{the worker ID}

\item{cache_path}{path to the drake cache}
}
\value{
\code{NULL}
}
\description{
Users should not need to call this function directly.
}
\examples{
# No examples here. This function is not for end users.
}
\keyword{internal}
