% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api-rename.R
\name{rename_targets}
\alias{rename_targets}
\title{Rename targets without having to rebuild them}
\usage{
rename_targets(from, to, cache = drake::drake_cache(path = path, verbose
  = verbose), path = NULL, verbose = 1L)
}
\arguments{
\item{from}{Character vector of original target names.}

\item{to}{Character vector of new target names. Must be the same length
as \code{from}.}

\item{cache}{drake cache. See \code{\link[=new_cache]{new_cache()}}.
If supplied, \code{path} is ignored.}

\item{path}{Path to a \code{drake} cache
(usually a hidden \code{.drake/} folder) or \code{NULL}.}

\item{verbose}{Integer, control printing to the console/terminal.
\itemize{
\item \code{0}: print nothing.
\item \code{1}: print targets, retries, and failures.
\item \code{2}: also show a spinner when preprocessing tasks are underway.
}}
}
\description{
Changes the names of already-built targets
in the cache.
}
\details{
Behind the scenes, \code{\link[=rename_targets]{rename_targets()}}
registers new names for targets in the cache. Remarks:
\itemize{
\item You also need to change the target names
in your \code{\link[=drake_plan]{drake_plan()}} as appropriate.
\item Any name change will invalidate downstream targets that depend
on the original names.
\item The old names are still valid until you change them in the plan
and call \code{\link[=make]{make()}}.
\item The underlying data is not duplicated.
This is because \code{drake}'s cache is powered by
\href{https://github.com/richfitz/storr}{storr}, which allows
multiple keys for each data value.
}
}
\examples{
\dontrun{
isolate_example("Let's rename some targets.", {
# Let's start with this plan.
plan <- drake_plan(
  temp = target(
    matrix(runif(10 ^ 3) + offset, ncol = 5),
    transform = map(offset = c(2, 5))
  )
)
make(plan)

# Later on, we add two more targets.
plan <- drake_plan(
  temp = target(
    matrix(runif(10 ^ 3) + offset, ncol = ncol),
    transform = cross(
      ncol = c(5, 2),
      offset = c(2, 5)
    )
  )
)

# But then our original targets have different names,
# and they are no longer up to date.
config <- drake_config(plan)
outdated(config)

# Let's rename those old targets.
rename_targets(
  from = c("temp_2", "temp_5"),
  to = c("temp_5_2", "temp_5_5")
)

# Now, those old targets are up to date and make() skips them.
outdated(config)
make(plan)
})
}
}
