% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/defunct.R
\name{render_graph}
\alias{render_graph}
\title{render_graph \ifelse{html}{\figure{lifecycle-defunct.svg}{options: alt='Defunct lifecycle'}}{\strong{Defunct}}}
\usage{
render_graph(...)
}
\arguments{
\item{...}{Arguments}
}
\description{
2019-02-15
}
\keyword{internal}
