% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-deprecate.R
\name{render_static_drake_graph}
\alias{render_static_drake_graph}
\title{Deprecated: render a \code{ggraph}/\code{ggplot2} representation
of your drake project.}
\usage{
render_static_drake_graph(graph_info, main = graph_info$default_title)
}
\arguments{
\item{graph_info}{List of data frames generated by
\code{\link[=drake_graph_info]{drake_graph_info()}}.
There should be 3 data frames: \code{nodes}, \code{edges},
and \code{legend_nodes}.}

\item{main}{Character string, title of the graph.}
}
\value{
A \code{ggplot2} object, which you can modify with more layers,
show with \code{plot()}, or save as a file with \code{ggsave()}.
}
\description{
Use \code{\link[=render_drake_ggraph]{render_drake_ggraph()}} instead.
}
\details{
Deprecated on 2018-25-07.
}
\examples{
# See render_drake_ggraph()
}
\keyword{internal}
