% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rstudio.R
\name{rs_addin_r_outdated}
\alias{rs_addin_r_outdated}
\title{RStudio addin for r_outdated()
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}}
\usage{
rs_addin_r_outdated(r_args = list(), .print = TRUE)
}
\arguments{
\item{r_args}{List of arguments to \code{r_fn}, not including \code{func} or \code{args}.
Example:
\code{r_make(r_fn = callr::r_bg, r_args = list(stdout = "stdout.log"))}.}

\item{.print}{Logical, whether to \code{print()} the result
to the console. Required for the addin.}
}
\value{
A character vector of outdated targets.
}
\description{
Call \code{\link[=r_outdated]{r_outdated()}} in an RStudio addin.
}
\keyword{internal}
