% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rstudio.R
\name{rs_addin_readd}
\alias{rs_addin_readd}
\title{Readd target at cursor into global environment
\lifecycle{stable}}
\usage{
rs_addin_readd(context = NULL)
}
\arguments{
\item{context}{an RStudio document context.
Read from the active document if not supplied.
This is used for testing purposes.}
}
\value{
Nothing.
}
\description{
This function provides an RStudio addin that will
read the target at the
current cursor location from the cache.
}
\details{
If you are using a non-standard \code{drake} cache,
you must supply it to the \code{"rstudio_drake_cache"} global option,
e.g. \code{options(rstudio_drake_cache = storr::storr_rds("my_cache"))}.
}
\keyword{internal}
