% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run.R
\name{run}
\alias{run}
\title{Function \command{run}}
\usage{
run(plan, targets = plan$target, envir = parent.frame(),
  parallelism = c("mclapply", "Makefile"), jobs = 1)
}
\arguments{
\item{plan}{workflow plan data frame}

\item{targets}{targets to bulid}

\item{envir}{environment to use}

\item{parallelism}{character, type of parallelism to use}

\item{jobs}{number of parallel processes or jobs}
}
\description{
Run your project.
}

