% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/defunct.R
\name{session}
\alias{session}
\title{session \ifelse{html}{\figure{lifecycle-defunct.svg}{options: alt='Defunct lifecycle'}}{\strong{Defunct}}}
\usage{
session(...)
}
\arguments{
\item{...}{Arguments}
}
\description{
2019-02-15
}
\keyword{internal}
