% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-deprecate.R
\name{short_hash}
\alias{short_hash}
\title{Deprecated. \code{drake} now only uses one hash algorithm per cache.}
\usage{
short_hash(cache = drake::get_cache(verbose = verbose), verbose = 1L)
}
\arguments{
\item{cache}{drake cache. See \code{\link[=new_cache]{new_cache()}}.
If supplied, \code{path} and \code{search} are ignored.}

\item{verbose}{Integer, control printing to the console/terminal.
\itemize{
\item \code{0}: print nothing.
\item \code{1}: print targets, retries, and failures.
\item \code{2}: also show a spinner when preprocessing tasks are underway.
}}
}
\value{
A character vector naming a hash algorithm.
}
\description{
Deprecated on 2018-12-12.
}
\examples{
# deprecated
}
\keyword{internal}
