% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache.R
\name{status}
\alias{status}
\title{Function \code{status}}
\usage{
status(path = getwd(), search = FALSE)
}
\arguments{
\item{path}{Root directory of the drake project,
or if \code{search} is \code{TRUE}, either the
project root or a subdirectory of the project.}

\item{search}{If \code{TRUE}, search parent directories
to find the nearest drake cache. Otherwise, look in the
current working directory only.}
}
\value{
data frame containing the build status
of the last session
}
\description{
Get the build status of the last call to 
\code{\link{run}()}: which objects were skipped, built,
and imported.
}
\seealso{
\code{\link{session}},
\code{\link{built}}, \code{\link{imported}},
\code{\link{readd}}, 
\code{\link{plan}}, \code{\link{run}}
}

