% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api-plan.R
\name{[.drake_lang}
\alias{[.drake_lang}
\title{Internal formatting function.}
\usage{
\method{[}{drake_lang}(x, i)
}
\arguments{
\item{x}{S3 \code{drake_lang} character vector.}

\item{i}{Numeric index.}
}
\description{
Internal.
}
\keyword{internal}
