% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api-plan.R
\name{[.langs}
\alias{[.langs}
\title{Internal formatting function.}
\usage{
\method{[}{langs}(x, i)
}
\arguments{
\item{x}{S3 \code{langs} character vector.}

\item{i}{Numeric index.}
}
\description{
Internal.
}
\keyword{internal}
