% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test.R
\name{test_with_dir}
\alias{test_with_dir}
\title{Internal function \code{test_with_dir}}
\usage{
test_with_dir(desc, ...)
}
\arguments{
\item{desc}{character, description of the test}

\item{...}{code to test}
}
\value{
nothing
}
\description{
Run a test in a way that quarantines
the side effects from your workspace and file system.
}
\examples{
\dontrun{
test_with_dir(
  "Write a file to a temporary folder",
  writeLines("hello", "world.txt")
)
file.exists("world.txt") # FALSE
}
}
\keyword{internal}
