% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text_drake_graph.R
\name{text_drake_graph_impl}
\alias{text_drake_graph_impl}
\title{Internal function with a drake_config() argument}
\usage{
text_drake_graph_impl(
  config,
  from = NULL,
  mode = c("out", "in", "all"),
  order = NULL,
  subset = NULL,
  targets_only = FALSE,
  make_imports = TRUE,
  from_scratch = FALSE,
  group = NULL,
  clusters = NULL,
  show_output_files = TRUE,
  nchar = 1L,
  print = TRUE
)
}
\arguments{
\item{config}{A \code{\link[=drake_config]{drake_config()}} object.}

\item{make_imports}{Logical, whether to make the imports first.
Set to \code{FALSE} to save some time and risk obsolete output.}
}
\description{
Not a user-side function.
}
\keyword{internal}
