% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-deprecate.R
\name{this_cache}
\alias{this_cache}
\title{Get the cache at the exact file path specified.}
\usage{
this_cache(path = NULL, force = FALSE, verbose = 1L,
  fetch_cache = NULL, console_log_file = NULL)
}
\arguments{
\item{path}{File path of the cache.}

\item{force}{Deprecated.}

\item{verbose}{Integer, control printing to the console/terminal.
\itemize{
\item \code{0}: print nothing.
\item \code{1}: print targets, retries, and failures.
\item \code{2}: also show a spinner when preprocessing tasks are underway.
}}

\item{fetch_cache}{Deprecated.}

\item{console_log_file}{Optional character scalar of a file name or
connection object (such as \code{stdout()}) to dump maximally verbose
log information for \code{\link[=make]{make()}}. Independent of the \code{verbose} argument.}
}
\value{
A drake/storr cache at the specified path, if it exists.
}
\description{
This function does not apply to
in-memory caches such as \code{storr_environment()}.
}
\keyword{internal}
