% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecate.R
\name{time_stamps}
\alias{time_stamps}
\title{Deprecated}
\usage{
time_stamps(config)
}
\arguments{
\item{config[drake_config()]}{list}
}
\value{
nothing
}
\description{
2019-01-03
}
\examples{
# deprecated
}
\keyword{internal}
