% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deps.R
\name{tracked}
\alias{tracked}
\title{List the targets and imports that are reproducibly tracked.
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}}
\usage{
tracked(config)
}
\arguments{
\item{config}{An output list from \code{\link[=drake_config]{drake_config()}}.}
}
\value{
A character vector with the names of reproducibly-tracked targets.
}
\description{
List all the spec
in your project's dependency network.
}
\examples{
\dontrun{
isolate_example("Quarantine side effects.", {
if (suppressWarnings(require("knitr"))) {
load_mtcars_example() # Load the canonical example for drake.
# List all the targets/imports that are reproducibly tracked.
config <- drake_config(my_plan)
tracked(config)
}
})
}
}
