% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/triggers.R
\name{trigger}
\alias{trigger}
\title{Set the trigger of a target.}
\usage{
trigger(condition = FALSE, command = TRUE, depends = TRUE, file = TRUE,
  change = NULL)
}
\arguments{
\item{condition}{R code (expression or language object)
that returns a logical. The target will rebuild
if the code evaluates to \code{TRUE}.}

\item{command}{logical, whether to rebuild the target if the
\code{\link[=drake_plan]{drake_plan()}} command changes.}

\item{depends}{logical, whether to rebuild if a
non-file dependency changes.}

\item{file}{logical, whether to rebuild the target
if a \code{\link[=file_in]{file_in()}}/\code{\link[=file_out]{file_out()}}/\code{\link[=knitr_in]{knitr_in()}} file changes.}

\item{change}{R code (expression or language object)
that returns any value. The target will rebuild
if that value is different from last time
or not already cached.}
}
\value{
a list of trigger specification details that
\code{drake} processes internally when it comes time to decide
whether to build the target.
}
\description{
Use this function inside a target's command
in your \code{\link[=drake_plan]{drake_plan()}}. The target will rebuild if and only if:
\itemize{
\item Any of \code{command}, \code{depends}, or \code{file} is \code{TRUE}, or
\item \code{condition} evaluates to \code{TRUE}, or
\item \code{change} evaluates to a value different from last time.
}
}
\seealso{
\code{\link[=drake_plan]{drake_plan()}}, \code{\link[=make]{make()}}
}
