% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache_types.R
\name{type_of_cache}
\alias{type_of_cache}
\title{experimental function type_of_cache}
\usage{
type_of_cache(path)
}
\arguments{
\item{path}{path to the cache}
}
\description{
Try to get the type of a drake
file system cache. Only works on known caches
with known file systems.
}
\details{
Experimental function for a possible
new feature in the future.
It will come in handy if/when multiple cache types
are supported.
}
\examples{
\dontrun{
load_basic_example()
make(my_plan)
type_of_cache(".drake")
}
}
