% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/walk_code.R
\name{walk_code}
\alias{walk_code}
\title{Static code analysis}
\usage{
walk_code(expr, results, locals, restrict)
}
\arguments{
\item{expr}{A function or expression.}

\item{results}{A \code{drake_deps} object.}

\item{locals}{An environment, a hash table of local variables.}

\item{restrict}{An environment,
a hash table for whitelisting global symbols.}
}
\description{
Static code analysis.
}
\keyword{internal}
