% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workflow.R
\name{workflow}
\alias{workflow}
\title{Function \code{workflow}}
\usage{
workflow(...)
}
\arguments{
\item{...}{workflow named with their respective targets}
}
\value{
data frame of remake targets and commands
}
\description{
Turns a named collection of \code{remake} commands into 
a data frame of \code{remake} targets and \code{remake} workflow.
\code{\link{workflow_string}} and \code{\link{workflow_batch}} are similar.
}
\seealso{
\code{\link{workflow_string}}, \code{\link{workflow_batch}},
}

