% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workflow.R
\name{workflow_batch}
\alias{workflow_batch}
\title{Function \code{workflow_batch}}
\usage{
workflow_batch(x = NULL)
}
\arguments{
\item{x}{named character vector with targets as names and workflow as elements}
}
\value{
data frame of remake targets and workflow
}
\description{
Similar to \code{\link{workflow}} except that commands are
given as a named character vector with targets as names and commands as elements.
\code{\link{workflow_string}} is another alternative.
}
\seealso{
\code{\link{workflow}}, \code{\link{workflow_string}},
}

