% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workflow.R
\name{workflow_string}
\alias{workflow_string}
\title{Function \code{workflow_string}}
\usage{
workflow_string(...)
}
\arguments{
\item{...}{workflow named with their respective targets}
}
\value{
data frame of skipthat targets and commands
}
\description{
Similar to \code{\link{workflow}} except that commands are
parsed as strings rather than symbols. \code{\link{workflow_batch}} is
another alternative.
}
\seealso{
\code{\link{workflow}}, \code{\link{workflow_batch}},
}

