% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate.R
\name{analyses}
\alias{analyses}
\title{Function \code{analyses}}
\usage{
analyses(commands, datasets)
}
\arguments{
\item{commands}{output of the \code{\link{workflow}} function.}

\item{datasets}{Data frame of commands to generate datasets}
}
\value{
Preprocessed data frame of \code{skipthat} analysis commands
}
\description{
Produces a data frame of \code{skipthat} analysis commands.
}

