% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate.R
\name{analyses}
\alias{analyses}
\title{Function \code{analyses}}
\usage{
analyses(plan, datasets)
}
\arguments{
\item{plan}{data frame with unevaluated wildcards}

\item{datasets}{dataframe with the drake plan for datasets}
}
\value{
data frame of instructions for analyses (wildcards unevaluated)
}
\description{
Generate a portion of a \code{drake} plan for
analyzing multiple datasets with multiple methods.
}
\seealso{
\code{\link{summaries}},
 \code{\link{run}}, \code{\link{plan}}
}

