% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecate.R
\name{analyses}
\alias{analyses}
\title{Function \code{analyses}}
\usage{
analyses(plan, datasets)
}
\arguments{
\item{plan}{Same as for \code{\link{plan_analyses}()}.}

\item{datasets}{Same as for \code{\link{plan_analyses}()}.}
}
\value{
The same return value as \code{\link{plan_analyses}()}.
}
\description{
Use \code{\link{plan_analyses}()} instead.
}
\details{
Deprecated on 2017-11-12.
}
\examples{
# See ?future_backend for examples.
}
\seealso{
\code{\link{future_backend}}
}
