% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{analysis_wildcard}
\alias{analysis_wildcard}
\title{Show the analysis wildcard
used in \code{\link[=plan_summaries]{plan_summaries()}}.
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}}
\usage{
analysis_wildcard()
}
\value{
The analysis wildcard used in \code{\link[=plan_summaries]{plan_summaries()}}.
}
\description{
Deprecated on 2019-01-12.
}
\details{
Used to generate workflow plan data frames.
}
\seealso{
\code{\link[=plan_summaries]{plan_summaries()}}
}
\keyword{internal}
