% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate.R
\name{analysis_wildcard}
\alias{analysis_wildcard}
\title{Function analysis_wildcard}
\usage{
analysis_wildcard()
}
\value{
The analysis wildcard used in \code{\link{plan_summaries}()}.
}
\description{
Show the analysis wildcard
used in \code{\link{plan_summaies}()}.
}
\examples{
# See ?plan_analyses for examples
}
\seealso{
\code{\link{plan_summaries}()}
}
