% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strings.R
\name{as_drake_filename}
\alias{as_drake_filename}
\title{Function \code{as_drake_filename}}
\usage{
as_drake_filename(x)
}
\arguments{
\item{x}{character string to be turned into a filename
understandable by drake (i.e., a string with literal
single quotes on both ends).}
}
\value{
A single-quoted character string: i.e., a filename
understandable by drake.
}
\description{
Converts an ordinary character string
into a filename understandable by drake. In other words,
\code{as_drake_filename(x)} just wraps single quotes around \code{x}.
}
\examples{
# Wraps the string in single quotes.
as_drake_filename("my_file.rds") # "'my_file.rds'"
}
