% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecate.R
\name{as_file}
\alias{as_file}
\title{Deprecated function \code{as_file}}
\usage{
as_file(x)
}
\arguments{
\item{x}{Same as for \code{\link[=file_store]{file_store()}}.}
}
\value{
The same return value as \code{\link[=file_store]{file_store()}}.
}
\description{
Use \code{\link[=file_store]{file_store()}} instead.
}
\details{
Deprecated on 2017-11-12.
}
\examples{
# See ?file_store for examples.
}
\seealso{
\code{\link[=file_store]{file_store()}}
}
\keyword{internal}
