% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecate.R
\name{as_file}
\alias{as_file}
\title{Deprecated function \code{as_file}}
\usage{
as_file(x)
}
\arguments{
\item{x}{Same as for \code{\link[=as_drake_filename]{as_drake_filename()}}.}
}
\value{
The same return value as \code{\link[=as_drake_filename]{as_drake_filename()}}.
}
\description{
Use \code{\link[=as_drake_filename]{as_drake_filename()}} instead.
}
\details{
Deprecated on 2017-11-12.
}
\examples{
# See ?as_drake_filename for examples.
}
\seealso{
\code{\link[=as_drake_filename]{as_drake_filename()}}
}
\keyword{internal}
