% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plan.R
\name{as_file}
\alias{as_file}
\title{Function \code{as_file}}
\usage{
as_file(x)
}
\arguments{
\item{x}{character string to be turned into a filename
understandable by drake (i.e., a string with literal
single quotes on both ends).}
}
\value{
a single-quoted character string: i.e., a filename
understandable by drake.
}
\description{
Converts an ordinary character string
into a filename understandable by drake. In other words,
\code{as_file(x)} just wraps single quotes around \code{x}
}

