% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hash.R
\name{available_hash_algos}
\alias{available_hash_algos}
\title{List the available hash algorithms for drake caches.}
\usage{
available_hash_algos()
}
\value{
A character vector of names of available hash algorithms.
}
\description{
See the advanced storage tutorial
at \url{https://github.com/ropensci/drake/blob/master/vignettes/storage.Rmd}
for details.
}
\examples{
available_hash_algos()
}
