% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hash.R
\name{available_hash_algos}
\alias{available_hash_algos}
\title{Function \code{available_hash_algos}}
\usage{
available_hash_algos()
}
\value{
A character vector of names of available hash algorithms.
}
\description{
List the available hash algorithms.
}
\examples{
available_hash_algos()
}
