% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecate.R
\name{available_hash_algos}
\alias{available_hash_algos}
\title{Deprecated. List the available hash algorithms for drake caches.}
\usage{
available_hash_algos()
}
\value{
A character vector of names of available hash algorithms.
}
\description{
Deprecated on 2018-12-12.
}
\examples{
# deprecated
}
\keyword{internal}
