% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hash_algorithms.R
\name{available_hash_algos}
\alias{available_hash_algos}
\title{Function available_hash_algos}
\usage{
available_hash_algos()
}
\description{
List the available hash algorithms.
}
\examples{
available_hash_algos()
}
