% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecate.R
\name{backend}
\alias{backend}
\title{Function \code{future_backend}}
\usage{
backend(...)
}
\arguments{
\item{...}{Arguments to \code{future::plan()}.}
}
\value{
The same return value as \code{future::plan()}.
}
\description{
Use \code{\link{future_backend}()} instead.
}
\details{
Deprecated on 2017-11-12.
}
\examples{
# See ?future_backend for examples.
}
\seealso{
\code{\link{future_backend}}
}
