% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build.R
\name{build}
\alias{build}
\title{Internal function build}
\usage{
build(target, meta_list, config)
}
\arguments{
\item{target}{name of the target}

\item{meta_list}{list of metadata that tell which
targets are up to date}

\item{config}{internal configuration list}
}
\description{
Function to build a target.
For internal use only.
the only reason this function is exported
is to set up PSOCK clusters efficiently.
}
