% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph.R
\name{build_drake_graph}
\alias{build_drake_graph}
\title{Create the \code{igraph} dependency network of your project.}
\usage{
build_drake_graph(plan = read_drake_plan(),
  targets = drake::possible_targets(plan), envir = parent.frame(),
  verbose = drake::default_verbose(), jobs = 1, sanitize_plan = TRUE)
}
\arguments{
\item{plan}{workflow plan data frame, same as for function
\code{\link[=make]{make()}}.}

\item{targets}{names of targets to build, same as for function
\code{\link[=make]{make()}}.}

\item{envir}{environment to import from, same as for function
\code{\link[=make]{make()}}.}

\item{verbose}{logical or numeric, control printing to the console.
Set the defauld value with \code{pkgconfig}:
for example, \code{pkgconfig::set_config("drake::verbose" = 2)}.
\describe{
\item{0 or \code{FALSE}:}{print nothing.}
\item{1 or \code{TRUE}:}{print only targets to build.}
\item{2:}{in addition, print checks and cache info.}
\item{3:}{in addition, print any potentially missing items.}
\item{4:}{in addition, print imports. Full verbosity.}
}}

\item{jobs}{number of jobs to accelerate the construction
of the dependency graph. A light \code{mclapply()}-based
parallelism is used if your operating system is not Windows.}

\item{sanitize_plan}{logical, whether to sanitize the workflow plan first.}
}
\value{
An igraph object representing
the workflow plan dependency network.
}
\description{
This function returns an igraph object representing how
the targets in your workflow plan data frame
depend on each other.
(\code{help(package = "igraph")}). To plot the graph, call
to \code{\link[=plot.igraph]{plot.igraph()}} on your graph, or just use
\code{\link[=vis_drake_graph]{vis_drake_graph()}} from the start.
}
\examples{
\dontrun{
test_with_dir("Quarantine side effects.", {
load_basic_example() # Get the code with drake_example("basic").
# Make the igraph network connecting all the targets and imports.
g <- build_drake_graph(my_plan)
class(g) # "igraph"
})
}
}
\seealso{
\code{\link[=vis_drake_graph]{vis_drake_graph()}}
}
