% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache.R
\name{build_times}
\alias{build_times}
\title{Function \code{build_times}}
\usage{
build_times(path = getwd(), search = TRUE, digits = 0)
}
\arguments{
\item{path}{Root directory of the drake project,
or if \code{search} is \code{TRUE}, either the
project root or a subdirectory of the project.}

\item{search}{logical. If \code{TRUE}, search parent directories
to find the nearest drake cache. Otherwise, look in the
current working directory only.}

\item{digits}{How many digits to round the times to.}
}
\value{
data.frame of times from \code{\link{system.time}}
}
\description{
List all the build times. This doesn't include the amount of time 
spent loading and saving objects!
}
\seealso{
\code{\link{built}}
}
