% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_times.R
\name{build_times}
\alias{build_times}
\title{Function \code{build_times}}
\usage{
build_times(path = getwd(), search = TRUE, digits = 3,
  cache = get_cache(path = path, search = search, verbose = verbose),
  targets_only = FALSE, verbose = TRUE)
}
\arguments{
\item{path}{Root directory of the drake project,
or if \code{search} is \code{TRUE}, either the
project root or a subdirectory of the project.}

\item{search}{logical. If \code{TRUE}, search parent directories
to find the nearest drake cache. Otherwise, look in the
current working directory only.}

\item{digits}{How many digits to round the times to.}

\item{cache}{optional drake cache. If supplied,
the \code{path} and \code{search} arguments are ignored.}

\item{targets_only}{logical, whether to only return the
build times of the targets (exclude the imports).}

\item{verbose}{whether to print console messages}
}
\value{
data.frame of times from \code{\link{system.time}}
}
\description{
List all the build times.
This does not include the amount of time
spent loading and saving objects!
}
\examples{
\dontrun{
# Show the build times for the basic example.
load_basic_example() # Load the example.
make(my_plan) # Build all the targets.
build_times() # Show how long it took to build each target.
}
}
\seealso{
\code{\link{built}}
}
