% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{built}
\alias{built}
\title{List all the built targets (non-imports) in the cache.
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}}
\usage{
built(
  path = getwd(),
  search = TRUE,
  cache = drake::get_cache(path = path, search = search, verbose = verbose),
  verbose = 1L,
  jobs = 1
)
}
\arguments{
\item{path}{Path to a \code{drake} cache
(usually a hidden \verb{.drake/} folder) or \code{NULL}.}

\item{search}{Deprecated.}

\item{cache}{drake cache. See \code{\link[=new_cache]{new_cache()}}.
If supplied, \code{path} is ignored.}

\item{verbose}{Deprecated on 2019-09-11.}

\item{jobs}{Number of jobs/workers for parallel processing.}
}
\value{
Character vector naming the built targets in the cache.
}
\description{
Deprecated on 2019-01-08.
}
\details{
Targets are listed in the workflow plan
data frame (see \code{\link[=drake_plan]{drake_plan()}}.
}
\seealso{
\code{\link[=cached]{cached()}}, \code{\link[=loadd]{loadd()}}
}
\keyword{internal}
