% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache_interaction.R
\name{built}
\alias{built}
\title{Function \code{built}}
\usage{
built(path = getwd(), search = TRUE, cache = drake::get_cache(path = path,
  search = search))
}
\arguments{
\item{path}{Root directory of the drake project,
or if \code{search} is \code{TRUE}, either the
project root or a subdirectory of the project.}

\item{search}{logical. If \code{TRUE}, search parent directories
to find the nearest drake cache. Otherwise, look in the
current working directory only.}

\item{cache}{drake cache. See \code{\link{new_cache}()}.
If supplied, \code{path} and \code{search} are ignored.}
}
\value{
list of imported objects in the cache
}
\description{
List all the built (non-imported) objects in the drake
cache.
}
\examples{
\dontrun{
load_basic_example()
make(my_plan)
built()
}
}
\seealso{
\code{\link{cached}}, \code{\link{loadd}},
\code{link{imported}}
}
