% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-deprecate.R
\name{built}
\alias{built}
\title{Deprecated. List all the built targets (non-imports) in the cache.}
\usage{
built(path = getwd(), search = TRUE, cache = drake::get_cache(path =
  path, search = search, verbose = verbose), verbose = 1L, jobs = 1)
}
\arguments{
\item{path}{Root directory of the drake project,
or if \code{search} is \code{TRUE}, either the
project root or a subdirectory of the project.
Ignored if a \code{cache} is supplied.}

\item{search}{Logical. If \code{TRUE}, search parent directories
to find the nearest drake cache. Otherwise, look in the
current working directory only.
Ignored if a \code{cache} is supplied.}

\item{cache}{drake cache. See \code{\link[=new_cache]{new_cache()}}.
If supplied, \code{path} and \code{search} are ignored.}

\item{verbose}{Logical or numeric, control printing to the console.
\itemize{
\item \code{0} or \code{FALSE}: print nothing.
\item \code{1} or \code{TRUE}: print only targets to build.
\item \code{2}: plus checks and cache info.
\item \code{3}: plus missing imports.
\item \code{4}: plus all imports.
\item \code{5}: plus execution and total build times for targets.
\item \code{6}: plus notifications when targets are being stored.
}}

\item{jobs}{Number of jobs/workers for parallel processing.}
}
\value{
Character vector naming the built targets in the cache.
}
\description{
Deprecated on 2019-01-08.
}
\details{
Targets are listed in the workflow plan
data frame (see \code{\link[=drake_plan]{drake_plan()}}.
}
\seealso{
\code{\link[=cached]{cached()}}, \code{\link[=loadd]{loadd()}}
}
\keyword{internal}
