% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-deprecate.R
\name{cache_namespaces}
\alias{cache_namespaces}
\title{Deprecated. List all the \code{storr} cache namespaces used by drake.}
\usage{
cache_namespaces(default = storr::storr_environment()$default_namespace)
}
\arguments{
\item{default}{name of the default \code{storr} namespace}
}
\value{
A character vector of \code{storr} namespaces used for drake.
}
\description{
Deprecated on 2019-01-12.
}
\details{
Ordinary users do not need to worry about this function.
It is just another window into \code{drake}'s internals.
}
\seealso{
\code{\link[=make]{make()}}
}
\keyword{internal}
