% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache.R
\name{cache_namespaces}
\alias{cache_namespaces}
\title{Function cache_namespaces}
\usage{
cache_namespaces(default = storr::storr_environment()$default_namespace)
}
\arguments{
\item{default}{name of the default \code{storr} namespace}
}
\value{
A character vector of \code{storr} namespaces used for drake.
}
\description{
List the important \code{storr} cache namespaces
that need to be inventoried periodically in a call to \code{\link{make}()}.
Ordinary users do not need to worry about this function.
It is just another window into \code{drake}'s internals.
}
\examples{
cache_namespaces()
}
\seealso{
\code{\link{make}}
}
