% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache.R
\name{cache_namespaces}
\alias{cache_namespaces}
\title{Function drake_namespaces}
\usage{
cache_namespaces(default = storr_environment()$default_namespace)
}
\arguments{
\item{default}{name of the default namespace}
}
\description{
List the default namespaces of a \code{storr}
cache for drake.
}
\examples{
cache_namespaces()
}
