% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache_namespaces.R
\name{cache_namespaces}
\alias{cache_namespaces}
\title{Function \code{cache_namespaces}}
\usage{
cache_namespaces(default = storr::storr_environment()$default_namespace)
}
\arguments{
\item{default}{name of the default \code{storr} namespace}
}
\value{
A character vector of \code{storr} namespaces used for drake.
}
\description{
List all the \code{storr} cache namespaces used by drake.
Ordinary users do not need to worry about this function.
It is just another window into \code{drake}'s internals.
}
\examples{
cache_namespaces()
}
\seealso{
\code{\link{make}}
}
