% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache.R
\name{cache_path}
\alias{cache_path}
\title{Function cache_path}
\usage{
cache_path(cache = NULL)
}
\arguments{
\item{cache}{the cache whose file path
you want to know}
}
\description{
Returns the file path
where the cache is stored. Currently
only works with \code{storr} file system
caches.
}
