% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-deprecate.R
\name{cache_path}
\alias{cache_path}
\title{Deprecated. Return the file path where the cache is stored,
if applicable.}
\usage{
cache_path(cache = NULL)
}
\arguments{
\item{cache}{The cache whose file path you want to know.}
}
\value{
File path where the cache is stored.
}
\description{
Deprecated on 2019-01-12.
}
\details{
Currently only works with
\code{\link[storr:storr_rds]{storr::storr_rds()}} file system caches.
}
\keyword{internal}
