% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache.R
\name{cached}
\alias{cached}
\title{Function \code{cached}}
\usage{
cached(path = getwd(), search = FALSE)
}
\arguments{
\item{path}{Root directory of the drake project, 
or if \code{search} is \code{TRUE}, either the 
project root or a subdirectory of the project.}

\item{search}{If \code{TRUE}, search parent directories
to find the nearest drake cache. Otherwise, look in the
current working directory only.}
}
\value{
character vector of cached items
}
\description{
See the cached items stored by drake.
Load an item with \code{\link{readd}}.
}
\seealso{
\code{\link{built}}, \code{\link{imported}},
\code{\link{readd}}
\code{\link{plan}}, \code{\link{run}}
}

