% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache.R
\name{cached}
\alias{cached}
\title{Function \command{cached}}
\usage{
cached(path = getwd(), search = FALSE)
}
\arguments{
\item{path}{Root directory of the drake project, 
or if \command{search} is \command{TRUE}, either the 
project root or a subdirectory of the project.}

\item{search}{If \command{TRUE}, search parent directories
to find the nearest drake cache. Otherwise, look in the
current working directory only.}
}
\value{
character vector of cached items
}
\description{
See the cached items stored by drake.
Load an item with \command{\link{readd}}.
}
\seealso{
\command{\link{built}}, \command{\link{imported}},
\command{\link{readd}}
\command{\link{plan}}, \command{\link{run}}
}

