% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache.R
\name{cached}
\alias{cached}
\title{Function \code{cached}}
\usage{
cached(..., list = character(0), path = getwd(), search = FALSE)
}
\arguments{
\item{...}{objects to load from the cache, as names (unquoted)
or character strings (quoted). Similar to \code{...} in
\code{\link{remove}(...)}.}

\item{list}{character vector naming objects to be loaded from the
cache. Similar to the \code{list} argument of \code{\link{remove}()}.}

\item{path}{Root directory of the drake project, 
or if \code{search} is \code{TRUE}, either the 
project root or a subdirectory of the project.}

\item{search}{logical. If \code{TRUE}, search parent directories
to find the nearest drake cache. Otherwise, look in the
current working directory only.}
}
\value{
Either a named logical indicating whether the given
targets or cached or a character vector listing all cached
items, depending on whether any targets are specified
}
\description{
Check whether targets are in the cache.
If no targets are specified with \code{...} or \code{list}, 
then \code{cached()} lists 
all the items in the drake cache.
Read/load a cached item with \code{\link{readd}()} or 
\code{\link{loadd}()}.
}
\seealso{
\code{\link{built}}, \code{\link{imported}},
\code{\link{readd}}, \code{\link{loadd}},
\code{\link{plan}}, \code{\link{run}}
}

