% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecate.R
\name{check}
\alias{check}
\title{Deprecated function \code{check}}
\usage{
check(plan = drake_plan(), targets = drake::possible_targets(plan),
  envir = parent.frame(), cache = drake::get_cache(verbose = verbose),
  verbose = TRUE)
}
\arguments{
\item{plan}{Same as for \code{\link[=check_plan]{check_plan()}}.}

\item{targets}{Same as for \code{\link[=check_plan]{check_plan()}}.}

\item{envir}{Same as for \code{\link[=check_plan]{check_plan()}}.}

\item{cache}{Same as for \code{\link[=check_plan]{check_plan()}}.}

\item{verbose}{Same as for \code{\link[=check_plan]{check_plan()}}.}
}
\value{
Same as for \code{\link[=check_plan]{check_plan()}}.
}
\description{
Use \code{\link[=check_plan]{check_plan()}} instead.
}
\details{
Deprecated on 2017-11-12.
}
\examples{
# See ?check_plan for examples.
}
\seealso{
\code{\link[=check_plan]{check_plan()}}
}
\keyword{internal}
