% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.R
\name{check}
\alias{check}
\title{Function \code{check}}
\usage{
check(plan, targets = plan$target, envir = parent.frame())
}
\arguments{
\item{plan}{workflow plan data frame, possibly from 
\code{\link{plan}()}.}

\item{targets}{character vector of targets to make}

\item{envir}{environment containing user-defined functions}
}
\description{
Check a workflow plan, etc. for obvious 
errors such as circular dependencies and 
missing input files.
}
\seealso{
\code{link{plan}}, \code{\link{make}}
}

