% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{check_plan}
\alias{check_plan}
\title{Check a workflow plan data frame for obvious errors.
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}}
\usage{
check_plan(
  plan = NULL,
  targets = NULL,
  envir = parent.frame(),
  cache = drake::get_cache(verbose = verbose),
  verbose = 1L,
  jobs = 1
)
}
\arguments{
\item{plan}{Workflow plan data frame, possibly from
\code{\link[=drake_plan]{drake_plan()}}.}

\item{targets}{Character vector of targets to make.}

\item{envir}{Environment containing user-defined functions.}

\item{cache}{Optional drake cache. See \code{\link[=new_cache]{new_cache()}}.}

\item{verbose}{Deprecated on 2019-09-11.}

\item{jobs}{Number of jobs/workers for parallel processing.}
}
\value{
Invisibly return \code{plan}.
}
\description{
Deprecated on 2019-01-12.
}
\details{
Possible obvious errors include circular dependencies and
missing input files.
}
\seealso{
\code{\link[=drake_plan]{drake_plan()}}, \code{\link[=make]{make()}}
}
\keyword{internal}
