% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-check.R
\name{check_plan}
\alias{check_plan}
\title{Check a workflow plan data frame for obvious errors.}
\usage{
check_plan(plan = NULL, targets = NULL, envir = parent.frame(),
  cache = drake::get_cache(verbose = verbose), verbose = 1L,
  jobs = 1)
}
\arguments{
\item{plan}{workflow plan data frame, possibly from
\code{\link[=drake_plan]{drake_plan()}}.}

\item{targets}{character vector of targets to make}

\item{envir}{environment containing user-defined functions}

\item{cache}{optional drake cache. See \code{\link[=new_cache]{new_cache()}}.}

\item{verbose}{logical or numeric, control printing to the console.
\itemize{
\item \code{0} or \code{FALSE}: print nothing.
\item \code{1} or \code{TRUE}: print only targets to build.
\item \code{2}: plus checks and cache info.
\item \code{3}: plus missing imports.
\item \code{4}: plus all imports.
\item \code{5}: plus execution and total build times for targets.
\item \code{6}: plus notifications when targets are being stored.
}}

\item{jobs}{number of jobs/workers for parallel processing}
}
\value{
Invisibly return \code{plan}.
}
\description{
Possible obvious errors include
circular dependencies and
missing input files.
}
\examples{
\dontrun{
test_with_dir("Quarantine side effects.", {
load_mtcars_example() # Get the code with drake_example("mtcars").
check_plan(my_plan) # Check the workflow plan dataframe for obvious errors.
unlink("report.Rmd") # Remove an import file mentioned in the plan.
# If you un-suppress the warnings, check_plan()
# will tell you that 'report.Rmd' is missing.
suppressWarnings(check_plan(my_plan))
})
}
}
\seealso{
\code{\link[=drake_plan]{drake_plan()}}, \code{\link[=make]{make()}}
}
