% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean.R
\name{clean}
\alias{clean}
\title{Function \code{clean}}
\usage{
clean(destroy = FALSE)
}
\arguments{
\item{destroy}{logical, whether to totally remove the drake cache. 
If \code{destroy} is \code{FALSE}, only the targets 
from \code{run}()
are removed. If \code{TRUE}, the whole cache is removed, including
session metadata.}
}
\description{
Cleans up all work done by \code{\link{run}}. 
Your working directory (\code{\link{getwd}()}) must be the 
root directory of your drake project.
WARNING:
This deletes ALL \code{\link{run}} target, which includes 
file targets as well as the entire drake cache. Only use \code{clean}
if you're sure you won't lose any important work.
}
\seealso{
\code{\link{prune}}, \code{\link{run}},
}

