% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean.R
\name{clean}
\alias{clean}
\title{Function \code{clean}}
\usage{
clean(..., list = character(0), destroy = FALSE)
}
\arguments{
\item{...}{targets to remove from the cache, as names (unquoted)
or character strings (quoted). Similar to \code{...} in
\code{\link{remove}(...)}.}

\item{list}{character vector naming targets to be removed from the
cache. Similar to the \code{list} argument of \code{\link{remove}()}.}

\item{destroy}{logical, whether to totally remove the drake cache. 
If \code{destroy} is \code{FALSE}, only the targets 
from \code{run}()
are removed. If \code{TRUE}, the whole cache is removed, including
session metadata, etc.}
}
\description{
Cleans up all work done by \code{\link{run}()}. 
You must be in your project's working directory.
WARNING: This deletes ALL work done with \code{\link{run}()}, which includes 
file targets as well as the entire drake cache. Only use \code{clean()}
if you're sure you won't lose anything important.
}
\seealso{
\code{\link{prune}}, \code{\link{run}},
}

