% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean.R
\name{clean}
\alias{clean}
\title{Function \code{clean}}
\usage{
clean(..., list = character(0), destroy = FALSE, path = getwd(),
  search = TRUE, cache = NULL, verbose = TRUE, jobs = 1)
}
\arguments{
\item{...}{targets to remove from the cache, as names (unquoted)
or character strings (quoted). Similar to \code{...} in
\code{\link{remove}(...)}.
The symbols must not match other (formal) arguments of \code{clean()},
such as \code{destroy}, \code{cache}, \code{path}, \code{search},
\code{verbose}, or \code{jobs}. If there are name conflicts,
use the \code{list} argument instead of \code{...}.}

\item{list}{character vector naming targets to be removed from the
cache. Similar to the \code{list} argument of \code{\link{remove}()}.}

\item{destroy}{logical, whether to totally remove the drake cache.
If \code{destroy} is \code{FALSE}, only the targets
from \code{make}()
are removed. If \code{TRUE}, the whole cache is removed, including
session metadata, etc.}

\item{path}{Root directory of the drake project,
or if \code{search} is \code{TRUE}, either the
project root or a subdirectory of the project.}

\item{search}{logical. If \code{TRUE}, search parent directories
to find the nearest drake cache. Otherwise, look in the
current working directory only.}

\item{cache}{optional drake cache. See code{\link{new_cache}()}. If
If \code{cache} is supplied,
the \code{path} and \code{search} arguments are ignored.}

\item{verbose}{whether to print console messages}

\item{jobs}{Number of jobs for light parallelism
(disabled on Windows).}
}
\description{
Cleans up all work done by \code{\link{make}()}.
}
\details{
You must be in your project's working directory
or a subdirectory of it.
\code{clean(search = TRUE)} searches upwards in your folder structure
for the drake cache and acts on the first one it sees. Use
\code{search == FALSE} to look within the current working
directory only.
WARNING: This deletes ALL work done with \code{\link{make}()},
which includes
file targets as well as the entire drake cache. Only use \code{clean()}
if you're sure you won't lose anything important.
}
\examples{
\dontrun{
load_basic_example()
make(my_plan)
cached(no_imported_objects = TRUE)
clean(summ_regression1_large, small)
cached(no_imported_objects = TRUE)
make(my_plan)
clean(jobs = 2)
clean(destroy = TRUE)
}
}
\seealso{
\code{\link{prune}}, \code{\link{make}},
}
