% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{clean_main_example}
\alias{clean_main_example}
\title{Deprecated: clean the main example from \code{drake_example("main")}
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}}
\usage{
clean_main_example()
}
\value{
Nothing.
}
\description{
This function deletes files. Use at your own risk.
Destroys the \verb{.drake/} cache and the \code{report.Rmd} file
in the current working directory. Your working directory
(\code{getcwd()}) must be the folder from which you first ran
\code{load_main_example()} and \code{make(my_plan)}.
}
\details{
Deprecated 2018-12-31.
}
\keyword{internal}
