% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecate.R
\name{clean_main_example}
\alias{clean_main_example}
\title{Deprecated: clean the main example from \code{drake_example("main")}}
\usage{
clean_main_example()
}
\value{
Nothing.
}
\description{
This function deletes files. Use at your own risk.
Destroys the \code{.drake/} cache and the \code{report.Rmd} file
in the current working directory. Your working directory
(\code{getcwd()}) must be the folder from which you first ran
\code{load_main_example()} and \code{make(my_plan)}.
}
\details{
Deprecated December 2018.
}
\examples{
\dontrun{
# The code for this example is hosted at
# https://github.com/wlandau/drake-examples/tree/master/main
# You can download it with drake_example("main")
# or watch a video tutorial about it at
# https://ropenscilabs.github.io/drake-manual/.
}
}
\keyword{internal}
