% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/examples.R
\name{clean_main_example}
\alias{clean_main_example}
\title{Clean the main example from \code{drake_example("main")}}
\usage{
clean_main_example()
}
\value{
Nothing.
}
\description{
This function deletes files. Use at your own risk.
Destroys the \code{.drake/} cache and the \code{report.Rmd} file
in the current working directory. Your working directory
(\code{getcwd()}) must be the folder from which you first ran
\code{load_main_example()} and \code{make(my_plan)}.
}
\examples{
\dontrun{
test_with_dir("Quarantine side effects.", {
if (requireNamespace("downloader")) {
# Populate your workspace and write 'report.Rmd' and 'raw_data.xlsx'.
load_main_example() # Get the code: drake_example("main")
make(plan)
readd(hist) # Show the ggplot2 histogram.
# Clean up the example.
clean_main_example()
}
})
}
}
\seealso{
\code{\link[=load_main_example]{load_main_example()}}, \code{\link[=clean]{clean()}}
}
