% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-deprecate.R
\name{cleaned_namespaces}
\alias{cleaned_namespaces}
\title{Deprecated utility function}
\usage{

  cleaned_namespaces(default = storr::storr_environment()$default_namespace)
}
\arguments{
\item{default}{Name of the default \code{storr} namespace.}
}
\value{
A character vector of \code{storr} namespaces
that are cleaned during \code{\link[=clean]{clean()}}.
}
\description{
2019-02-13
}
\keyword{internal}
