% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache.R
\name{cleaned_namespaces}
\alias{cleaned_namespaces}
\title{Function cleaned_namespaces}
\usage{
cleaned_namespaces(default = storr::storr_environment()$default_namespace)
}
\arguments{
\item{default}{Name of the default \code{storr} namespace.}
}
\value{
A character vector of \code{storr} namespaces
that are cleaned during \code{\link{clean}()}.
}
\description{
List the \code{storr} namespaces that are cleaned
during a call to \code{\link{clean}()}. All these
namespaces store target-level data, but not all
target-level namespaces are cleaned during
\code{\link{clean}()}.
}
\examples{
cleaned_namespaces()
}
\seealso{
\code{\link{cache_namespaces}}, \code{\link{clean}}
}
