% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backend_clustermq.R
\name{cmq_build}
\alias{cmq_build}
\title{Build a target using the clustermq backend
\lifecycle{maturing}}
\usage{
cmq_build(target, meta, deps, spec, ht_is_subtarget, config)
}
\arguments{
\item{target}{Target name.}

\item{meta}{List of metadata.}

\item{deps}{Named list of target dependencies.}

\item{spec}{Internal, part of the full \code{config$spec}.}

\item{ht_is_subtarget}{Internal, part of \code{config}}

\item{config}{A \code{\link[=drake_config]{drake_config()}} list.}
}
\description{
For internal use only
}
\keyword{internal}
