% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backend-clustermq.R
\name{cmq_build}
\alias{cmq_build}
\title{Build a target using the clustermq backend}
\usage{
cmq_build(target, meta, deps, config)
}
\arguments{
\item{target}{target name}

\item{meta}{list of metadata}

\item{deps}{named list of target dependencies}

\item{config}{a \code{\link[=drake_config]{drake_config()}} list}
}
\description{
For internal use only
}
\keyword{internal}
