% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecate.R
\name{config}
\alias{config}
\title{Deprecated function \code{config}}
\usage{
config(plan = read_drake_plan(), targets = plan$target,
  envir = parent.frame(), verbose = 1, hook = default_hook,
  cache = drake::get_cache(verbose = verbose, force = force),
  parallelism = drake::default_parallelism(), jobs = 1,
  packages = rev(.packages()), prework = character(0),
  prepend = character(0), command = drake::default_Makefile_command(),
  args = drake::default_Makefile_args(jobs = jobs, verbose = verbose),
  recipe_command = drake::default_recipe_command(), timeout = Inf,
  cpu = timeout, elapsed = timeout, retries = 0, force = FALSE,
  log_progress = FALSE, graph = NULL, trigger = "any",
  skip_imports = FALSE)
}
\arguments{
\item{plan}{same as for \code{\link[=make]{make()}}}

\item{targets}{same as for \code{\link[=make]{make()}}}

\item{envir}{same as for \code{\link[=make]{make()}}}

\item{verbose}{same as for \code{\link[=make]{make()}}}

\item{hook}{same as for \code{\link[=make]{make()}}}

\item{cache}{same as for \code{\link[=make]{make()}}}

\item{parallelism}{same as for \code{\link[=make]{make()}}}

\item{jobs}{same as for \code{\link[=make]{make()}}}

\item{packages}{same as for \code{\link[=make]{make()}}}

\item{prework}{same as for \code{\link[=make]{make()}}}

\item{prepend}{same as for \code{\link[=make]{make()}}}

\item{command}{same as for \code{\link[=make]{make()}}}

\item{args}{same as for \code{\link[=make]{make()}}}

\item{recipe_command}{same as for \code{\link[=make]{make()}}}

\item{timeout}{same as for \code{\link[=make]{make()}}}

\item{cpu}{same as for \code{\link[=make]{make()}}}

\item{elapsed}{same as for \code{\link[=make]{make()}}}

\item{retries}{same as for \code{\link[=make]{make()}}}

\item{force}{same as for \code{\link[=make]{make()}}}

\item{log_progress}{same as for \code{\link[=drake_config]{drake_config()}}}

\item{graph}{same as for \code{\link[=drake_config]{drake_config()}}}

\item{trigger}{same as for \code{\link[=make]{make()}}}

\item{skip_imports}{same as for \code{\link[=drake_config]{drake_config()}}}
}
\value{
The master internal configuration list of a project.
}
\description{
Use \code{\link[=drake_config]{drake_config()}} instead.
}
\details{
Deprecated on 2017-11-12.
}
\examples{
# See ?drake_config for the examples.
}
\seealso{
\code{\link[=drake_config]{drake_config()}}

\code{\link[=drake_config]{drake_config()}}
}
\keyword{internal}
