% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecate.R
\name{dataframes_graph}
\alias{dataframes_graph}
\title{Deprecated function \code{dataframes_graph}}
\usage{
dataframes_graph(config = drake::read_drake_config(), from = NULL,
  mode = c("out", "in", "all"), order = NULL, subset = NULL,
  build_times = "build", digits = 3, targets_only = FALSE,
  split_columns = NULL, font_size = 20, from_scratch = FALSE,
  make_imports = TRUE, full_legend = TRUE)
}
\arguments{
\item{config}{deprecated}

\item{from}{deprecated}

\item{mode}{deprecated}

\item{order}{deprecated}

\item{subset}{deprecated}

\item{build_times}{deprecated}

\item{digits}{deprecated}

\item{targets_only}{deprecated}

\item{split_columns}{deprecated}

\item{font_size}{deprecated}

\item{from_scratch}{deprecated}

\item{make_imports}{deprecated}

\item{full_legend}{deprecated}
}
\value{
\code{args} for \code{\link{system2}(command, args)}
}
\description{
Use \code{\link[=drake_graph_info]{drake_graph_info()}} instead.
}
\details{
Deprecated on 2018-06-27.
}
\examples{
# See ?drake_graph_info for examples.
}
\keyword{internal}
