% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotgraph.R
\name{dataframes_graph}
\alias{dataframes_graph}
\title{Function \code{dataframes_graph}}
\usage{
dataframes_graph(plan, targets = drake::possible_targets(plan),
  envir = parent.frame(), verbose = FALSE, font_size = 20, graph = NULL)
}
\arguments{
\item{plan}{workflow plan data frame, same as for function 
\code{\link{make}()}.}

\item{targets}{names of targets to bulid, same as for function
\code{\link{make}()}.}

\item{envir}{environment to import from, same as for function
\code{\link{make}()}.}

\item{verbose}{logical, whether to output messages to the console.}

\item{font_size}{numeric, font size of the node labels in the graph}

\item{graph}{an igraph object if one has already been built with 
\code{\link{build_graph}()}.}
}
\value{
a list of three data frames: one for nodes, one for edges, and one for
the legend/key nodes.
}
\description{
Get the information about nodes, edges, and the legend/key
so you can plot your own custom \code{visNetwork}.
IMPORTANT: you must be in the root directory of your project.
}
\examples{
\dontrun{
load_basic_example()
raw_graph = dataframes_graph(my_plan)
str(raw_graph)
# Plot your own custom visNetwork graph
library(magrittr)
library(visNetwork)
visNetwork(nodes = raw_graph$nodes, edges = raw_graph$edges) \%>\%
  visLegend(useGroups = FALSE, addNodes = raw_graph$legend_nodes) \%>\% 
  visHierarchicalLayout(direction = "LR")
}
}
\seealso{
\code{\link{plot_graph}}, \code{\link{build_graph}}
}
