% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate.R
\name{dataset_wildcard}
\alias{dataset_wildcard}
\title{Function dataset_wildcard}
\usage{
dataset_wildcard()
}
\value{
The dataset wildcard used in
\code{\link{plan_analyses}()} and \code{\link{plan_summaries}()}.
}
\description{
Show the dataset wildcard
used in \code{\link{plan_analyses}()} and \code{\link{plan_summaries}()}.
}
\examples{
# See ?plan_analyses for examples
}
\seealso{
\code{\link{plan_analyses}()}
}
