% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate.R
\name{dataset_wildcard}
\alias{dataset_wildcard}
\title{Show the dataset wildcard
used in \code{\link[=plan_analyses]{plan_analyses()}} and \code{\link[=plan_summaries]{plan_summaries()}}.}
\usage{
dataset_wildcard()
}
\value{
The dataset wildcard used in
\code{\link[=plan_analyses]{plan_analyses()}} and \code{\link[=plan_summaries]{plan_summaries()}}.
}
\description{
Used to generate workflow plan data frames.
}
\examples{
# See ?plan_analyses for examples
}
\seealso{
\code{\link[=plan_analyses]{plan_analyses()}}
}
