% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api-debug.R
\name{debug_and_run}
\alias{debug_and_run}
\title{Run a function in debug mode.}
\usage{
debug_and_run(f)
}
\arguments{
\item{f}{A function.}
}
\value{
The return value of \code{f}.
}
\description{
Internal function for \code{\link[=drake_debug]{drake_debug()}}. Not for general use.
}
\seealso{
\code{\link[=drake_debug]{drake_debug()}}
}
\keyword{internal}
