% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache.R
\name{default_cache_path}
\alias{default_cache_path}
\title{Function default_cache_path}
\usage{
default_cache_path()
}
\description{
Return the default file path of the drake/storr cache.
}
\examples{
default_cache_path()
}
