% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache.R
\name{default_cache_path}
\alias{default_cache_path}
\title{Return the default file path of the drake/storr cache.}
\usage{
default_cache_path(root = getwd())
}
\arguments{
\item{root}{character scalar, path to the root of the
\code{drake} project with the cache.}
}
\value{
Default file path of the drake/storr cache.
}
\description{
Applies to file system caches only.
}
\examples{
\dontrun{
default_cache_path()
}
}
\keyword{internal}
