% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drake_graph_info_utils.R
\name{default_graph_title}
\alias{default_graph_title}
\title{Return the default title for graph visualizations}
\usage{
default_graph_title(split_columns = FALSE)
}
\arguments{
\item{split_columns}{deprecated}
}
\value{
A character scalar with the default graph title.
}
\description{
For internal use only.
}
\examples{
default_graph_title()
}
\keyword{internal}
