% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drake_graph_info_utils.R
\name{default_graph_title}
\alias{default_graph_title}
\title{Return the default title of the graph for
\code{\link[=vis_drake_graph]{vis_drake_graph()}}.}
\usage{
default_graph_title(split_columns = FALSE)
}
\arguments{
\item{split_columns}{deprecated}
}
\value{
a character scalar with the default graph title for
\code{\link[=vis_drake_graph]{vis_drake_graph()}}.
}
\description{
For internal use only.
}
\examples{
default_graph_title()
}
\seealso{
\code{\link[=drake_graph_info]{drake_graph_info()}}, \code{\link[=vis_drake_graph]{vis_drake_graph()}}
}
\keyword{internal}
