% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataframes_graph_utils.R
\name{default_graph_title}
\alias{default_graph_title}
\title{Return the default title of the graph for
\code{\link[=vis_drake_graph]{vis_drake_graph()}}.}
\usage{
default_graph_title(split_columns = FALSE)
}
\arguments{
\item{split_columns}{logical, whether the columns were split
in \code{\link[=dataframes_graph]{dataframes_graph()}} or \code{\link[=vis_drake_graph]{vis_drake_graph()}}
with the \code{split_columns} argument.}
}
\value{
a character scalar with the default graph title for
\code{\link[=vis_drake_graph]{vis_drake_graph()}}.
}
\description{
For internal use only.
}
\examples{
default_graph_title()
}
\seealso{
\code{\link[=dataframes_graph]{dataframes_graph()}}, \code{\link[=vis_drake_graph]{vis_drake_graph()}}
}
\keyword{internal}
