% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataframes_graph_utils.R
\name{default_graph_title}
\alias{default_graph_title}
\title{Function default_graph_title}
\usage{
default_graph_title(parallelism = drake::parallelism_choices(), split_columns)
}
\arguments{
\item{parallelism}{Mode of parallelism intended for the workflow.
See \code{\link{parallelism_choices}()}.}

\item{split_columns}{logical, whether the columns were split
in \code{\link{dataframes_graph}()} or \code{\link{plot_graph}()}
with the \code{split_columns} argument.}
}
\description{
Default title of the graph from
\code{\link{plot_graph}()}.
}
\seealso{
\code{\link{dataframes_graph}}, \code{\link{plot_graph}}
}
