% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hooks.R
\name{default_hook}
\alias{default_hook}
\title{Function default_hook}
\usage{
default_hook(code)
}
\arguments{
\item{code}{Placeholder for the code to build a target/import.}
}
\description{
The default \code{hook} argument to \code{\link{make}()}.
}
\examples{
\dontrun{
load_basic_example()
make(my_plan, hook = default_hook) # Nothing gets built!
cached() # character(0) # nolint
}
}
