% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hooks.R
\name{default_hook}
\alias{default_hook}
\title{Function default_hook}
\usage{
default_hook(code)
}
\arguments{
\item{code}{Placeholder for the code to build a target/import.}
}
\value{
A function that you can supply to the \code{hook} argument
of \code{\link{make}()}.
}
\description{
The default \code{hook} argument to \code{\link{make}()}.
}
\examples{
\dontrun{
test_with_dir("Quarantine side effects.", {
load_basic_example() # Load the canonical drake example.
# Everything gets built normally.
make(my_plan, hook = default_hook)
cached() # List the cached targets and imports.
})
}
}
