% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/choose_hash_algo.R
\name{default_long_hash_algo}
\alias{default_long_hash_algo}
\title{Default long hash algorithm}
\usage{
default_long_hash_algo()
}
\description{
This hash is long in order
to avoid collisions. It is used
used for hashes that are not file names,
so we do not need to worry about it being too long.
}
