% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Makefile.R
\name{default_Makefile_args}
\alias{default_Makefile_args}
\title{Function \code{default_Makefile_args}}
\usage{
default_Makefile_args(jobs, verbose)
}
\arguments{
\item{jobs}{number of jobs}

\item{verbose}{logical, whether to be verbose}
}
\value{
\code{args} for \code{\link{system2}(command, args)}
}
\description{
Configures default
arguments to \code{\link{system2}()} to run Makefiles.
}
\examples{
default_Makefile_args(jobs = 2, verbose = FALSE)
default_Makefile_args(jobs = 4, verbose = TRUE)
}
