% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecate.R
\name{default_Makefile_args}
\alias{default_Makefile_args}
\title{Deprecated}
\usage{
default_Makefile_args(jobs, verbose)
}
\arguments{
\item{jobs}{number of jobs}

\item{verbose}{logical or numeric, control printing to the console.
\itemize{
\item \code{0} or \code{FALSE}: print nothing.
\item \code{1} or \code{TRUE}: print only targets to build.
\item \code{2}: plus checks and cache info.
\item \code{3}: plus missing imports.
\item \code{4}: plus all imports.
\item \code{5}: plus execution and total build times for targets.
\item \code{6}: plus notifications when targets are being stored.
}}
}
\value{
\code{args} for \code{system2(command, args)}
}
\description{
2019-01-03
}
\examples{
# deprecated
}
\keyword{internal}
