% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Makefile.R
\name{default_Makefile_command}
\alias{default_Makefile_command}
\title{Function \code{default_Makefile_command}}
\usage{
default_Makefile_command()
}
\value{
A character scalar naming a Linux/Unix command
to run a Makefile.
}
\description{
Give the default \code{command}
argument to \code{\link{make}()}. Relevant for
\code{"Makefile"} parallelism only.
}
\examples{
default_Makefile_command()
}
