% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Makefile.R
\name{default_Makefile_command}
\alias{default_Makefile_command}
\title{Function \code{default_Makefile_command}}
\usage{
default_Makefile_command()
}
\description{
Give the default \code{command}
argument to \code{\link{make}()}
}
\examples{
default_Makefile_command()
}
