% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-deprecate.R
\name{default_Makefile_command}
\alias{default_Makefile_command}
\title{Deprecated}
\usage{
default_Makefile_command()
}
\value{
A character scalar
}
\description{
2019-01-03
}
\examples{
# deprecated
}
\keyword{internal}
