% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{default_parallelism}
\alias{default_parallelism}
\title{Default parallel backend
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}}
\usage{
default_parallelism()
}
\value{
character
}
\description{
2019-01-02
}
\keyword{internal}
