% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parallel_ui.R
\name{default_parallelism}
\alias{default_parallelism}
\title{Function \code{default_parallelism}}
\usage{
default_parallelism()
}
\value{
The default parallelism option for your system.
}
\description{
Show the default parallelism for \code{\link{make}()}
on your system:
\code{'parLapply'} for Windows machines and \code{'mclapply'}
for other platforms.
}
\examples{
default_parallelism()
}
\seealso{
\code{\link{make}}, \code{\link{shell_file}}
}
