% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parallel_ui.R
\name{default_parallelism}
\alias{default_parallelism}
\title{Show the default \code{parallelism} argument
to \code{\link[=make]{make()}} for your system.}
\usage{
default_parallelism()
}
\value{
The default parallelism option for your system.
}
\description{
Returns \code{'parLapply'} for Windows machines
and \code{'mclapply'} for other platforms.
}
\examples{
default_parallelism()
}
\seealso{
\code{\link[=make]{make()}}, \code{\link[=shell_file]{shell_file()}}
}
