% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parallel_ui.R
\name{default_parallelism}
\alias{default_parallelism}
\title{Function \code{default_parallelism}}
\usage{
default_parallelism()
}
\value{
default parallelism option for the current platform
}
\description{
Default parallelism for \code{\link{make}()}:
\code{"parLapply"} for Windows machines and \code{"mclapply"}
for other platforms.
}
\examples{
default_parallelism()
}
\seealso{
\code{\link{make}}, \code{\link{shell_file}}
}
