% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{default_recipe_command}
\alias{default_recipe_command}
\title{Default Makefile recipe command
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}}
\usage{
default_recipe_command()
}
\value{
A character scalar with the default recipe command.
}
\description{
2019-01-02
}
\keyword{internal}
