% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecate.R
\name{default_recipe_command}
\alias{default_recipe_command}
\title{Deprecated}
\usage{
default_recipe_command()
}
\value{
A character scalar with the default recipe command.
}
\description{
See the help file of \code{\link[=Makefile_recipe]{Makefile_recipe()}}
for details and examples.
}
\examples{
# deprecated
}
