% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recipe.R
\name{default_recipe_command}
\alias{default_recipe_command}
\title{Show the default
recipe command for \code{\link{make}(..., parallelism = "Makefile")}.}
\usage{
default_recipe_command()
}
\value{
A character scalar with the default recipe command.
}
\description{
See the help file of \code{\link[=Makefile_recipe]{Makefile_recipe()}}
for details and examples.
}
\examples{
default_recipe_command()
}
\seealso{
\code{\link[=Makefile_recipe]{Makefile_recipe()}}
}
