% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Makefile.R
\name{default_recipe_command}
\alias{default_recipe_command}
\title{default_recipe_command}
\usage{
default_recipe_command()
}
\description{
Function to give the default
recipe command for Makefile parallelism.
}
\examples{
default_recipe_command()
\dontrun{
load_basic_example()
# Look at the Makefile generated by the following.
make(my_plan, paralleliem = "Makefile")
# Generates a Makefile with "R -q" rather than
# Rscript -e.
make(my_plan, parallelism = "Makefile",
  recipe_command = "R -q -e")
}
}
