% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecate.R
\name{default_recipe_command}
\alias{default_recipe_command}
\title{Deprecated}
\usage{
default_recipe_command()
}
\value{
A character scalar with the default recipe command.
}
\description{
2019-01-02
}
\examples{
# deprecated
}
\keyword{internal}
