% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/choose_hash_algo.R
\name{default_short_hash_algo}
\alias{default_short_hash_algo}
\title{Default short hash algorithm}
\usage{
default_short_hash_algo()
}
\description{
This is the hash algorithm
used most of the time. These hashes
are used as file names in the cache
and elsewhere, so they cannot be too long.
Windows has an upper limit on file path sizes.
}
