% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-deprecate.R
\name{default_short_hash_algo}
\alias{default_short_hash_algo}
\title{Deprecated. Return the default short hash algorithm for \code{make()}.}
\usage{
default_short_hash_algo(cache = NULL)
}
\arguments{
\item{cache}{optional drake cache.
When you \code{\link[=configure_cache]{configure_cache()}} without
supplying a short hash algorithm,
\code{default_short_hash_algo(cache)} is the short
hash algorithm that drake picks for you.}
}
\value{
A character vector naming a hash algorithm.
}
\description{
Deprecated. drake now only uses one hash algorithm per cache.
}
\details{
Deprecated on 2018-12-12
}
\examples{
# deprecated
}
\keyword{internal}
