% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecate.R
\name{default_system2_args}
\alias{default_system2_args}
\title{Deprecated function \code{default_system2_args}}
\usage{
default_system2_args(jobs, verbose)
}
\arguments{
\item{jobs}{number of jobs}

\item{verbose}{logical, whether to be verbose}
}
\value{
\code{args} for \code{\link{system2}(command, args)}
}
\description{
Use \code{\link[=default_Makefile_args]{default_Makefile_args()}} instead.
}
\details{
Deprecated on 2017-10.
}
\examples{
# See ?default_Makefile_args for examples.
}
\seealso{
\code{\link[=default_Makefile_args]{default_Makefile_args()}}
}
\keyword{internal}
