% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/triggers.R
\name{default_trigger}
\alias{default_trigger}
\title{Return the default trigger.}
\usage{
default_trigger()
}
\value{
A character scalar naming the default trigger.
}
\description{
Triggers are target-level rules
that tell \code{\link{make}()} how to check if
target is up to date or outdated.
}
\examples{
default_trigger()
# See ?triggers for more examples.
}
\seealso{
\code{\link{triggers}}, \code{\link{make}}
}
