% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/console.R
\name{default_verbose}
\alias{default_verbose}
\title{Default verbosity for \code{drake}}
\usage{
default_verbose()
}
\value{
a logical or integer with the value of
the default \code{verbose} argument to \code{drake} functions.
}
\description{
Set with \code{pkgconfig}: for example,
\code{pkgconfig::set_config("drake::verbose" = 2)}.
}
\keyword{internal}
