% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dependencies.R
\name{dependency_profile}
\alias{dependency_profile}
\title{Function \code{dependency_profile}}
\usage{
dependency_profile(target, config)
}
\arguments{
\item{target}{name of the target}

\item{config}{configuration list output by
\code{\link{config}} or \code{\link{make}}}
}
\value{
A list of information that drake takes into account
when examining the dependencies of the target.
}
\description{
Return the detailed dependency profile
of the target. Useful for debugging.
For up to date targets, like elements
of the returned list should agree: for example,
\code{cached_dependency_hash} and
\code{current_dependency_hash}.
}
\examples{
\dontrun{
test_with_dir("Quarantine side effects.", {
load_basic_example() # Load drake's canonical exmaple.
con <- make(my_plan) # Run the project, build the targets.
# Get some example dependency profiles of targets.
dependency_profile("small", config = con)
dependency_profile("'report.md'", config = con)
clean(large) # Clear out most of the information about the target 'large'.
# The dependency hashes are still there.
dependency_profile("large", config = con)
# Should agree with `$hashes_of_dependencies`.
con$cache$get_hash("simulate",
  namespace = "kernels")
})
}
}
\seealso{
\code{\link{deps}}, \code{\link{make}},
\code{\link{config}}
}
