% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dependencies.R
\name{dependency_profile}
\alias{dependency_profile}
\title{Function dependency_profile}
\usage{
dependency_profile(target, config)
}
\arguments{
\item{target}{name of the target}

\item{config}{configuration list output by
\code{\link{config}} or \code{\link{make}}}
}
\description{
Return the detailed dependency profile
of the target. Useful for debugging.
For up to date targets, like elements
of the returned list should agree: for example,
\code{cached_dependency_hash} and
\code{current_dependency_hash}.
}
\examples{
\dontrun{
load_basic_example()
con <- make(my_plan)
dependency_profile("small", config = con)
dependency_profile("'report.md'", config = con)
clean(large) # most information gone
dependency_profile("large", config = con) # hashes of deps still available
# should agree with `$meta_of_dependencies`
con$cache$get_hash("simulate",
  namespace = "reproducibly_tracked")
}
}
\seealso{
\code{\link{deps}}, \code{\link{make}},
\code{\link{config}}
}
