% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dependencies.R
\name{dependency_profile}
\alias{dependency_profile}
\title{Find out why a target is out of date.}
\usage{
dependency_profile(target, config = drake::read_drake_config(),
  character_only = FALSE)
}
\arguments{
\item{target}{name of the target}

\item{config}{configuration list output by
\code{\link[=drake_config]{drake_config()}} or \code{\link[=make]{make()}}}

\item{character_only}{logical, whether to assume \code{target}
is a character string rather than a symbol}
}
\value{
A data frame of the old hashes and
new hashes of each dependency and trigger, along with
a logical column to indicate which hashes changed since
the last \code{\link[=make]{make()}}.
}
\description{
The dependency profile can tell you
why a target is out of date. It tells you the specific
dependencies that changed since the last \code{\link[=make]{make()}},
as well as whether any of the built-in \code{\link[=trigger]{trigger()}}s
(the command, depend, and file triggers) will be activated.
}
\examples{
\dontrun{
test_with_dir("Quarantine side effects.", {
load_mtcars_example() # Load drake's canonical example.
config <- make(my_plan) # Run the project, build the targets.
# Get some example dependency profiles of targets.
dependency_profile(small, config = config)
# Change a dependency.
simulate <- function(x){}
# Update the in-memory imports in the cache
# so dependency_profile can detect changes to them.
# Changes to targets are already cached.
make(my_plan, skip_targets = TRUE)
# The dependency hash changed.
dependency_profile(small, config = config)
})
}
}
\seealso{
\code{\link[=diagnose]{diagnose()}},
\code{\link[=deps_code]{deps_code()}}, \code{\link[=make]{make()}},
\code{\link[=drake_config]{drake_config()}}
}
