% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-deprecate.R
\name{deps}
\alias{deps}
\title{Deprecated.
List the dependencies of a function, workflow plan command,
or knitr report source file.}
\usage{
deps(x)
}
\arguments{
\item{x}{Either a function or a string.
Strings are commands from your workflow plan data frame.}
}
\value{
A character vector, names of dependencies.
Files wrapped in single quotes.
The other names listed are functions or generic R objects.
}
\description{
Deprecated. Use \code{\link[=deps_code]{deps_code()}} or \code{\link[=deps_target]{deps_target()}} instead.
These functions are intended for debugging and checking your project.
The dependency structure of the components of your analysis
decides which targets are built and when.
}
\details{
Deprecated on 2018-05-08.
}
\examples{
# See deps_code() for examples.
}
\keyword{internal}
