% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-deprecate.R
\name{deps_targets}
\alias{deps_targets}
\title{Deprecated.}
\usage{
deps_targets(targets, config, reverse = FALSE)
}
\arguments{
\item{targets}{A character vector of target names.}

\item{config}{An output list from \code{\link[=drake_config]{drake_config()}}}

\item{reverse}{Logical, whether to compute reverse dependencies
(targets immediately downstream) instead of ordinary dependencies.}
}
\value{
Names of dependencies listed by type (object, input file, etc).
}
\description{
Deprecated. Use \code{\link[=deps_target]{deps_target()}} (singular) instead.
}
\details{
Deprecated on 2018-08-30.
}
\examples{
\dontrun{
test_with_dir("Quarantine side effects.", {
load_mtcars_example() # Get the code with drake_example("mtcars").
config <- drake_config(my_plan)
deps_targets("regression1_small", config = config)
deps_targets(c("small", "large"), config = config, reverse = TRUE)
})
}
}
\keyword{internal}
