% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{do_prework}
\alias{do_prework}
\title{Internal function do_prework}
\usage{
do_prework(config, verbose_packages)
}
\arguments{
\item{config}{internal configuration list}

\item{verbose_packages}{logical, whether to print
package startup messages}
}
\description{
Run the \code{prework} of a \code{\link{make}()}.
For internal use only.
the only reason this function is exported
is to set up PSOCK clusters efficiently.
}
