% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-deprecate.R
\name{drake_batchtools_tmpl_file}
\alias{drake_batchtools_tmpl_file}
\title{Deprecated. Get a template file for execution on a cluster.}
\usage{
drake_batchtools_tmpl_file(example = drake::drake_hpc_template_files(),
  to = getwd(), overwrite = FALSE)
}
\arguments{
\item{example}{name of template file}

\item{to}{Character vector, where to write the file.}

\item{overwrite}{Logical, whether to overwrite an existing file of the
same name.}
}
\description{
Deprecated. Use \code{\link[=drake_hpc_template_file]{drake_hpc_template_file()}} instead.
}
\details{
Deprecated on 2018-06-27.
}
\examples{
# See drake_hpc_template_file() for examples.
}
\keyword{internal}
