% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build.R
\name{drake_build}
\alias{drake_build}
\title{Build/process a single target or import.}
\usage{
drake_build(target, meta, config)
}
\arguments{
\item{target}{name of the target}

\item{meta}{list of metadata that tell which
targets are up to date (from \code{drake:::meta()}).}

\item{config}{internal configuration list}
}
\value{
The value of the target right after it is built.
}
\description{
For internal use only.
the only reason this function is exported
is to set up parallel socket (PSOCK) clusters
without much of a fuss.
}
\examples{
\dontrun{
test_with_dir("Quarantine side effects.", {
# This example is not really a user-side demonstration.
# It just walks through a dive into the internals.
# Populate your workspace and write 'report.Rmd'.
load_basic_example()
# Create the master internal configuration list.
config <- drake_config(my_plan)
# Compute metadata on 'small', including a hash/fingerprint
# of the dependencies.
meta <- drake:::meta(target = "small", config = config)
# Should not yet include 'small'.
cached()
# Build 'small'
drake_build(target = "small", meta = meta, config = config)
# Should now include 'small'
cached()
readd(small)
})
}
}
\keyword{internal}
