% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api-build.R
\name{drake_build}
\alias{drake_build}
\title{Build/process a single target or import.}
\usage{
drake_build(target, config = NULL, meta = NULL,
  character_only = FALSE, envir = NULL, jobs = 1, replace = FALSE)
}
\arguments{
\item{target}{Name of the target.}

\item{config}{Internal configuration list.}

\item{meta}{Deprecated.}

\item{character_only}{Logical, whether \code{name} should be treated
as a character or a symbol
(just like \code{character.only} in \code{\link[=library]{library()}}).}

\item{envir}{Environment to load objects into. Defaults to the
calling environment (current workspace).}

\item{jobs}{Number of parallel jobs for loading objects. On
non-Windows systems, the loading process for multiple objects
can be lightly parallelized via \code{parallel::mclapply()}.
just set jobs to be an integer greater than 1. On Windows,
\code{jobs} is automatically demoted to 1.}

\item{replace}{Logical. If \code{FALSE},
items already in your environment
will not be replaced.}
}
\value{
The value of the target right after it is built.
}
\description{
Also load the target's dependencies beforehand.
}
\examples{
\dontrun{
test_with_dir("Quarantine side effects.", {
if (suppressWarnings(require("knitr"))) {
# This example is not really a user-side demonstration.
# It just walks through a dive into the internals.
# Populate your workspace and write 'report.Rmd'.
load_mtcars_example() # Get the code with drake_example("mtcars").
# Create the master internal configuration list.
config <- drake_config(my_plan)
out <- drake_build(small, config = config)
# Now includes `small`.
cached()
head(readd(small))
# `small` was invisibly returned.
head(out)
# If you previously called make(),
# `config` is just read from the cache.
make(my_plan, verbose = FALSE)
config <- drake_config(my_plan)
result <- drake_build(small, config = config)
head(result)
}
})
}
}
\seealso{
\code{\link[=drake_debug]{drake_debug()}}
}
