% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build.R
\name{drake_build}
\alias{drake_build}
\title{Internal function drake_build}
\usage{
drake_build(target, meta_list, config)
}
\arguments{
\item{target}{name of the target}

\item{meta_list}{list of metadata that tell which
targets are up to date}

\item{config}{internal configuration list}
}
\value{
The value of the target right after it is built.
}
\description{
Function to build a target or import.
For internal use only.
the only reason this function is exported
is to set up PSOCK clusters efficiently.
}
\examples{
\dontrun{
# This example is not really a user-side demonstration.
# It just walks through a dive into the internals.
# Populate your workspace and write 'report.Rmd'.
load_basic_example()
# Create the master internal configuration list.
config <- drake_config(my_plan)
# Compute metadata on 'small', including a hash/fingerprint
# of the dependencies.
meta_list <- list(
  small = drake:::meta(target = "small", config = config)
)
# Should not yet include 'small'.
cached()
# Build 'small'
drake_build(target = "small", meta_list = meta_list, config = config)
# Should now include 'small'
cached()
readd(small)
}
}
