% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{drake_cache_log_file}
\alias{drake_cache_log_file}
\title{Generate a flat text log file to represent the state of
the cache.
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}}
\usage{
drake_cache_log_file(
  file = "drake_cache.log",
  path = getwd(),
  search = TRUE,
  cache = drake::get_cache(path = path, search = search, verbose = verbose),
  verbose = 1L,
  jobs = 1L,
  targets_only = FALSE
)
}
\arguments{
\item{file}{character scalar, name of the flat text log file.}

\item{path}{Path to a \code{drake} cache
(usually a hidden \verb{.drake/} folder) or \code{NULL}.}

\item{search}{Deprecated.}

\item{cache}{drake cache. See \code{\link[=new_cache]{new_cache()}}.
If supplied, \code{path} is ignored.}

\item{verbose}{Deprecated on 2019-09-11.}

\item{jobs}{Number of jobs/workers for parallel processing.}

\item{targets_only}{Logical, whether to output information only on the
targets in your workflow plan data frame. If \code{targets_only} is \code{FALSE}, the
output will include the hashes of both targets and imports.}
}
\value{
There is no return value, but a log file is generated.
}
\description{
Deprecated on 2019-03-09.
}
\details{
Calling this function to create a log file and later calling
\code{make()} makes the log file out of date. Therefore, we recommend using
\code{make()} with the \code{cache_log_file} argument to create the cache log. This
way ensures that the log is always up to date with \code{make()} results.
}
\seealso{
\code{\link[=drake_cache_log]{drake_cache_log()}}, \code{\link[=make]{make()}}, \code{\link[=get_cache]{get_cache()}}
}
\keyword{internal}
