% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/examples.R
\name{drake_examples}
\alias{drake_examples}
\title{List the names of all the drake examples.}
\usage{
drake_examples()
}
\value{
Names of all the drake examples.
}
\description{
The \code{'mtcars'} example is
documented at https://ropenscilabs.github.io/drake-manual/mtcars.html.
}
\examples{
\dontrun{
test_with_dir("Quarantine side effects.", {
drake_examples() # List all the drake examples.
# Sets up the same example as https://ropenscilabs.github.io/drake-manual/mtcars.html # nolint
drake_example("mtcars")
# Sets up the SLURM example.
drake_example("slurm")
})
}
}
\seealso{
\code{\link[=drake_example]{drake_example()}}, \code{\link[=make]{make()}}
}
