% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/examples.R
\name{drake_examples}
\alias{drake_examples}
\title{Function \code{drake_examples}}
\usage{
drake_examples()
}
\value{
Names of all the drake examples.
}
\description{
List the names of all the drake examples.
The \code{'basic'} example is the one from the
quickstart vignette: \code{vignette('quickstart')}.
}
\examples{
\dontrun{
drake_examples() # List all the drake examples.
# Sets up the same example as the quickstart vignette.
drake_example("basic")
# Sets up the SLURM example.
drake_example("slurm")
}
}
\seealso{
\code{\link{drake_example}}, \code{\link{make}}
}
