% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean.R
\name{drake_gc}
\alias{drake_gc}
\title{Function drake_gc()}
\usage{
drake_gc(path = getwd(), search = TRUE, verbose = TRUE, force = FALSE)
}
\arguments{
\item{path}{file path to the folder containing the cache.
Yes, this is the parent directory containing the cache,
not the cache itself, and it assumes the cache is in the
`.drake` folder. If you are looking for a different cache
with a known folder different from `.drake`, use
the \code{\link{this_cache}()} function.}

\item{search}{logical, whether to search back in the file system
for the cache.}

\item{verbose}{logical, whether to print the location of the cache}

\item{force}{logical, whether to load the cache
despite any back compatibility issues with the
running version of drake.}
}
\value{
\code{NULL}
}
\description{
Do garbage collection on the cache
}
\examples{
\dontrun{
load_basic_example() # Load drake's canonical example.
make(my_plan) # Run the project, build the targets.
# At this point, check the size of the '.drake/' cache folder.
# Clean without garbage collection.
clean(garbage_collection = FALSE)
# The '.drake/' cache folder is still about the same size.
drake_gc() # Do garbage collection on the cache.
# The '.drake/' cache folder should have gotten much smaller.
}
}
\seealso{
\code{\link{clean}}
}
