% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api-history.R
\name{drake_history}
\alias{drake_history}
\title{Show history}
\usage{
drake_history(config, show_args = FALSE, show_status = FALSE,
  make_imports = TRUE, do_prework = TRUE)
}
\arguments{
\item{config}{A list returned from \code{drake_config()}.}

\item{show_args}{Logical, whether to include atomic arguments
to function calls in \code{\link[=drake_plan]{drake_plan()}} commands.
Could be slow because this requires parsing and analyzing
lots of R code.}

\item{show_status}{Logical, whether to show the status of each target
based on the results of \code{outdated()}. Could be slow.}

\item{make_imports}{Logical, whether to make the imports first.
Set to \code{FALSE} to save some time and risk obsolete output.}

\item{do_prework}{Whether to do the \code{prework}
normally supplied to \code{\link[=make]{make()}}.}
}
\value{
A data frame of target history.
}
\description{
See the history of your targets:
what you ran, when you ran it, and where
it lives. Optionally show the settings
under which you ran things (could be slow)
and whether each target is up to date (also could be slow).
}
