% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parallel_ui.R
\name{drake_hpc_template_files}
\alias{drake_hpc_template_files}
\title{List the available example template files for deploying
work to a cluster / job scheduler.}
\usage{
drake_hpc_template_files()
}
\value{
a character vector of example tempalte files that
you can write with \code{\link[=drake_hpc_template_file]{drake_hpc_template_file()}}.
}
\description{
See the example files from
\code{\link[=drake_examples]{drake_examples()}} and \code{\link[=drake_example]{drake_example()}}
for example usage.
}
\examples{
\dontrun{
test_with_dir("Quarantine side effects.", {
load_mtcars_example() # Get the code with drake_example("mtcars").
# List the available template files.
drake_hpc_template_files()
# Write a SLURM template file from the SLURM example.
drake_hpc_template_file("slurm_future.tmpl") # Writes slurm_future.tmpl.
library(future.batchtools)
# future::plan(batchtools_slurm, template = "slurm_future.tmpl") # nolint
# make(my_plan, parallelism = "future", jobs = 2) # nolint
})
}
}
\seealso{
\code{\link[=drake_hpc_template_file]{drake_hpc_template_file()}},
\code{\link[=drake_examples]{drake_examples()}}, \code{\link[=drake_example]{drake_example()}},
\code{\link[=shell_file]{shell_file()}}
}
