% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aesthetics.R, R/color.R
\name{drake_palette}
\alias{drake_palette}
\alias{drake_palette}
\title{Function palette}
\usage{
drake_palette()

drake_palette()
}
\description{
show color palette for drake.
Your console must have the crayon package enabled.

show color palette for drake.
Your console must have the crayon package enabled.
}
\examples{
drake_palette()
drake_palette()
}
