% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color.R
\name{drake_palette}
\alias{drake_palette}
\title{Function palette}
\usage{
drake_palette()
}
\value{
There is a console message,
but the actual return value is \code{NULL}.
}
\description{
show color palette for drake.
Used in both the console and \code{\link{vis_drake_graph}()}
Your console must have the crayon package enabled.
}
\details{
This palette applies to console output
(internal functions \code{console()} and
\code{console_many_targets()}) and the node colors
in \code{\link{vis_drake_graph}()}.
So if you want to contribute improvements to the palette,
please both \code{drake_palette()} and
\code{visNetwork::visNetwork(nodes = \link{legend_nodes}())}
}
\examples{
# Show drake's color palette as text.
drake_palette()
# Show part of the palette as an interactive visNetwork graph.
# These are the nodes in the legend of vis_drake_graph().
\dontrun{
visNetwork::visNetwork(nodes = legend_nodes())
}
}
