% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strings.R
\name{drake_quotes}
\alias{drake_quotes}
\title{Function \code{drake_quotes}}
\usage{
drake_quotes(x = NULL, single = FALSE)
}
\arguments{
\item{x}{character vector or object to be coerced to character.}

\item{single}{Add single quotes if \code{TRUE}
and double quotes otherwise.}
}
\value{
character vector with quotes around it
}
\description{
Put quotes around each element of a character vector.
}
\examples{
# Single-quote this string.
drake_quotes("abcd", single = TRUE) # "'abcd'"
# Double-quote this string.
drake_quotes("abcd") # "\\"abcd\\""
}
\seealso{
\code{\link{drake_unquote}}, \code{\link{drake_strings}}
}
