% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecate.R
\name{drake_quotes}
\alias{drake_quotes}
\title{Put quotes around each element of a character vector.}
\usage{
drake_quotes(x = NULL, single = FALSE)
}
\arguments{
\item{x}{character vector or object to be coerced to character.}

\item{single}{Add single quotes if \code{TRUE}
and double quotes otherwise.}
}
\value{
Character vector with quotes around it.
}
\description{
Deprecated on 2019-01-01
}
\examples{
# deprecated
}
\seealso{
\code{\link[=drake_unquote]{drake_unquote()}}, \code{\link[=drake_strings]{drake_strings()}}
}
