% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/progress.R
\name{drake_session}
\alias{drake_session}
\title{Function \code{drake_session}}
\usage{
drake_session(path = getwd(), search = TRUE, cache = drake::get_cache(path
  = path, search = search, verbose = verbose), verbose = TRUE)
}
\arguments{
\item{path}{Root directory of the drake project,
or if \code{search} is \code{TRUE}, either the
project root or a subdirectory of the project.}

\item{search}{If \code{TRUE}, search parent directories
to find the nearest drake cache. Otherwise, look in the
current working directory only.}

\item{cache}{optional drake cache. See code{\link{new_cache}()}.
If \code{cache} is supplied,
the \code{path} and \code{search} arguments are ignored.}

\item{verbose}{whether to print console messages}
}
\value{
\code{\link{sessionInfo}()} of the last
call to \code{\link{make}()}
}
\description{
Load the \code{\link{sessionInfo}()}
of the last call to \code{\link{make}()}.
}
\examples{
\dontrun{
load_basic_example() # Load drake's canonical example.
make(my_plan) # Run the project, build the targets.
drake_session() # Retrieve the cached sessionInfo() of the last make().
}
}
\seealso{
\code{\link{diagnose}}, \code{\link{built}}, \code{\link{imported}},
\code{\link{readd}}, \code{\link{workplan}}, \code{\link{make}}
}
