% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecate.R
\name{drake_session}
\alias{drake_session}
\title{Deprecated. Return the \code{\link[=sessionInfo]{sessionInfo()}}
of the last call to \code{\link[=make]{make()}}.}
\usage{
drake_session(path = getwd(), search = TRUE,
  cache = drake::get_cache(path = path, search = search, verbose =
  verbose), verbose = 1L)
}
\arguments{
\item{path}{Root directory of the drake project,
or if \code{search} is \code{TRUE}, either the
project root or a subdirectory of the project.
Ignored if a \code{cache} is supplied.}

\item{search}{logical. If \code{TRUE}, search parent directories
to find the nearest drake cache. Otherwise, look in the
current working directory only.
Ignored if a \code{cache} is supplied.}

\item{cache}{drake cache. See \code{\link[=new_cache]{new_cache()}}.
If supplied, \code{path} and \code{search} are ignored.}

\item{verbose}{logical or numeric, control printing to the console.
\itemize{
\item \code{0} or \code{FALSE}: print nothing.
\item \code{1} or \code{TRUE}: print only targets to build.
\item \code{2}: also print checks and cache info.
\item \code{3}: also print any potentially missing items.
\item \code{4}: also print imports and writes to the cache.
}}
}
\value{
\code{\link[=sessionInfo]{sessionInfo()}} of the last call to \code{\link[=make]{make()}}
}
\description{
Deprecated. Use \code{\link[=drake_get_session_info]{drake_get_session_info()}} instead.
}
\details{
Deprecated on 2018-12-06.
}
\examples{
# See ?drake_get_session_info for examples.
}
\seealso{
\code{\link[=diagnose]{diagnose()}}, \code{\link[=built]{built()}}, \code{\link[=imported]{imported()}},
\code{\link[=readd]{readd()}}, \code{\link[=drake_plan]{drake_plan()}}, \code{\link[=make]{make()}}
}
