% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{drake_strings}
\alias{drake_strings}
\title{Turn valid expressions into character strings.
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}}
\usage{
drake_strings(...)
}
\arguments{
\item{...}{Unquoted symbols to turn into character strings.}
}
\value{
A character vector.
}
\description{
Deprecated on 2019-01-01
}
\keyword{internal}
