% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-deprecate.R
\name{drake_strings}
\alias{drake_strings}
\title{Turn valid expressions into character strings.}
\usage{
drake_strings(...)
}
\arguments{
\item{...}{Unquoted symbols to turn into character strings.}
}
\value{
A character vector.
}
\description{
Deprecated on 2019-01-01
}
\examples{
# deprecated
}
\keyword{internal}
