% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecate.R
\name{drake_strings}
\alias{drake_strings}
\title{Turn valid expressions into character strings.}
\usage{
drake_strings(...)
}
\arguments{
\item{...}{unquoted symbols to turn into character strings.}
}
\value{
A character vector.
}
\description{
Deprecated on 2019-01-01
}
\examples{
# deprecated
}
\seealso{
\code{\link[=drake_quotes]{drake_quotes()}}, \code{\link[=drake_unquote]{drake_unquote()}}
}
