% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strings.R
\name{drake_strings}
\alias{drake_strings}
\title{Turn valid expressions into character strings.}
\usage{
drake_strings(...)
}
\arguments{
\item{...}{unquoted symbols to turn into character strings.}
}
\value{
a character vector
}
\description{
This function may be useful for
constructing workflow plan data frames.
}
\examples{
# Turn symbols into strings.
drake_strings(a, b, c, d) # [1] "a" "b" "c" "d"
}
\seealso{
\code{\link[=drake_quotes]{drake_quotes()}}, \code{\link[=drake_unquote]{drake_unquote()}}
}
