% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tips.R
\name{drake_tip}
\alias{drake_tip}
\title{Function drake_tip}
\usage{
drake_tip()
}
\description{
Output a random tip about drake.
Tips are usually related to news and usage.
}
\examples{
drake_tip()
message(drake_tip())
}
