% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-tips.R
\name{drake_tip}
\alias{drake_tip}
\title{Output a random tip about drake.}
\usage{
drake_tip()
}
\value{
A character scalar with a tip on how to use drake.
}
\description{
Tips are usually related to news and usage.
}
\examples{
drake_tip() # Show a tip about using drake.
message(drake_tip()) # Print out a tip as a message.
}
