% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strings.R
\name{drake_unquote}
\alias{drake_unquote}
\title{Function \code{drake_unquote}}
\usage{
drake_unquote(x = NULL, deep = FALSE)
}
\arguments{
\item{x}{character vector}

\item{deep}{remove all outer quotes if \code{TRUE}
and only the outermost set otherwise. Single and double
quotes are treated interchangeably, and matching is not checked.}
}
\value{
character vector without leading
or trailing escaped quotes around
the elements
}
\description{
Remove leading and trailing
escaped quotes from character strings.
}
\examples{
x <- "'abcd'"
# Remove the literal quotes around x.
drake_unquote(x) # "abcd"
}
\seealso{
\code{\link{drake_quotes}}, \code{\link{drake_strings}}
}
