% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-deprecate.R
\name{drake_unquote}
\alias{drake_unquote}
\title{Remove leading and trailing
escaped quotes from character strings.}
\usage{
drake_unquote(x = NULL)
}
\arguments{
\item{x}{Character vector.}
}
\value{
Character vector without leading
or trailing escaped quotes around
the elements.
}
\description{
Deprecated on 2019-01-01
}
\examples{
# deprecated
}
\keyword{internal}
