% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{eager_load_target}
\alias{eager_load_target}
\title{Load a target right away (internal function)}
\usage{
eager_load_target(target, cache, namespace, envir, verbose)
}
\arguments{
\item{cache}{optional drake cache. See code{\link{new_cache}()}.
If \code{cache} is supplied,
the \code{path} and \code{search} arguments are ignored.}

\item{namespace}{character scalar,
name of an optional storr namespace to load from.}

\item{envir}{environment to load objects into. Defaults to the
calling environment (current workspace).}

\item{verbose}{logical, whether to print console messages}
}
\description{
This function is only exported
to make active bindings work safely.
It is not actually a user-side function.
}
\keyword{internal}
