% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecate.R
\name{evaluate}
\alias{evaluate}
\title{Deprecated function evaluate}
\usage{
evaluate(plan, rules = NULL, wildcard = NULL, values = NULL,
  expand = TRUE)
}
\arguments{
\item{plan}{Same as for \code{\link{evaluate_plan}}}

\item{rules}{Same as for \code{\link{evaluate_plan}}}

\item{wildcard}{Same as for \code{\link{evaluate_plan}}}

\item{values}{Same as for \code{\link{evaluate_plan}}}

\item{expand}{Same as for \code{\link{evaluate_plan}}}
}
\value{
Same as for \code{\link{evaluate_plan}}
}
\description{
Use \code{\link{evaluate_plan}()} instead.
}
\examples{
# See ?evaluate_plan for examples.
}
\seealso{
\code{\link{evaluate_plan}}
}
