% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/defunct.R
\name{example_drake}
\alias{example_drake}
\title{example_drake \ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#defunct}{\figure{lifecycle-defunct.svg}{options: alt='[Defunct]'}}}{\strong{[Defunct]}}}
\usage{
example_drake(...)
}
\arguments{
\item{...}{Arguments}
}
\description{
2019-02-15
}
\keyword{internal}
