% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecate.R
\name{example_drake}
\alias{example_drake}
\title{Deprecated function \code{example_drake}}
\usage{
example_drake(example = drake::drake_examples(), destination = getwd())
}
\arguments{
\item{example}{Same as for \code{\link[=drake_example]{drake_example()}}}

\item{destination}{Same as for \code{\link[=drake_example]{drake_example()}}}
}
\value{
\code{NULL}
}
\description{
Use \code{\link[=drake_example]{drake_example()}} instead.
}
\details{
Deprecated on 2017-11-12.
}
\examples{
# See ?drake_example for examples.
}
\seealso{
\code{\link[=drake_example]{drake_example()}}
}
\keyword{internal}
