% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/examples.R
\name{example_drake}
\alias{example_drake}
\title{Function \code{example_drake}}
\usage{
example_drake(example = drake::examples_drake(), destination = getwd())
}
\arguments{
\item{example}{name of the example. 
To see all the available example names, 
run \code{\link{examples_drake}}.}

\item{destination}{character scalar, file path, where
to write the folder containing the code files for the example.}
}
\description{
Copy a folder of code files for a 
drake example to the current working directory.
Call \code{example_drake("basic")} to generate the code files from the
quickstart vignette: \code{vignette("quickstart")}.
To see the names of all the examples, run \code{\link{examples_drake}}.
}
\examples{
\dontrun{
example_drake("basic") # Walkthrough: vignette("quickstart")
}
}
\seealso{
\code{\link{examples_drake}}, \code{\link{make}}
}
