% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/examples.R
\name{examples_drake}
\alias{examples_drake}
\title{Function \code{examples_drake}}
\usage{
examples_drake()
}
\value{
Names of all the drake examples.
}
\description{
List the names of all the drake examples.
The \code{'basic'} example is the one from the
quickstart vignette: \code{vignette('quickstart')}.
}
\examples{
\dontrun{
examples_drake() # List all the drake examples.
# Sets up the same example as the quickstart vignette.
example_drake("basic")
# Sets up the SLURM example.
example_drake("slurm")
}
}
\seealso{
\code{\link{example_drake}}, \code{\link{make}}
}
