% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecate.R
\name{examples_drake}
\alias{examples_drake}
\title{Deprecated function \code{examples_drake}}
\usage{
examples_drake()
}
\value{
Names of all the drake examples.
}
\description{
Use \code{\link{drake_examples}()} instead.
}
\details{
Deprecated on 2017-11-12.
}
\examples{
# See ?drake_examples for the examples.
}
\seealso{
\code{\link{drake_examples}}
}
\keyword{internal}
