% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate.R
\name{expand}
\alias{expand}
\title{Function \code{expand}}
\usage{
expand(plan, values = NULL)
}
\arguments{
\item{plan}{workflow plan data frame}

\item{values}{values to expand over. These will be appended to
the names of the new targets.}
}
\value{
an expanded workflow plan data frame
}
\description{
Expands a workflow plan data frame by duplicating rows.
This generates multiple replicates of targets with the same commands.
}
\examples{
datasets <- workflow(
  small = simulate(5),
  large = simulate(50))
expand(datasets, values = c("rep1", "rep2", "rep3"))
}
