% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate.R
\name{expand}
\alias{expand}
\title{Function \command{expand}}
\usage{
expand(x, values = NULL)
}
\arguments{
\item{x}{argument data frame}

\item{values}{values to expand over}
}
\value{
an expanded data frame
}
\description{
Expands a dataframe of drake instructions by duplicating rows.
}

