% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expose_imports.R
\name{expose_imports}
\alias{expose_imports}
\title{Expose all the imports in a package so \code{make()} can detect
all the package's nested functions.}
\usage{
expose_imports(package, character_only = FALSE, envir = parent.frame(),
  jobs = 1)
}
\description{
When \code{drake} analyzes the functions in your environment,
it can
}
