% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{failed}
\alias{failed}
\title{List failed targets.
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}}
\usage{
failed(
  path = NULL,
  search = NULL,
  cache = drake::drake_cache(path = path),
  verbose = 1L,
  upstream_only = NULL
)
}
\arguments{
\item{path}{Path to a \code{drake} cache
(usually a hidden \verb{.drake/} folder) or \code{NULL}.}

\item{search}{Deprecated.}

\item{cache}{drake cache. See \code{\link[=new_cache]{new_cache()}}.
If supplied, \code{path} is ignored.}

\item{verbose}{Deprecated on 2019-09-11.}

\item{upstream_only}{Deprecated.}
}
\value{
A character vector of target names.
}
\description{
Deprecated on 2020-03-23. Use \code{\link[=drake_failed]{drake_failed()}} instead.
}
\keyword{internal}
