% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dependencies.R
\name{file_output}
\alias{file_output}
\title{Declare the file outputs of a workflow plan command.}
\usage{
file_output(...)
}
\arguments{
\item{...}{Symbols or character vectors denoting file outputs.
file_output(summaries.txt)
file_output(plot.pdf, 'report.ps', "report.html")}
}
\value{
A character vector of declared files.
}
\description{
Use this function to help write the commands
in your workflow plan data frame.
}
\seealso{
file_input knitr_input
}
