% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/defunct.R
\name{find_knitr_doc}
\alias{find_knitr_doc}
\title{find_knitr_doc \ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#defunct}{\figure{lifecycle-defunct.svg}{options: alt='[Defunct]'}}}{\strong{[Defunct]}}}
\usage{
find_knitr_doc(...)
}
\arguments{
\item{...}{Arguments}
}
\description{
2019-02-15
}
\keyword{internal}
