% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find.R
\name{find_project}
\alias{find_project}
\title{Function \code{find_project}}
\usage{
find_project(path = getwd())
}
\arguments{
\item{path}{starting path for search back for the project.
Should be a subdirectory of the drake project.}
}
\value{
File path of the nearest drake project or \code{NULL}
if no drake project is found.
}
\description{
Return the file path of the nearest drake
project (searching upwards for directories
containing a drake cache).
Only works if the cache is a file system in a folder named \code{.drake}.
}
\examples{
\dontrun{
load_basic_example()
make(my_plan)
find_project()
}
}
\seealso{
\code{\link{workplan}}, \code{\link{make}}
}
