% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find.R
\name{find_project}
\alias{find_project}
\title{Search up the file system
for the nearest root path of a drake project.}
\usage{
find_project(path = getwd())
}
\arguments{
\item{path}{starting path for search back for the project.
Should be a subdirectory of the drake project.}
}
\value{
File path of the nearest drake project or \code{NULL}
if no drake project is found.
}
\description{
Only works if the cache is a file system
in a folder named \code{.drake} (default).
}
\examples{
\dontrun{
test_with_dir("Quarantine side effects.", {
load_basic_example() # Get the code with drake_example("basic").
make(my_plan) # Run the project, build the target.
# Find the root directory of the current drake project.
# Search up through parent directories if necessary.
find_cache()
})
}
}
\seealso{
\code{\link{drake_plan}}, \code{\link{make}}
}
