% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/defunct.R
\name{from_plan}
\alias{from_plan}
\title{from_plan \ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#defunct}{\figure{lifecycle-defunct.svg}{options: alt='[Defunct]'}}}{\strong{[Defunct]}}}
\usage{
from_plan(column)
}
\arguments{
\item{column}{Character, name of a column in your \code{drake} plan.}
}
\description{
The \code{from_plan()} function is now defunct
in order to reduce the demands on memory usage.
}
\details{
2019-03-28
}
\seealso{
\code{\link[=drake_envir]{drake_envir()}}
}
\keyword{internal}
