% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exec-memory.R
\name{from_plan}
\alias{from_plan}
\title{Get a target's plan info from inside a plan's command.}
\usage{
from_plan(column)
}
\arguments{
\item{column}{Character, name of a column in your \code{drake} plan.}
}
\value{
\code{plan[target, column]}, where \code{plan} is your workflow
plan data frame, \code{target} is the target being built,
and \code{column} is the name of the column of the plan you provide.
}
\description{
Call this function inside the commands in your plan
to get an entry from any column in the plan. Changes to custom
columns referred to this way
(for example, \code{a} in \code{drake_plan(x = target(from_plan("a"), a = 123))})
do not invalidate targets, so be careful. Only use \code{from_plan()}
to reference data that does not actually affect the output value
of the target. For example, you might use \code{from_plan()} to set the
number of parallel workers within a target:
\code{drake_plan(x = target(mclapply(..., from_plan("cores"), cores = 4))}.
Now, if you change the \code{cores} column of the plan, the parallelism will
change, but the target \code{x} will stay up to date.
}
\examples{
plan <- drake_plan(my_target = target(from_plan("a"), a = "a_value"))
plan

cache <- storr::storr_environment()
make(plan, cache = cache, session_info = FALSE)
readd(my_target, cache = cache)

# Why do we care?
# Because we can customize parallel computing within targets this way.
# Notice how we treat `mc.cores` for `mclapply()`
plan <- drake_plan(
  a = target(
    mclapply(1:8, my_function, mc.cores = from_plan("cores")),
    cores = 4
  ),
  b = target(
    mclapply(1:4, my_function, mc.cores = from_plan("cores")),
    cores = 2
  )
)

plan

# make() # nolint

# Now, if you change the `cores` column of the plan, the parallelism will
# change, but the targets will stay up to date.
plan$cores <- c(1, 1)
plan

# make(plan) # nolint
}
\seealso{
\code{\link[=drake_envir]{drake_envir()}}
}
