% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backend-future.R
\name{future_build}
\alias{future_build}
\title{Task passed to individual futures in the \code{"future"} backend}
\usage{
future_build(target, meta, config, protect)
}
\arguments{
\item{target}{name of the target}

\item{meta}{list of metadata}

\item{config}{\code{\link[=drake_config]{drake_config()}} list}

\item{protect}{Names of targets that still need their
dependencies available in memory.}
}
\value{
Either the target value or a list of build results.
}
\description{
For internal use only. Only exported to make available
to futures.
}
\keyword{internal}
