% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate.R
\name{gather}
\alias{gather}
\title{Function \code{gather}}
\usage{
gather(x = NULL, target = "target", gather = "list")
}
\arguments{
\item{x}{argument data frame}

\item{target}{name of aggregated target object}

\item{gather}{function used to gather the targets}
}
\value{
data frame with a command to gather the targets in \code{x}
}
\description{
Aggregate/gather the targets of a previous set of drake instructions.
}

