% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate.R
\name{gather}
\alias{gather}
\title{Function \code{gather}}
\usage{
gather(plan = NULL, target = "target", gather = "list")
}
\arguments{
\item{plan}{workflow plan data frame of prespecified targets}

\item{target}{name of the new aggregated target}

\item{gather}{function used to gather the targets. Should be 
one of \code{\link{list}(...)}, \code{\link{c}(...)},
\code{\link{rbind}(...)}, or similar.}
}
\value{
workflow plan data frame for aggregating prespecified targets
}
\description{
Create a new workflow plan data frame with a single new 
target. This new target is a list, vector, or other aggregate of
a collection of existing targets in another workflow plan data frame.
}
\examples{
datasets = plan(
  small = simulate(5),
  large = simulate(50))
gather(datasets, target = "my_datasets")
gather(datasets, target = "aggregated_data", gather = "rbind")
}
