% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecate.R
\name{gather}
\alias{gather}
\title{Deprecated function \code{gather}}
\usage{
gather(plan = NULL, target = "target", gather = "list")
}
\arguments{
\item{plan}{Same as for \code{\link{gather_plan}}}

\item{target}{Same as for \code{\link{gather_plan}}}

\item{gather}{Same as for \code{\link{gather_plan}}}
}
\value{
Same as for \code{\link{gather_plan}}
}
\description{
Use \code{\link{gather_plan}()} instead.
}
\details{
Deprecated on 2017-11-12.
}
\examples{
# See ?gather_plan for examples.
}
\seealso{
\code{\link{gather_plan}}
}
\keyword{internal}
