% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache.R
\name{get_cache}
\alias{get_cache}
\title{Function get_cache}
\usage{
get_cache(path = getwd(), search = TRUE)
}
\arguments{
\item{path}{path to the cache}

\item{search}{logical, whether to search back in the file system
for the cache.}
}
\description{
get the drake cache
}
\examples{
\dontrun{
get_cache()
load_basic_example()
make(my_plan)
get_cache()
}
}
\seealso{
\code{\link{config}}
}
