% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache.R
\name{get_cache}
\alias{get_cache}
\title{Function get_cache}
\usage{
get_cache(path = getwd(), search = TRUE, verbose = TRUE)
}
\arguments{
\item{path}{file path to the folder containing the cache.
Yes, this is the parent directory containing the cache,
not the cache itself, and it assumes the cache is in the
`.drake` folder. If you are looking for a different cache
with a known folder different from `.drake`, use
the \code{\link{this_cache}()} function.}

\item{search}{logical, whether to search back in the file system
for the cache.}

\item{verbose}{logical, whether to print the location of the cache}
}
\description{
Search for and return a drake file system cache.
}
\examples{
\dontrun{
get_cache()
load_basic_example()
make(my_plan)
x <- get_cache()
x$list(namespace = "readd")
}
}
\seealso{
\code{\link{this_cache}}, \code{\link{new_cache}},
\code{\link{recover_cache}}, \code{\link{config}}
}
