% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache.R
\name{get_cache}
\alias{get_cache}
\title{Function \code{get_cache}}
\usage{
get_cache(path = getwd(), search = TRUE, verbose = 1, force = FALSE,
  fetch_cache = NULL)
}
\arguments{
\item{path}{file path to the folder containing the cache.
Yes, this is the parent directory containing the cache,
not the cache itself, and it assumes the cache is in the
`.drake` folder. If you are looking for a different cache
with a known folder different from `.drake`, use
the \code{\link{this_cache}()} function.}

\item{search}{logical, whether to search back in the file system
for the cache.}

\item{verbose}{logical, whether to print the location of the cache}

\item{force}{logical, whether to load the cache
despite any back compatibility issues with the
running version of drake.}
}
\value{
A drake/storr cache, if available. \code{NULL} otherwise.
}
\description{
Search for and return a drake file system cache.
}
\examples{
\dontrun{
test_with_dir("Quarantine side effects.", {
clean(destroy = TRUE)
# No cache is available.
get_cache() # NULL
load_basic_example() # Load drake's canonical example.
make(my_plan) # Run the project, build the targets.
x <- get_cache() # Now, there is a cache.
# List the objects readable from the cache with readd().
x$list() # Or x$list(namespace = x$default_namespace)
})
}
}
\seealso{
\code{\link{this_cache}}, \code{\link{new_cache}},
\code{\link{recover_cache}}, \code{\link{config}}
}
