% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{get_trace}
\alias{get_trace}
\title{Deprecated, get a trace of a dynamic target's value.
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}}
\usage{
get_trace(trace, value)
}
\arguments{
\item{trace}{Character, name of the trace
you want to extract. Such trace names are declared
in the \code{.trace} argument of \code{map()}, \code{cross()} or \code{group()}..}

\item{value}{Value of the dynamic target}
}
\value{
The dynamic trace of one target in another:
a vector of values from a grouping variable.
}
\description{
Deprecated on 2019-12-10. Use \code{\link[=read_trace]{read_trace()}} instead.
}
\keyword{internal}
