% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workplan.R
\name{ignore}
\alias{ignore}
\title{Ignore components of commands and imported functions.}
\usage{
ignore(x)
}
\arguments{
\item{x}{code to ignore}
}
\value{
the argument
}
\description{
In commands and imported functions, use
\code{ignore(some_code)} to
\enumerate{
\item Force \code{drake} to not track dependencies in \code{some_code}, and
\item Ignore any changes in \code{some_code} when it comes to
deciding which target are out of date.
}
}
\seealso{
\code{\link[=file_in]{file_in()}}, \code{\link[=file_out]{file_out()}}, \code{\link[=knitr_in]{knitr_in()}}
}
