% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache.R
\name{imported}
\alias{imported}
\title{Function \command{imported}}
\usage{
imported(path = getwd(), search = FALSE)
}
\arguments{
\item{path}{Root directory of the drake project,
or if \command{search} is \command{TRUE}, either the
project root or a subdirectory of the project.}

\item{search}{If \command{TRUE}, search parent directories
to find the nearest drake cache. Otherwise, look in the
current working directory only.}
}
\value{
list of imported objects in the cache
}
\description{
List all the imported objects in the drake cache
}
\seealso{
\command{\link{cached}}, \command{\link{loadd}},
\command{link{built}}
}

