% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-deprecate.R
\name{imported}
\alias{imported}
\title{Deprecated. List all the imports in the drake cache.}
\usage{
imported(files_only = FALSE, path = getwd(), search = TRUE,
  cache = drake::get_cache(path = path, search = search, verbose =
  verbose), verbose = 1L, jobs = 1)
}
\arguments{
\item{files_only}{logical, whether to show imported files only
and ignore imported objects. Since all your functions and
all their global variables are imported, the full list of
imported objects could get really cumbersome.}

\item{path}{Root directory of the drake project,
or if \code{search} is \code{TRUE}, either the
project root or a subdirectory of the project.
Ignored if a \code{cache} is supplied.}

\item{search}{logical. If \code{TRUE}, search parent directories
to find the nearest drake cache. Otherwise, look in the
current working directory only.
Ignored if a \code{cache} is supplied.}

\item{cache}{drake cache. See \code{\link[=new_cache]{new_cache()}}.
If supplied, \code{path} and \code{search} are ignored.}

\item{verbose}{logical or numeric, control printing to the console.
\itemize{
\item \code{0} or \code{FALSE}: print nothing.
\item \code{1} or \code{TRUE}: print only targets to build.
\item \code{2}: plus checks and cache info.
\item \code{3}: plus missing imports.
\item \code{4}: plus all imports.
\item \code{5}: plus execution and total build times for targets.
\item \code{6}: plus notifications when targets are being stored.
}}

\item{jobs}{number of jobs/workers for parallel processing}
}
\value{
Character vector naming the imports in the cache.
}
\description{
Deprecated on 2019-01-08.
}
\details{
An import is a non-target object processed
by \code{\link[=make]{make()}}. Targets in the workflow
plan data frame (see \code{\link[=drake_config]{drake_config()}}
may depend on imports.
}
\examples{
\dontrun{
test_with_dir("Quarantine side effects.", {
load_mtcars_example() # Load the canonical example.
make(my_plan) # Run the project, build the targets.
imported() # List all the imported objects/files in the cache.
# For imported files, only the fingerprints/hashes are stored.
})
}
}
\seealso{
\code{\link[=cached]{cached()}}, \code{\link[=loadd]{loadd()}}
}
