% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache_types.R
\name{in_memory_cache_types}
\alias{in_memory_cache_types}
\title{Function in_memory_cache_types}
\usage{
in_memory_cache_types()
}
\description{
List the supported drake in-memory cache types.
}
\examples{
cache_types()
}
\seealso{
\code{\link{cache_types}},
\code{\link{new_cache}}, \code{\link{get_cache}},
\code{\link{default_cache_type}}
}
