% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{isolate_example}
\alias{isolate_example}
\title{Isolate the side effects of an example.
\ifelse{html}{\out{<a href='https://www.tidyverse.org/lifecycle/#stable'><img src='figures/lifecycle-stable.svg' alt='Stable lifecycle'></a>}}{\strong{Stable}}}
\usage{
isolate_example(desc, code)
}
\arguments{
\item{desc}{Character, description of the example.}

\item{...}{Code to run.}
}
\value{
Nothing.
}
\description{
Runs code in a temporary directory
in a controlled environment with a controlled
set of options.
}
\keyword{internal}
