% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{isolate_example}
\alias{isolate_example}
\title{Isolate the side effects of an example.
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}}
\usage{
isolate_example(desc, code)
}
\arguments{
\item{desc}{Character, description of the example.}

\item{code}{Code to run.}
}
\value{
Nothing.
}
\description{
Runs code in a temporary directory
in a controlled environment with a controlled
set of options.
}
\keyword{internal}
