% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-deprecate.R
\name{knitr_deps}
\alias{knitr_deps}
\title{Deprecated in favor of \code{\link[=deps_knitr]{deps_knitr()}}}
\usage{
knitr_deps(target)
}
\arguments{
\item{target}{Encoded file path}
}
\value{
Data frame of dependencies
}
\description{
Deprecated on 2019-02-14
\code{knit("your_report.Rmd")} or
\code{knit("your_report.Rmd", quiet = TRUE)}.
}
\keyword{internal}
