% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-knitr.R
\name{knitr_deps}
\alias{knitr_deps}
\title{Find the drake dependencies of a dynamic knitr report target.}
\usage{
knitr_deps(target)
}
\arguments{
\item{target}{File path to the file or name of the file target,
source text of the document.}
}
\value{
A character vector of the names of dependencies.
}
\description{
To enable drake to watch for the dependencies
of a knitr report, the command in your workflow plan data frame
must call \code{\link[knitr:knit]{knitr::knit()}} directly.
In other words,
the command must look something like
\code{knit("your_report.Rmd")} or
\code{knit("your_report.Rmd", quiet = TRUE)}.
}
\details{
drake looks for dependencies in the document by
analyzing evaluated code chunks for other targets/imports
mentioned in \code{\link[=loadd]{loadd()}} and \code{\link[=readd]{readd()}}.
}
\examples{
\dontrun{
test_with_dir("Quarantine side effects.", {
load_mtcars_example() # Get the code with drake_example("mtcars").
knitr_deps(file_store("report.Rmd"))
})
}
}
\seealso{
\code{\link[=deps_code]{deps_code()}},
\code{\link[=make]{make()}}, \code{\link[=load_mtcars_example]{load_mtcars_example()}}
}
