% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataframes_graph_utils.R
\name{legend_nodes}
\alias{legend_nodes}
\title{Function legend_nodes}
\usage{
legend_nodes(font_size = 20)
}
\arguments{
\item{font_size}{font size of the node label text}
}
\description{
Output a \code{visNetwork}-friendly
data frame of nodes. It tells you what
the colors and shapes mean
in \code{\link{plot_graph}()}.
}
\examples{
\dontrun{
visNetwork::visNetwork(nodes = legend_nodes())
}
}
\seealso{
\code{\link{drake_palette}()},
\code{\link{plot_graph}()},
\code{\link{dataframes_graph}()}
}
