% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/defunct.R
\name{load_basic_example}
\alias{load_basic_example}
\title{load_basic_example \ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#defunct}{\figure{lifecycle-defunct.svg}{options: alt='[Defunct]'}}}{\strong{[Defunct]}}}
\usage{
load_basic_example(...)
}
\arguments{
\item{...}{Arguments}
}
\description{
2019-02-15
}
\keyword{internal}
