% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/examples.R
\name{load_basic_example}
\alias{load_basic_example}
\title{Function \code{load_basic_example}}
\usage{
load_basic_example(envir = parent.frame())
}
\arguments{
\item{envir}{The environment to load the example into. Defaults to your workspace.
For an insulated workspace, set \code{envir = new.env(parent = globalenv())}.}
}
\description{
Load the basic example into your workspace (or the environment you specify).
Also writes/overwrites the file \code{report.Rmd}.
}
\examples{
\dontrun{
load_basic_example()
deps(reg1)
deps(my_plan$command[1])
deps(my_plan$command[4])
plot_graph(my_plan)
make(my_plan)
clean(destroy = TRUE)
unlink("report.Rmd")
}
}
