% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic_example.R
\name{load_basic_example}
\alias{load_basic_example}
\title{Function \code{load_basic_example}}
\usage{
load_basic_example(envir = parent.frame(), cache = NULL,
  report_file = "report.Rmd", overwrite = FALSE, to = report_file,
  verbose = TRUE, force = FALSE)
}
\arguments{
\item{envir}{The environment to load the example into.
Defaults to your workspace.
For an insulated workspace,
set \code{envir = new.env(parent = globalenv())}.}

\item{cache}{Optional \code{storr} cache to use.}

\item{report_file}{where to write the report file \code{report.Rmd}.}

\item{overwrite}{logical, whether to overwrite an
existing file \code{report.Rmd}}

\item{to}{deprecated, where to write the dynamic report source file
\code{report.Rmd}}

\item{verbose}{logical, whether to print console messages.}

\item{force}{logical, whether to force the loading of a
non-back-compatible cache from a previous version of drake.}
}
\value{
A \code{\link{drake_config}()} configuration list.
}
\description{
Loads the basic example into your workspace
(or the environment you specify).
Also writes/overwrites the file \code{report.Rmd}.
For a thorough walkthrough of how to set up this example, see the
quickstart vignette: \code{vignette('quickstart')}. Alternatively,
call \code{\link{drake_example}('basic')} to generate an R script
that builds up this example step by step.
}
\examples{
\dontrun{
# Populate your workspace and write 'report.Rmd'.
load_basic_example()
# Check the dependencies of an imported function.
deps(reg1)
# Check the dependencies of commands in the workflow plan.
deps(my_plan$command[1])
deps(my_plan$command[4])
# Plot the interactive network visualization of the workflow.
vis_drake_graph(my_plan)
# Run the workflow to build all the targets in the plan.
make(my_plan)
# Remove the whole cache.
clean(destroy = TRUE)
# Clean up the imported file.
unlink('report.Rmd')
}
}
