% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecate.R
\name{load_basic_example}
\alias{load_basic_example}
\title{Deprecated function \code{load_basic_example}}
\usage{
load_basic_example(envir = parent.frame(), seed = NULL, cache = NULL,
  report_file = "report.Rmd", overwrite = FALSE, to = report_file,
  verbose = drake::default_verbose(), force = FALSE)
}
\arguments{
\item{envir}{The environment to load the example into.
Defaults to your workspace.
For an insulated workspace,
set \code{envir = new.env(parent = globalenv())}.}

\item{report_file}{where to write the report file. Deprecated.
In a future release, the report file will always be
\code{report.Rmd} and will always be written to your
working directory (current default).}

\item{overwrite}{logical, whether to overwrite an
existing file \code{report.Rmd}}

\item{force}{logical, whether to force the loading of a
non-back-compatible cache from a previous version of drake.}
}
\value{
A config list, as in \code{\link[=load_mtcars_example]{load_mtcars_example()}}.
}
\description{
Use \code{\link[=load_mtcars_example]{load_mtcars_example()}} instead.
}
\details{
Deprecated on 2018-04-21.
}
\examples{
# See ?load_mtcars_example for examples.
}
\seealso{
\code{\link[=load_mtcars_example]{load_mtcars_example()}}
}
\keyword{internal}
