% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{load_main_example}
\alias{load_main_example}
\title{Load the main example.
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}}
\usage{
load_main_example(
  envir = parent.frame(),
  report_file = "report.Rmd",
  overwrite = FALSE,
  force = FALSE
)
}
\arguments{
\item{envir}{The environment to load the example into.
Defaults to your workspace.
For an insulated workspace,
set \code{envir = new.env(parent = globalenv())}.}

\item{report_file}{Where to write the report file \code{report.Rmd}.}

\item{overwrite}{Logical, whether to overwrite an
existing file \code{report.Rmd}}

\item{force}{Deprecated.}
}
\value{
A \code{\link[=drake_config]{drake_config()}} configuration list.
}
\description{
The main example lives at
\verb{https://github.com/wlandau/drake-examples/tree/main/main}.
Use \code{drake_example("main")} to download its code.
This function also writes/overwrites
the files \code{report.Rmd} and \code{raw_data.xlsx}.
}
\details{
Deprecated 2018-12-31.
}
\keyword{internal}
