% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hash.R
\name{long_hash}
\alias{long_hash}
\title{Get the long hash algorithm of a drake cache.}
\usage{
long_hash(cache = drake::get_cache(verbose = verbose),
  verbose = drake::default_verbose())
}
\arguments{
\item{cache}{drake cache}

\item{verbose}{logical or numeric, control printing to the console.
Use \code{pkgconfig} to set the default value of \code{verbose}
for your R session:
for example, \code{pkgconfig::set_config("drake::verbose" = 2)}.
\describe{
\item{0 or \code{FALSE}:}{print nothing.}
\item{1 or \code{TRUE}:}{print only targets to build.}
\item{2:}{in addition, print checks and cache info.}
\item{3:}{in addition, print any potentially missing items.}
\item{4:}{in addition, print imports. Full verbosity.}
}}
}
\value{
A character vector naming a hash algorithm.
}
\description{
See the advanced storage tutorial
at \url{https://github.com/ropensci/drake/blob/master/vignettes/storage.Rmd}
for details.
}
\examples{
\dontrun{
test_with_dir("Quarantine side effects.", {
load_basic_example() # Get the code with drake_example("basic").
# Run the project and return the internal master configuration list.
config <- make(my_plan)
# Locate the storr cache.
cache <- config$cache
# Get the long hash algorithm of the cache.
long_hash(cache)
})
}
}
\seealso{
\code{\link[=default_short_hash_algo]{default_short_hash_algo()}},
\code{\link[=default_long_hash_algo]{default_long_hash_algo()}}
}
