% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-deprecate.R
\name{long_hash}
\alias{long_hash}
\title{Deprecated. \code{drake} now has just one hash algorithm per cache.}
\usage{
long_hash(cache = drake::get_cache(verbose = verbose), verbose = 1L)
}
\arguments{
\item{cache}{drake cache. See \code{\link[=new_cache]{new_cache()}}.
If supplied, \code{path} and \code{search} are ignored.}

\item{verbose}{logical or numeric, control printing to the console.
\itemize{
\item \code{0} or \code{FALSE}: print nothing.
\item \code{1} or \code{TRUE}: print only targets to build.
\item \code{2}: plus checks and cache info.
\item \code{3}: plus missing imports.
\item \code{4}: plus all imports.
\item \code{5}: plus execution and total build times for targets.
\item \code{6}: plus notifications when targets are being stored.
}}
}
\value{
A character vector naming a hash algorithm.
}
\description{
Deprecated on 2018-12-12
}
\examples{
# deprecated
}
\keyword{internal}
