% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecate.R
\name{make}
\alias{make}
\title{Deprecated function \code{make}. Use \code{link{run}()} instead.}
\usage{
make(...)
}
\arguments{
\item{...}{arguments to \code{\link{run}()}}
}
\description{
Deprecated function \code{make}. Use \code{link{run}()} instead.
}
\seealso{
\code{\link{run}()}
}

