% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run.R
\name{make}
\alias{make}
\title{Function \code{make}}
\usage{
make(plan, targets = possible_targets(plan), envir = parent.frame(),
  verbose = TRUE, parallelism = parallelism_choices(), jobs = 1,
  packages = (.packages()), prework = character(0),
  prepend = character(0), command = "make",
  args = default_system2_args(jobs = jobs, verbose = verbose))
}
\arguments{
\item{plan}{same as in function \code{\link{run}()}}

\item{targets}{same as in function \code{\link{run}()}}

\item{envir}{same as in function \code{\link{run}()}}

\item{verbose}{same as in function \code{\link{run}()}}

\item{parallelism}{same as in function \code{\link{run}()}}

\item{jobs}{same as in function \code{\link{run}()}}

\item{packages}{same as in function \code{\link{run}()}}

\item{prework}{same as in function \code{\link{run}()}}

\item{prepend}{same as in function \code{\link{run}()}}

\item{command}{same as in function \code{\link{run}()}}

\item{args}{same as in function \code{\link{run}()}}
}
\description{
Same as \code{\link{run}}. 
Type \code{?\link{run}} for more.
}

