% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make.R
\name{make_imports}
\alias{make_imports}
\title{Function make_imports}
\usage{
make_imports(config)
}
\arguments{
\item{config}{a configuration list returned by \code{\link{config}()}}
}
\description{
just make the imports
}
\examples{
\dontrun{
load_basic_example()
con <- config(my_plan)
make_imports(config = con)
}
}
\seealso{
\code{\link{make}}, \code{\link{config}}
}
