% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make.R
\name{make_targets}
\alias{make_targets}
\title{Just build the targets.}
\usage{
make_targets(config)
}
\arguments{
\item{config}{a configuration list returned by \code{\link{config}()}}
}
\value{
The master internal configuration list
used by \code{\link{make}()}.
}
\description{
Targets are the workflow steps listed
in your workflow plan data frame.
(See \code{\link{drake_plan}()}.)
}
\examples{
\dontrun{
test_with_dir("Quarantine side effects.", {
load_basic_example() # Load the canonical example.
# Generate the master internal configuration list.
con <- drake_config(my_plan)
# Just cache the imports, do not build any targets.
make_imports(config = con)
# Just make the targets
make_targets(config = con)
})
}
}
\seealso{
\code{\link{make}}, \code{\link{config}},
\code{\link{make_imports}}
}
