% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make.R
\name{make_with_config}
\alias{make_with_config}
\title{Run \code{\link[=make]{make()}},
on an existing internal configuration list.}
\usage{
make_with_config(config)
}
\arguments{
\item{config}{An input internal configuration list}
}
\value{
An output internal configuration list
}
\description{
Use \code{\link[=drake_config]{drake_config()}}
to create the \code{config} argument.
}
\examples{
\dontrun{
test_with_dir("Quarantine side effects.", {
load_basic_example() # Get the code with drake_example("basic").
# The following lines are the same as make(my_plan)
config <- drake_config(my_plan) # Create the internal config list.
make_with_config(config = config) # Run the project, build the targets.
})
}
}
\seealso{
\code{\link[=make]{make()}}, \code{\link[=drake_config]{drake_config()}}
}
