% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-deprecate.R
\name{make_with_config}
\alias{make_with_config}
\title{deprecated}
\usage{
make_with_config(config = drake::read_drake_config())
}
\arguments{
\item{config}{a configuration list returned by \code{\link[=drake_config]{drake_config()}}}
}
\value{
nothing
}
\description{
Deprecated on 2019-01-04
}
\examples{
# deprecated
}
\seealso{
\code{\link[=make]{make()}}, \code{\link[=drake_config]{drake_config()}}
}
\keyword{internal}
