% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecate.R
\name{Makefile_recipe}
\alias{Makefile_recipe}
\title{Deprecated}
\usage{
Makefile_recipe(recipe_command = drake::default_recipe_command(),
  target = "your_target", cache_path = NULL)
}
\arguments{
\item{recipe_command}{character scalar}

\item{target}{character scalar}

\item{cache_path}{character scalar}
}
\value{
A character scalar
}
\description{
2019-01-03
}
\examples{
# deprecated
}
\keyword{internal}
