% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exec-memory.R
\name{manage_memory}
\alias{manage_memory}
\title{Manage in-memory targets}
\usage{
manage_memory(targets, config, downstream = NULL, jobs = 1)
}
\arguments{
\item{targets}{Character vector of targets.}

\item{config}{\code{\link[=drake_config]{drake_config()}} list.}

\item{downstream}{Optional, character vector of any targets
assumed to be downstream.}

\item{jobs}{Number of jobs for local parallel computing}
}
\value{
Nothing.
}
\description{
Load targets that you need to build the targets
and unload the ones you will never need again in the
current runthrough of the pipeline. This function should
not be used directly by users. Only exported for
internal reasons.
}
\examples{
# Users should use make().
}
\keyword{internal}
