% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/defunct.R
\name{max_useful_jobs}
\alias{max_useful_jobs}
\title{max_useful_jobs \ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#defunct}{\figure{lifecycle-defunct.svg}{options: alt='[Defunct]'}}}{\strong{[Defunct]}}}
\usage{
max_useful_jobs(...)
}
\arguments{
\item{...}{Arguments}
}
\description{
2019-05-16
}
\keyword{internal}
