% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hooks.R
\name{message_sink_hook}
\alias{message_sink_hook}
\title{Function message_sink_hook}
\usage{
message_sink_hook(code)
}
\arguments{
\item{code}{code to run to build the target.}
}
\description{
an example \code{hook} argument to
\code{make()} that redirects error messages to files.
}
\examples{
\dontrun{
message_sink_hook({
  cat(1234)
  stop(5678)
})
x <- workplan(loud = cat(1234), bad = stop(5678))
make(x, hook = message_sink_hook)
}
}
\seealso{
\code{\link{make}}, \code{\link{silencer_hook}},
\code{\link{output_sink_hook}}
}
