% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/defunct.R
\name{migrate_drake_project}
\alias{migrate_drake_project}
\title{migrate_drake_project \ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#defunct}{\figure{lifecycle-defunct.svg}{options: alt='[Defunct]'}}}{\strong{[Defunct]}}}
\usage{
migrate_drake_project(...)
}
\arguments{
\item{...}{Arguments}
}
\description{
2019-05-16
}
\keyword{internal}
