% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecate.R
\name{migrate_drake_project}
\alias{migrate_drake_project}
\title{Deprecated: reconfigure an old project (built with drake <= 4.4.0)
to be compatible with later versions of drake.}
\usage{
migrate_drake_project(path = NULL, jobs = 1)
}
\arguments{
\item{path}{Full path to the cache.}

\item{jobs}{number of jobs for light parallelism.}
}
\description{
Deprecated on May 4, 2018.
This function was intended to migrate a project/cache from
drake 4.4.0 or earlier
to be compatible with the version of drake on your system.
}
\examples{
# This function is deprecated.
}
\seealso{
\code{\link[=rescue_cache]{rescue_cache()}}, \code{\link[=make]{make()}}
}
\keyword{internal}
