% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outdated.R
\name{missed}
\alias{missed}
\title{Function \code{missed}}
\usage{
missed(config)
}
\arguments{
\item{config}{internal runtime parameter list of
\code{\link{make}(...)},
produced by both \code{\link{drake_config}()} and \code{\link{make}()}.}
}
\value{
Character vector of names of missing objects and files.
}
\description{
Report any import objects required by your plan_drake
plan but missing from your workspace.
IMPORTANT: you must be in the root directory of your project.
}
\examples{
\dontrun{
test_with_dir("Quarantine side effects.", {
config <- load_basic_example() # Load the canonical example.
missed(config) # All the imported files and objects should be present.
rm(reg1) # Remove an import dependency from you workspace.
missed(config) # Should report that reg1 is missing.
})
}
}
\seealso{
\code{\link{outdated}}
}
