% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Makefile.R
\name{mk}
\alias{mk}
\title{Function \code{mk}}
\usage{
mk(target)
}
\arguments{
\item{target}{name of target to make}
}
\description{
Internal drake function to be called 
inside Makefiles only. Makes a single target.
Users, do not invoke directly.
}
