% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache.R
\name{new_cache}
\alias{new_cache}
\title{Function new_cache}
\usage{
new_cache(path = drake::default_cache_path(), verbose = 1,
  type = drake::default_cache_type(),
  short_hash_algo = drake::default_short_hash_algo(),
  long_hash_algo = drake::default_long_hash_algo(), ...)
}
\arguments{
\item{path}{file path to the cache if the cache
is a file system cache.}

\item{verbose}{logical, whether to print out the path of the cache.}

\item{type}{character scalar, type of the drake cache.
Must be among the list of supported caches
in \code{\link{cache_types}()}.}

\item{short_hash_algo}{short hash algorithm for the cache.
See \code{\link{default_short_hash_algo}()} and
\code{\link{make}()}}

\item{long_hash_algo}{long hash algorithm for the cache.
See \code{\link{default_long_hash_algo}()} and
\code{\link{make}()}}

\item{...}{other arguments to the cache constructor}
}
\value{
A newly created drake cache as a storr object.
}
\description{
Make a new drake cache. Could be any
type of cache in \code{\link{cache_types}()}.
}
\examples{
test_with_dir("Quarantine side effects.", {
\dontrun{
clean(destroy = TRUE)
unlink("not_hidden", recursive = TRUE)
cache1 <- new_cache() # Creates a new hidden '.drake' folder.
cache2 <- new_cache(path = "not_hidden", short_hash_algo = "md5")
clean(destroy = TRUE, cache = cache2)
}
# Create a storr_environment() cache from a custom environment.
e <- new.env() # Create a new environment.
ls(e) # Emtpy.
y <- new_cache(type = "storr_environment", envir = e)
ls(e) # Storr populates the environment for use in an in-memory cache.
})
}
\seealso{
\code{\link{default_short_hash_algo}},
\code{\link{default_long_hash_algo}},
\code{\link{make}}, \code{\link{cache_types}},
\code{\link{in_memory_cache_types}}
}
