% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parallel_stages.R
\name{next_stage}
\alias{next_stage}
\title{Function next_stage}
\usage{
next_stage(config)
}
\arguments{
\item{config}{A master configuration list produced by
\code{\link{drake_config}()} or \code{\link{make}()}}
}
\value{
A character vector of the targets to be made
in the first parallel stage of the next call to \code{\link{make}}.
}
\description{
List the targets that will be made in the
first parallel stage in the next call to \code{\link{make}}
}
\examples{
\dontrun{
test_with_dir("Quarantine side effects.", {
config <- load_basic_example() # drake's canonical example
next_stage(config = config)    # "small" and "large"
})
}
}
\seealso{
\code{\link{make}}, \code{\link{drake_config}}
}
