% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outdated.R
\name{outdated}
\alias{outdated}
\title{Function outdated}
\usage{
outdated(plan, targets = drake::possible_targets(plan),
  envir = parent.frame(), verbose = FALSE)
}
\arguments{
\item{plan}{Workflow plan data frame: that is, a data frame with a "target"
column and a "command" column. See \code{\link{plan}()}.}

\item{targets}{Character vector of targets. status() shows the status
of all the dependencies of the targets listed.}

\item{envir}{environment with your imported objects. Defaults to your workspace.}

\item{verbose}{logical, whether be verbose. If \code{TRUE}, objects are listed
as they are imported. (Objects need to be imported to check the status.)}
}
\description{
Check which targets are out of date and need to be rebuilt.
IMPORTANT: you must be in the root directory of your project.
}
\examples{
\dontrun{
load_basic_example()
outdated(my_plan)
make(my_plan)
outdated(my_plan)
}
}
\seealso{
\code{\link{plan}}, \code{\link{make}}, \code{\link{plot_graph}}
}
