% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outdated.R
\name{outdated}
\alias{outdated}
\title{Function outdated}
\usage{
outdated(plan = workflow(), targets = drake::possible_targets(plan),
  envir = parent.frame(), verbose = TRUE, cache = drake::get_cache(),
  parallelism = drake::default_parallelism(), jobs = 1,
  packages = (.packages()), prework = character(0), config = NULL)
}
\arguments{
\item{plan}{same as for \code{\link{make}}}

\item{targets}{same as for \code{\link{make}}}

\item{envir}{same as for \code{\link{make}}. Overrides
\code{config$envir}.}

\item{verbose}{same as for \code{\link{make}}}

\item{cache}{optional drake cache. See code{\link{new_cache}()}.
The \code{cache} argument is ignored if a
non-null \code{config} argument is supplied.}

\item{parallelism}{same as for \code{\link{make}}}

\item{jobs}{same as for \code{\link{make}}}

\item{packages}{same as for \code{\link{make}}}

\item{prework}{same as for \code{\link{make}}}

\item{config}{option internal runtime parameter list of
\code{\link{make}(...)},
produced with \code{\link{config}()}.
\code{config$envir} is ignored.
Overrides all the other arguments if not \code{NULL}.
For example,
\code{plan} is replaced with \code{config$plan}.
Computing \code{config}
in advance could save time if you plan multiple calls to
\code{outdated()}.}
}
\description{
Check which targets are out of date and need to be rebuilt.
IMPORTANT: you must be in the root directory of your project.
}
\examples{
\dontrun{
load_basic_example()
outdated(my_plan)
make(my_plan)
outdated(my_plan)
}
}
\seealso{
\code{\link{missed}}, \code{\link{workflow}},
\code{\link{make}}, \code{\link{plot_graph}}
}
