% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outdated.R
\name{outdated}
\alias{outdated}
\title{Function outdated}
\usage{
outdated(plan = workplan(), targets = drake::possible_targets(plan),
  envir = parent.frame(), verbose = 2, hook = default_hook,
  cache = drake::get_cache(verbose = verbose),
  parallelism = drake::default_parallelism(), jobs = 1,
  packages = rev(.packages()), prework = character(0), graph = NULL,
  config = NULL, make_imports = TRUE)
}
\arguments{
\item{plan}{same as for \code{\link{make}}}

\item{targets}{same as for \code{\link{make}}}

\item{envir}{same as for \code{\link{make}}. Overrides
\code{config$envir}.}

\item{verbose}{same as for \code{\link{make}}}

\item{hook}{same as for \code{\link{make}}}

\item{cache}{optional drake cache. See code{\link{new_cache}()}.
The \code{cache} argument is ignored if a
non-null \code{config} argument is supplied.}

\item{parallelism}{same as for \code{\link{make}}}

\item{jobs}{same as for \code{\link{make}}}

\item{packages}{same as for \code{\link{make}}}

\item{prework}{same as for \code{\link{make}}}

\item{graph}{same as for \code{\link{make}}}

\item{config}{option internal runtime parameter list of
\code{\link{make}(...)},
produced with \code{\link{config}()}.
\code{config$envir} is ignored.
Overrides all the other arguments if not \code{NULL}.
For example,
\code{plan} is replaced with \code{config$plan}.}

\item{make_imports}{logical, whether to import external files
and objects from the user's workspace to determine
which targets are up to date. If \code{FALSE}, the computation
is faster, but all the relevant information is drawn from the cache
and may be out of date.}
}
\value{
Character vector of the names of outdated targets.
}
\description{
Check which targets are out of date and need to be rebuilt.
IMPORTANT: you must be in the root directory of your project.
}
\examples{
\dontrun{
load_basic_example() # Load the canonical example of drake.
outdated(my_plan) # Which targets are out of date?
make(my_plan) # Run the projects, build the targets.
# Now, everything should be up to date (no targets listed).
outdated(my_plan)
}
}
\seealso{
\code{\link{missed}}, \code{\link{workplan}},
\code{\link{make}}, \code{\link{vis_drake_graph}}
}
