% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hooks.R
\name{output_sink_hook}
\alias{output_sink_hook}
\title{Function output_sink_hook}
\usage{
output_sink_hook(code)
}
\arguments{
\item{code}{code to run to build the target.}
}
\description{
an example \code{hook} argument to
\code{make()} that redirects output messages to files.
}
\examples{
\dontrun{
output_sink_hook({
  cat(1234)
  stop(5678)
})
x <- workplan(loud = cat(1234), bad = stop(5678))
make(x, hook = output_sink_hook)
}
}
\seealso{
\code{\link{make}}, \code{\link{silencer_hook}},
\code{\link{message_sink_hook}}
}
