% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecate.R
\name{parallel_stages}
\alias{parallel_stages}
\title{Defunct function}
\usage{
parallel_stages(config = drake::read_drake_config(),
  from_scratch = FALSE)
}
\arguments{
\item{config}{An configuration list output by
\code{\link[=make]{make()}} or \code{\link[=drake_config]{drake_config()}}.}

\item{from_scratch}{logical, whether to assume
that the next \code{\link[=make]{make()}} will run from scratch
so that all targets are attempted.}
}
\value{
A data frame of information spelling out how
targets are divided into parallelizable stages
(according to the \code{stage} column).
}
\description{
Staged parallelism is removed from drake,
so this function is moot.
Drake uses a much better parallel algorithm now.
}
\details{
Made defunct on May 4, 2018.
}
\examples{
# Do not use this function.
}
\keyword{internal}
