% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{parallelism_choices}
\alias{parallelism_choices}
\title{Names of old parallel backends
\lifecycle{deprecated}}
\usage{
parallelism_choices(distributed_only = FALSE)
}
\arguments{
\item{distributed_only}{Logical.}
}
\value{
character vector
}
\description{
2019-01-03
}
\keyword{internal}
