% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-deprecate.R
\name{plan_analyses}
\alias{plan_analyses}
\title{Deprecated.}
\usage{
plan_analyses(plan, datasets, sep = "_")
}
\arguments{
\item{plan}{workflow plan data frame of analysis methods.
The commands in the \code{command} column must
have the \code{dataset__} wildcard where the datasets go.
For example, one command could be \code{lm(dataset__)}. Then,
the commands in the output will include \code{lm(your_dataset_1)},
\code{lm(your_dataset_2)}, etc.}

\item{datasets}{workflow plan data frame with instructions
to make the datasets.}

\item{sep}{character scalar, delimiter for creating
the names of new targets}
}
\value{
An evaluated workflow plan data frame of analysis targets.
}
\description{
Use \code{\link[=evaluate_plan]{evaluate_plan()}} and \code{\link[=expand_plan]{expand_plan()}} instead.
}
\details{
2019-01-13
}
\examples{
# Deprecated
}
\keyword{internal}
