% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecate.R
\name{plan_drake}
\alias{plan_drake}
\title{Deprecated function \code{plan_drake}}
\usage{
plan_drake(..., list = character(0), file_targets = FALSE,
  strings_in_dots = c("filenames", "literals"))
}
\arguments{
\item{...}{Same as for \code{\link[=drake_plan]{drake_plan()}}.}

\item{list}{Same as for \code{\link[=drake_plan]{drake_plan()}}.}

\item{file_targets}{Same as for \code{\link[=drake_plan]{drake_plan()}}.}

\item{strings_in_dots}{Same as for \code{\link[=drake_plan]{drake_plan()}}.}
}
\value{
A data frame of targets and commands.
}
\description{
Use \code{\link[=drake_plan]{drake_plan()}} instead.
}
\details{
Deprecated on 2017-12-12.
}
\examples{
# See ?drake_plan for examples.
}
\seealso{
\code{\link[=drake_plan]{drake_plan()}}
}
\keyword{internal}
