% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotgraph.R
\name{plot_graph}
\alias{plot_graph}
\title{Function \code{plot_graph}}
\usage{
plot_graph(plan, targets = drake::possible_targets(plan),
  envir = parent.frame(), verbose = FALSE, font_size = 20, graph = NULL,
  ...)
}
\arguments{
\item{plan}{workflow plan data frame, same as for function 
\code{\link{make}()}.}

\item{targets}{names of targets to bulid, same as for function
\code{\link{make}()}.}

\item{envir}{environment to import from, same as for function
\code{\link{make}()}.}

\item{verbose}{logical, whether to output messages to the console.}

\item{font_size}{numeric, font size of the node labels in the graph}

\item{graph}{an igraph object if one has already been built with 
\code{\link{build_graph}()}.}

\item{...}{arguments passed to \code{visNetwork::visNetwork()} to plot the graph.}
}
\description{
Plot the dependency structure of your workflow.
IMPORTANT: you must be in the root directory of your project.
}
\examples{
\dontrun{
load_basic_example()
plot_graph(my_plan)
make(my_plan)
plot_graph(my_plan)
}
}
\seealso{
\code{\link{build_graph}}
}
