% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotgraph.R
\name{plot_graph}
\alias{plot_graph}
\title{Function \code{plot_graph}}
\usage{
plot_graph(plan, targets = drake::possible_targets(plan),
  envir = parent.frame(), verbose = TRUE, jobs = 1,
  parallelism = drake::default_parallelism(), packages = (.packages()),
  prework = character(0), config = NULL, font_size = 20,
  navigationButtons = TRUE, ...)
}
\arguments{
\item{plan}{workflow plan data frame, same as for function 
\code{\link{make}()}.}

\item{targets}{names of targets to bulid, same as for function
\code{\link{make}()}.}

\item{envir}{environment to import from, same as for function
\code{\link{make}()}.}

\item{verbose}{logical, whether to output messages to the console.}

\item{jobs}{The \code{outdated()} function is called internally, 
and it needs to import objects and examine your 
input files to see what has been updated. This could take some time,
and parallel computing may be needed
to speed up the process. The \code{jobs} argument is number of parallel jobs 
to use for faster computation.}

\item{parallelism}{Choice of parallel backend to speed up the computation.
See \code{?parallelism_choices} for details. The Makefile option is not available
here. Drake will try to pick the best option for your system by default.}

\item{packages}{same as for \code{\link{make}()}.}

\item{prework}{same as for \code{\link{make}()}.}

\item{config}{option internal runtime parameter list of 
\code{\link{make}(...)},
produced with \code{\link{get_config}()}.
Computing this
in advance could save time if you plan multiple calls to 
\code{outdated()}.}

\item{font_size}{numeric, font size of the node labels in the graph}

\item{navigationButtons}{logical, whether to add navigation buttons with 
\code{visNetwork::visInteraction(navigationButtons = TRUE)}}

\item{...}{other arguments passed to \code{visNetwork::visNetwork()} to plot the graph.}
}
\description{
Plot the dependency structure of your workflow.
IMPORTANT: you must be in the root directory of your project.
To save time for repeated plotting, this function is divided into
\code{\link{dataframes_graph}()} and \code{\link{render_graph}()}.
}
\examples{
\dontrun{
load_basic_example()
plot_graph(my_plan, width = "100\%") # The width is passed to visNetwork().
make(my_plan)
plot_graph(my_plan) # The red nodes from before are now green.
}
}
\seealso{
\code{\link{build_graph}}
}
