% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecate.R
\name{plot_graph}
\alias{plot_graph}
\title{Deprecated function \code{plot_graph}}
\usage{
plot_graph(plan = read_drake_plan(), targets = plan$target,
  envir = parent.frame(), verbose = 1, hook = default_hook,
  cache = drake::get_cache(verbose = verbose), jobs = 1,
  parallelism = drake::default_parallelism(),
  packages = rev(.packages()), prework = character(0), config = NULL,
  file = character(0), selfcontained = FALSE, build_times = TRUE,
  digits = 3, targets_only = FALSE, split_columns = FALSE,
  font_size = 20, layout = "layout_with_sugiyama", main = NULL,
  direction = "LR", hover = TRUE, navigationButtons = TRUE,
  from = NULL, mode = c("out", "in", "all"), order = NULL,
  subset = NULL, ncol_legend = 1, make_imports = TRUE,
  from_scratch = FALSE, ...)
}
\arguments{
\item{plan}{deprecated}

\item{envir}{deprecated}

\item{verbose}{deprecated}

\item{hook}{deprecated}

\item{cache}{deprecated}

\item{jobs}{deprecated}

\item{packages}{deprecated}

\item{prework}{deprecated}

\item{config}{deprecated}

\item{file}{deprecated}

\item{build_times}{deprecated}

\item{digits}{deprecated}

\item{targets_only}{deprecated}

\item{split_columns}{deprecated}

\item{font_size}{deprecated}

\item{layout}{deprecated}

\item{main}{deprecated}

\item{direction}{deprecated}

\item{hover}{deprecated}

\item{navigationButtons}{deprecated}

\item{from}{deprecated}

\item{mode}{deprecated}

\item{order}{deprecated}

\item{subset}{deprecated}

\item{ncol_legend}{deprecated}

\item{make_imports}{deprecated}

\item{from_scratch}{deprecated}

\item{...}{deprecated}
}
\value{
A \code{visNetwork} graph.
}
\description{
Use \code{\link[=vis_drake_graph]{vis_drake_graph()}} instead.
}
\details{
Deprecated on 2017-10.
}
\examples{
# See ?vis_drake_graph for examples.
}
\keyword{internal}
