% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{possible_targets}
\alias{possible_targets}
\title{List the possible targets for the \code{targets}
argument to \code{\link[=make]{make()}}, given a workflow plan
data frame.}
\usage{
possible_targets(plan = read_drake_plan())
}
\arguments{
\item{plan}{workflow plan data frame}
}
\value{
Character vector of possible targets given the workflow plan.
}
\description{
Intended for internal use only.
}
\examples{
\dontrun{
test_with_dir("Quarantine side effects.", {
load_basic_example() # Get the code with drake_example("basic").
# List the possible targets you could choose for the
# `targets` argument to make(). You may choose any subset.
possible_targets(my_plan)
})
}
}
\seealso{
\code{\link[=make]{make()}}
}
\keyword{internal}
