% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/future_lapply_staged.R
\name{prepare_distributed}
\alias{prepare_distributed}
\title{Do the prep work for \code{future_lapply_staged} parallelism.}
\usage{
prepare_distributed(config)
}
\arguments{
\item{config}{Internal configuration list from
\code{\link[=drake_config]{drake_config()}}.}
}
\value{
Nothing.
}
\description{
For internal use only. Do not call directly.
}
\examples{
# Not a user-side function.
}
\keyword{internal}
