% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distributed.R
\name{prepare_distributed}
\alias{prepare_distributed}
\title{Do the preparatory work
for \code{\link{make}()} with a distributed computing
backend (see the \code{parallelism} argument of \code{\link{make}()}).}
\usage{
prepare_distributed(config)
}
\arguments{
\item{config}{Internal configuration list from
\code{\link{drake_config}()}.}
}
\value{
Nothing.
}
\description{
For internal use only. Exported to flesh out some
of the more advanced examples.
}
\examples{
\dontrun{
test_with_dir("Quarantine side effects.", {
load_basic_example() # Get the code with drake_example("basic").
config <- drake_config(my_plan)
prepare_distributed(config = config)
})
}
}
\keyword{internal}
