% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_imports.R
\name{process_import}
\alias{process_import}
\title{Process an imported data object
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}}
\usage{
process_import(import, config)
}
\arguments{
\item{import}{Character, name of an import to process}

\item{config}{\code{\link[=drake_config]{drake_config()}} object}
}
\description{
For internal use only. Not a user-side function.
}
\keyword{internal}
