% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-deprecate.R
\name{process_import}
\alias{process_import}
\title{Deprecated. An internal function only used
inside process_imports().}
\usage{
process_import(import, config)
}
\arguments{
\item{import}{character, name of an import to process}

\item{config}{\code{\link[=drake_config]{drake_config()}} object}
}
\description{
Deprecated on 2019-01-12.
}
\keyword{internal}
