% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exec-imports.R
\name{process_import}
\alias{process_import}
\title{internal function}
\usage{
process_import(import, config)
}
\arguments{
\item{import}{character, name of an import to process}

\item{config}{\code{\link[=drake_config]{drake_config()}} object}
}
\description{
only used inside process_imports(). Not a user-side function.
}
\examples{
# Not a user-side function.
}
\keyword{internal}
