% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean.R
\name{prune}
\alias{prune}
\title{Function \code{prune}}
\usage{
prune(plan)
}
\arguments{
\item{plan}{workflow plan data frame, as generated by \code{\link{plan}}.}
}
\description{
Removes any cached targets not listed in 
\code{plan$target}, including file targets and their 
cached fingerprints. 
Your working directory (\code{\link{getwd}()}) must be the
root directory of your project.
WARNING: this removes objects and files generated by drake.
Only do this if you're sure you won't lose any important work.
}
\seealso{
\code{\link{clean}}, \code{\link{run}}
}

