% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecate.R
\name{prune}
\alias{prune}
\title{Deprecated function \code{prune}}
\usage{
prune(plan = workplan())
}
\arguments{
\item{plan}{workflow plan data frame, as generated by
\code{\link{workplan}}.}
}
\description{
Use \code{\link{clean}()} instead
}
\examples{
\dontrun{
load_basic_example()
make(my_plan)
cached()
prune(my_plan[1:3,])
cached()
make(my_plan)
clean(destroy = TRUE)
}
}
\seealso{
\code{\link{clean}}, \code{\link{make}}
}
