% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecate.R
\name{prune}
\alias{prune}
\title{Deprecated function \code{prune}}
\usage{
prune(plan = workplan())
}
\arguments{
\item{plan}{workflow plan data frame, as generated by
\code{\link{workplan}}.}
}
\value{
\code{NULL}
}
\description{
Use \code{\link{clean}()} instead
}
\examples{
# See ?clean for examples.
}
\seealso{
\code{\link{clean}}, \code{\link{make}}
}
