% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean.R
\name{prune}
\alias{prune}
\title{Function \code{prune}}
\usage{
prune(plan)
}
\arguments{
\item{plan}{data frame, plan as generated with \code{\link{plan}}.}
}
\description{
Removes any cached target objects and generated 
files not listed in \code{plan$target$}. 
Your working directory (\code{\link{getwd}()}) must be the
root directory of your project.
WARNING: this removes files.
Only do this if you're sure you won't lose any important work.
}
\seealso{
\code{\link{clean}}, \code{\link{run}}
}

