% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean.R
\name{prune}
\alias{prune}
\title{Function \command{prune}}
\usage{
prune(plan)
}
\arguments{
\item{plan}{data frame, plan as generated with \command{\link{plan}}.}
}
\description{
Removes any cached target objects and generated 
files not listed in \command{plan$target$}. 
Your working directory (\command{\link{getwd}()}) must be the
root directory of your project.
WARNING: this removes files.
Only do this if you're sure you won't lose any important work.
}
\seealso{
\command{\link{clean}}, \command{\link{run}}
}

