% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean.R
\name{prune}
\alias{prune}
\title{Function \code{prune}}
\usage{
prune(plan, targets = plan$targets, envir = parent.frame())
}
\arguments{
\item{plan}{workflow plan data frame, as generated by 
\code{\link{plan}}.}

\item{targets}{character string, names of targets to
keep in the cache. All dependencies are kept too.}

\item{envir}{environment to use, same as in \code{\link{run}()}.
This is where your imported objects should have come from.
Use \code{help("run")} for details.}
}
\description{
Removes any cached targets not listed in 
\code{plan$target}, including file targets and their 
cached fingerprints. 
Your working directory (\code{\link{getwd}()}) must be the
root directory of your project.
WARNING: this removes objects and files generated by drake.
Only do this if you're sure you won't lose any important work.
}
\seealso{
\code{\link{clean}}, \code{\link{run}}
}

