% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecate.R
\name{prune}
\alias{prune}
\title{Deprecated function \code{prune}}
\usage{
prune(plan)
}
\arguments{
\item{plan}{workflow plan data frame, as generated by 
\code{\link{plan}}.}
}
\description{
Use \code{\link{clean}()} instead
}
\seealso{
\code{\link{clean}}, \code{\link{make}}
}

