% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecate.R
\name{prune}
\alias{prune}
\title{Deprecated function \code{prune}}
\usage{
prune(plan = workplan())
}
\arguments{
\item{plan}{workflow plan data frame, as generated by
\code{\link{workplan}}.}
}
\description{
Use \code{\link{clean}()} instead
}
\examples{
\dontrun{
load_basic_example() # Load the canonical example.
make(my_plan) # Run the targets, build the project.
cached() # List the targets/imports in the cache.
# Remove all the cached targets except the first 3 targets
# in the workflow plan.
prune(my_plan[1:3,])
cached() # Only the first 3 targets remain, plus all the imports.
make(my_plan) # Re-make the deleted targets.
clean(destroy = TRUE) # Completely remove the cache.
}
}
\seealso{
\code{\link{clean}}, \code{\link{make}}
}
