% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-deprecate.R
\name{prune_drake_graph}
\alias{prune_drake_graph}
\title{deprecated}
\usage{
prune_drake_graph(graph, to = igraph::V(graph)$name, jobs = 1)
}
\arguments{
\item{graph}{an igraph object}

\item{to}{character vector of vertices}

\item{jobs}{number of jobs for parallelism}
}
\value{
an \code{igraph} object
}
\description{
2019-01-08
}
\examples{
# deprectaed
}
\keyword{internal}
