% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{r_recipe_wildcard}
\alias{r_recipe_wildcard}
\title{Default Makefile recipe wildcard
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}}
\usage{
r_recipe_wildcard()
}
\value{
The R recipe wildcard.
}
\description{
2019-01-02
}
\keyword{internal}
