% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-deprecate.R
\name{r_recipe_wildcard}
\alias{r_recipe_wildcard}
\title{deprecated}
\usage{
r_recipe_wildcard()
}
\value{
The R recipe wildcard.
}
\description{
2019-01-02
}
\examples{
# deprecated
}
\keyword{internal}
