% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recipe.R
\name{r_recipe_wildcard}
\alias{r_recipe_wildcard}
\title{Show the R recipe wildcard
for \code{make(..., parallelism = "Makefile")}.}
\usage{
r_recipe_wildcard()
}
\value{
The R recipe wildcard.
}
\description{
Relevant to \code{"Makefile"} parallelism only.
}
\examples{
r_recipe_wildcard()
}
\seealso{
\code{\link[=default_recipe_command]{default_recipe_command()}}
}
