% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recipe.R
\name{r_recipe_wildcard}
\alias{r_recipe_wildcard}
\title{Function \code{r_recipe_wildcard}}
\usage{
r_recipe_wildcard()
}
\value{
The R recipe wildcard.
}
\description{
Gives the R recipe wildcard
for Makefiles.
}
\details{
See the help file of \code{\link{Makefile_recipe}}
for details and examples.
}
\examples{
r_recipe_wildcard()
}
\seealso{
\code{\link{default_recipe_command}}
}
