% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-deprecate.R
\name{rate_limiting_times}
\alias{rate_limiting_times}
\title{Defunct}
\usage{
rate_limiting_times(config = drake::read_drake_config(),
  targets = NULL, from_scratch = FALSE, targets_only = FALSE,
  future_jobs = 1, digits = 3)
}
\arguments{
\item{config}{option internal runtime parameter list of
\code{\link{make}(...)},
produced by both \code{\link[=make]{make()}} and
\code{\link[=drake_config]{drake_config()}}.}

\item{targets}{Character vector, names of targets.
Find the rate-limiting times for building these targets
plus dependencies.
Defaults to all targets.}

\item{from_scratch}{logical, whether to assume
next hypothetical call to \code{\link[=make]{make()}}
is a build from scratch (after \code{\link[=clean]{clean()}}).}

\item{targets_only}{logical, whether to factor in just the
targets or use times from everything, including the imports.}

\item{future_jobs}{hypothetical number of jobs
assumed for the predicted runtime.
assuming this number of jobs.}

\item{digits}{number of digits for rounding the times.}
}
\value{
A data frame of times of the worst-case scenario
rate-limiting targets in each parallelizable stage.
}
\description{
This function is now moot because
staged parallelism in \code{drake} was replaced
by a much better scheduling algorithm.
}
\details{
Made defunct on May 4, 2018
}
\examples{
# Do not use this function.
}
\keyword{internal}
