% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{read_config}
\alias{read_config}
\title{Function \code{read_config}}
\usage{
read_config(path = getwd(), search = TRUE)
}
\arguments{
\item{path}{Root directory of the drake project,
or if \code{search} is \code{TRUE}, either the
project root or a subdirectory of the project.}

\item{search}{logical. If \code{TRUE}, search parent directories
to find the nearest drake cache. Otherwise, look in the
current working directory only.}
}
\value{
a named list of configuration items
}
\description{
Read all the configuration parameters 
from your last attempted call to \code{\link{make}()}.
These include the workflow plan
}
\seealso{
\code{\link{make}}
}
