% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-deprecate.R
\name{read_drake_graph}
\alias{read_drake_graph}
\title{Deprecated}
\usage{
read_drake_graph(path = getwd(), search = TRUE, cache = NULL,
  verbose = 1L)
}
\arguments{
\item{path}{Root directory of the drake project,
or if \code{search} is \code{TRUE}, either the
project root or a subdirectory of the project.
Ignored if a \code{cache} is supplied.}

\item{search}{Logical. If \code{TRUE}, search parent directories
to find the nearest drake cache. Otherwise, look in the
current working directory only.
Ignored if a \code{cache} is supplied.}

\item{cache}{drake cache. See \code{\link[=new_cache]{new_cache()}}.
If supplied, \code{path} and \code{search} are ignored.}

\item{verbose}{Integer, control printing to the console/terminal.
\itemize{
\item \code{0}: print nothing.
\item \code{1}: print targets, retries, and failures.
\item \code{2}: also show a spinner when preprocessing tasks are underway.
}}
}
\description{
drake no longer stores the config object,
the plan, etc. in the cache during \code{make()}. This change
improves speed.
}
\details{
2019-01-06
}
\examples{
# deprecated
}
\keyword{internal}
