% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecate.R
\name{read_drake_meta}
\alias{read_drake_meta}
\title{Deprecated: read the metadata of a target or import.}
\usage{
read_drake_meta(targets = NULL, path = getwd(), search = TRUE,
  cache = NULL, verbose = 1, jobs = 1)
}
\arguments{
\item{targets}{character vector, names of the targets
to get metadata. If \code{NULL}, all metadata is collected.}

\item{path}{Root directory of the drake project,
or if \code{search} is \code{TRUE}, either the
project root or a subdirectory of the project.}

\item{search}{logical. If \code{TRUE}, search parent directories
to find the nearest drake cache. Otherwise, look in the
current working directory only.}

\item{cache}{optional drake cache. See code{\link{new_cache}()}.
If \code{cache} is supplied,
the \code{path} and \code{search} arguments are ignored.}

\item{verbose}{whether to print console messages}

\item{jobs}{number of jobs for light parallelism.
Supports 1 job only on Windows.}
}
\value{
The cached master internal configuration list
of the last \code{\link[=make]{make()}}.
}
\description{
Use \code{\link[=diagnose]{diagnose()}} instead.
}
\details{
Deprecated on 2018-02-09
}
\examples{
# See ?diagnose() for examples.
}
\seealso{
\code{\link[=diagnose]{diagnose()}} \code{\link[=dependency_profile]{dependency_profile()}}, \code{\link[=make]{make()}}
}
\keyword{internal}
