% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-deprecate.R
\name{read_drake_plan}
\alias{read_drake_plan}
\title{Deprecated}
\usage{
read_drake_plan(path = getwd(), search = TRUE, cache = NULL,
  verbose = 1L)
}
\arguments{
\item{path}{Root directory of the drake project,
or if \code{search} is \code{TRUE}, either the
project root or a subdirectory of the project.
Ignored if a \code{cache} is supplied.}

\item{search}{Logical. If \code{TRUE}, search parent directories
to find the nearest drake cache. Otherwise, look in the
current working directory only.
Ignored if a \code{cache} is supplied.}

\item{cache}{drake cache. See \code{\link[=new_cache]{new_cache()}}.
If supplied, \code{path} and \code{search} are ignored.}

\item{verbose}{Logical or numeric, control printing to the console.
\itemize{
\item \code{0} or \code{FALSE}: print nothing.
\item \code{1} or \code{TRUE}: print only targets to build.
\item \code{2}: plus checks and cache info.
\item \code{3}: plus missing imports.
\item \code{4}: plus all imports.
\item \code{5}: plus execution and total build times for targets.
\item \code{6}: plus notifications when targets are being stored.
}}
}
\description{
drake no longer stores the config object,
the plan, etc. in the cache during \code{make()}. This change
improves speed.
}
\details{
2019-01-06
}
\examples{
# deprecated
}
\keyword{internal}
