% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecate.R
\name{read_graph}
\alias{read_graph}
\title{Deprecated function \code{read_graph}}
\usage{
read_graph(path = getwd(), search = TRUE, cache = drake::get_cache(path =
  path, search = search, verbose = verbose), verbose = TRUE, ...)
}
\arguments{
\item{path}{Same as for \code{\link{read_drake_graph}()}.}

\item{search}{Same as for \code{\link{read_drake_graph}()}.}

\item{cache}{Same as for \code{\link{read_drake_graph}()}.}

\item{verbose}{Same as for \code{\link{read_drake_graph}()}.}

\item{...}{Same as for \code{\link{read_drake_graph}()}.}
}
\value{
An \code{igraph} object representing the dependency
network of the workflow.
}
\description{
Use \code{\link{read_drake_graph}()} instead.
}
\details{
Deprecated on 2017-11-12.
}
\examples{
# See ?read_drake_graph for the examples.
}
\seealso{
\code{\link{read_drake_graph}}.
}
\keyword{internal}
