% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{read_plan}
\alias{read_plan}
\title{Function \code{read_plan}}
\usage{
read_plan(path = getwd(), search = TRUE)
}
\arguments{
\item{path}{Root directory of the drake project,
or if \code{search} is \code{TRUE}, either the
project root or a subdirectory of the project.}

\item{search}{logical. If \code{TRUE}, search parent directories
to find the nearest drake cache. Otherwise, look in the
current working directory only.}
}
\value{
a workflow plan data frame
}
\description{
Read the workflow plan
from your last attempted call to \code{\link{make}()}.
}
\seealso{
\code{\link{read_config}}
}
