% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{read_plan}
\alias{read_plan}
\title{Function \code{read_plan}}
\usage{
read_plan(path = getwd(), search = TRUE, cache = drake::get_cache(path =
  path, search = search, verbose = verbose), verbose = TRUE)
}
\arguments{
\item{path}{Root directory of the drake project,
or if \code{search} is \code{TRUE}, either the
project root or a subdirectory of the project.}

\item{search}{logical. If \code{TRUE}, search parent directories
to find the nearest drake cache. Otherwise, look in the
current working directory only.}

\item{cache}{optional drake cache. See code{\link{new_cache}()}.
If \code{cache} is supplied,
the \code{path} and \code{search} arguments are ignored.}

\item{verbose}{whether to print console messages}
}
\value{
A workflow plan data frame.
}
\description{
Read the workflow plan
from your last attempted call to \code{\link{make}()}.
}
\examples{
\dontrun{
load_basic_example() # Load the canonical example.
make(my_plan) # Run the project, build the targets.
read_plan() # Retrieve the workflow plan data frame from the cache.
}
}
\seealso{
\code{\link{read_config}}
}
