% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache.R
\name{readd}
\alias{readd}
\title{Function \command{readd}}
\usage{
readd(x, character_only = FALSE, path = getwd(), search = FALSE,
  envir = parent.frame())
}
\arguments{
\item{x}{If \command{character_only} is \command{TRUE}, 
\command{x} is a character variable storing the name of the item to get 
from the cache. Otherwise, \command{x} is a symbol with the literal 
object name.}

\item{character_only}{\command{TRUE}/\command{FALSE}, whether \command{x} 
can be assumed to be a character string 
(just as in \command{\link{library}()}).}

\item{path}{Root directory of the drake project,
or if \command{search} is \command{TRUE}, either the
project root or a subdirectory of the project.}

\item{search}{If \command{TRUE}, search parent directories
to find the nearest drake cache. Otherwise, look in the
current working directory only.}

\item{envir}{environment of imported functions (for lexical scoping)}
}
\value{
drake target item from the cache
}
\description{
Read a drake target object from the cache.
Does not delete the item from the cache.
}
\seealso{
\command{\link{loadd}}, \command{\link{cached}}, 
\command{\link{built}}, \command{link{imported}}, \command{\link{plan}}, 
\command{\link{run}}
}

