% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate.R
\name{reduce_plan}
\alias{reduce_plan}
\title{Write commands to reduce several targets down to one.}
\usage{
reduce_plan(plan = NULL, target = "target", begin = "", op = " + ",
  end = "", pairwise = TRUE)
}
\arguments{
\item{plan}{workflow plan data frame of prespecified targets}

\item{target}{name of the new aggregated target}

\item{gather}{function used to gather the targets. Should be
one of \code{\link{list}(...)}, \code{\link{c}(...)},
\code{\link{rbind}(...)}, or similar.}
}
\value{
A workflow plan data frame that aggregates multiple
prespecified targets into one additional target downstream.
}
\description{
Creates a new workflow plan data frame with a single new
target. This target is formed by combining the targets
in the argument plan using successive applications of a binary
operator.
}
\examples{
# Workflow plan for datasets:
x_plan <- evaluate_plan(
  drake_plan(x = VALUE),
  wildcard = "VALUE",
  values = 1:8
)
# Create a new target from the sum of the others.
reduce_plan(x_plan, target = "x_sum")
# For memory efficiency and parallel computing,
# reduce pairwise:
reduce_plan(x_plan, target = "x_sum", pairwise = TRUE)
}
