% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecate.R
\name{render_graph}
\alias{render_graph}
\title{Deprecated function \code{render_graph}}
\usage{
render_graph(graph_info, file = character(0),
  layout = "layout_with_sugiyama", direction = "LR", hover = TRUE,
  main = graph_info$default_title, selfcontained = FALSE,
  navigationButtons = TRUE, ncol_legend = 1, ...)
}
\arguments{
\item{graph_info}{Same as for \code{\link[=render_drake_graph]{render_drake_graph()}}.}

\item{file}{Same as for \code{\link[=render_drake_graph]{render_drake_graph()}}.}

\item{layout}{Same as for \code{\link[=render_drake_graph]{render_drake_graph()}}.}

\item{direction}{Same as for \code{\link[=render_drake_graph]{render_drake_graph()}}.}

\item{hover}{Same as for \code{\link[=render_drake_graph]{render_drake_graph()}}.}

\item{main}{Same as for \code{\link[=render_drake_graph]{render_drake_graph()}}.}

\item{selfcontained}{Same as for \code{\link[=render_drake_graph]{render_drake_graph()}}.}

\item{navigationButtons}{Same as for \code{\link[=render_drake_graph]{render_drake_graph()}}. # nolint}

\item{ncol_legend}{Same as for \code{\link[=render_drake_graph]{render_drake_graph()}}.}

\item{...}{Same as for \code{\link[=render_drake_graph]{render_drake_graph()}}.}
}
\value{
Same as for \code{\link[=render_drake_graph]{render_drake_graph()}}.
}
\description{
Use \code{\link[=render_drake_graph]{render_drake_graph()}} instead.
}
\details{
Deprecated on 2017-10.
}
\examples{
# See ?render_drake_graph for examples.
}
\seealso{
\code{\link[=render_drake_graph]{render_drake_graph()}}
}
\keyword{internal}
