% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecate.R
\name{render_graph}
\alias{render_graph}
\title{Deprecated function \code{render_graph}}
\usage{
render_graph(graph_info, file = character(0),
  layout = "layout_with_sugiyama", direction = "LR", hover = TRUE,
  main = graph_info$default_title, selfcontained = FALSE,
  navigationButtons = TRUE, ncol_legend = 1, ...)
}
\arguments{
\item{graph_info}{deprecated}

\item{file}{deprecated}

\item{layout}{deprecated}

\item{direction}{deprecated}

\item{hover}{deprecated}

\item{main}{deprecated}

\item{selfcontained}{deprecated}

\item{navigationButtons}{deprecated}

\item{ncol_legend}{deprecated}

\item{...}{deprecated}
}
\value{
A \code{visNetwork} graph.
}
\description{
Use \code{\link[=render_drake_graph]{render_drake_graph()}} instead.
}
\details{
Deprecated on 2017-10.
}
\examples{
# See ?render_drake_graph for examples.
}
\keyword{internal}
