% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotgraph.R
\name{render_graph}
\alias{render_graph}
\title{Function \code{render_graph}}
\usage{
render_graph(graph, layout = "layout_as_tree", navigationButtons = TRUE,
  ...)
}
\arguments{
\item{graph}{list of data frames generated by \code{\link{dataframes_graph}()}.
There should be 3 data frames: \code{nodes}, \code{edges}, 
and \code{legend_nodes}.}

\item{layout}{name of an igraph layout to use, such as "layout_with_sugiyama"
or "layout_as_tree"}

\item{navigationButtons}{logical, whether to add navigation buttons with 
\code{visNetwork::visInteraction(navigationButtons = TRUE)}}

\item{...}{arguments passed to \code{visNetwork()}.}
}
\description{
render a graph from the data frames 
generated by \code{\link{dataframes_graph}()}
}
\examples{
\dontrun{
load_basic_example()
graph = dataframes_graph(my_plan)
render_graph(graph, width = "100\%") # The width is passed to visNetwork().
}
}
