% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache_interaction.R
\name{rescue_cache}
\alias{rescue_cache}
\title{Function rescue_cache}
\usage{
rescue_cache(..., jobs = 1)
}
\arguments{
\item{...}{arguments to \code{\link{get_cache}}}

\item{jobs}{number of jobs for light parallelism
(disabled on Windows)}
}
\value{
Invisibly returns the cache.
}
\description{
Sometimes, \code{storr} caches may have
dangling orphaned files that prevent you from loading or cleaning.
This function tries to remove those files so you can use the
cache normally again.
}
\seealso{
\code{\link{get_cache}}, \code{\link{cached}}
}
