% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rstudio.R
\name{rs_addin_r_make}
\alias{rs_addin_r_make}
\title{RStudio addin for r_make()
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}}
\usage{
rs_addin_r_make(r_args = list())
}
\arguments{
\item{r_args}{List of arguments to \code{r_fn}, not including \code{func} or \code{args}.
Example:
\code{r_make(r_fn = callr::r_bg, r_args = list(stdout = "stdout.log"))}.}
}
\value{
Nothing.
}
\description{
Call \code{\link[=r_make]{r_make()}} in an RStudio addin.
}
\keyword{internal}
