% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run.R
\name{run}
\alias{run}
\title{Function \command{run}}
\usage{
run(plan, targets = plan$target, envir = parent.frame(), verbose = TRUE,
  parallelism = c("mclapply", "Makefile"), jobs = 1,
  prepend = character(0), packages = character(0), prework = character(0))
}
\arguments{
\item{plan}{workflow plan data frame}

\item{targets}{targets to bulid}

\item{envir}{environment to use}

\item{verbose}{logical, whether to print output to the console}

\item{parallelism}{character, type of parallelism to use}

\item{jobs}{number of parallel processes or jobs}

\item{prepend}{lines to prepend to the Makefile}

\item{packages}{packages to load}

\item{prepwork}{prework to do before running the workflow}
}
\description{
Run your project.
}

