% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecate.R
\name{session}
\alias{session}
\title{Deprecated function \code{session}}
\usage{
session(path = getwd(), search = TRUE, cache = drake::get_cache(path
  = path, search = search, verbose = verbose), verbose = TRUE)
}
\arguments{
\item{path}{Same as for \code{\link[=drake_get_session_info]{drake_get_session_info()}}.}

\item{search}{Same as for \code{\link[=drake_get_session_info]{drake_get_session_info()}}.}

\item{cache}{Same as for \code{\link[=drake_get_session_info]{drake_get_session_info()}}.}

\item{verbose}{Same as for \code{\link[=drake_get_session_info]{drake_get_session_info()}}.}
}
\value{
Same as for \code{\link[=drake_get_session_info]{drake_get_session_info()}}.
}
\description{
Use \code{\link[=drake_get_session_info]{drake_get_session_info()}} instead
}
\details{
Deprecated on 2017-11-12.
}
\examples{
# See ?drake_get_session_info for examples.
}
\seealso{
\code{\link[=drake_get_session_info]{drake_get_session_info()}}
}
\keyword{internal}
