% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/status.R
\name{session}
\alias{session}
\title{Function \code{session}}
\usage{
session(path = getwd(), search = TRUE)
}
\arguments{
\item{path}{Root directory of the drake project,
or if \code{search} is \code{TRUE}, either the
project root or a subdirectory of the project.}

\item{search}{If \code{TRUE}, search parent directories
to find the nearest drake cache. Otherwise, look in the
current working directory only.}
}
\value{
\code{\link{sessionInfo}()} of the last
call to \code{\link{make}()}
}
\description{
Load the \code{\link{sessionInfo}()}
of the last call to \code{\link{make}()}.
}
\seealso{
\code{\link{built}}, \code{\link{imported}},
\code{\link{readd}}, \code{\link{plan}}, \code{\link{make}}
}
