% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{shell_file}
\alias{shell_file}
\title{Shell file for Makefile parallelism
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}}
\usage{
shell_file(path = "shell.sh", overwrite = FALSE)
}
\arguments{
\item{path}{Character.}

\item{overwrite}{Logical.}
}
\value{
logical
}
\description{
2019-01-03
}
\keyword{internal}
