% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-deprecate.R
\name{shell_file}
\alias{shell_file}
\title{Deprecated}
\usage{
shell_file(path = "shell.sh", overwrite = FALSE)
}
\arguments{
\item{path}{Character.}

\item{overwrite}{Logical.}
}
\value{
logical
}
\description{
2019-01-03
}
\examples{
# deprecated
}
\keyword{internal}
