% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecate.R
\name{shell_file}
\alias{shell_file}
\title{Deprecated}
\usage{
shell_file(path = "shell.sh", overwrite = FALSE)
}
\arguments{
\item{path}{character}

\item{overwrite}{logical}
}
\value{
logical
}
\description{
2019-01-03
}
\examples{
# deprecated
}
\keyword{internal}
