% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parallel_ui.R
\name{shell_file}
\alias{shell_file}
\title{Function shell_file}
\usage{
shell_file(path = "shell.sh", overwrite = FALSE)
}
\arguments{
\item{path}{file path of the shell file}

\item{overwrite}{logical, whether to overwrite a possible
destination file with the same name}
}
\description{
Write an example \code{shell.sh} file required by
\code{make(..., parallelism = 'Makefile', prepend = 'SHELL=./shell.sh')}
and do a `chmod +x` to enable execution.
Use this option to run your project in parallel on a computing cluster
or supercomputer.
}
\examples{
\dontrun{
# Write shell.sh to your working directory.
# Read the parallelism vignette to learn how it is used
# in Makefile parallelism.
shell_file()
}
}
\seealso{
\code{\link{make}}, \code{\link{max_useful_jobs}},
\code{\link{parallelism_choices}}
}
