% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parallel_ui.R
\name{shell_file}
\alias{shell_file}
\title{Write an example \code{shell.sh} file required by
\code{make(..., parallelism = 'Makefile', prepend = 'SHELL=./shell.sh')}.}
\usage{
shell_file(path = "shell.sh", overwrite = FALSE)
}
\arguments{
\item{path}{file path of the shell file}

\item{overwrite}{logical, whether to overwrite a possible
destination file with the same name}
}
\value{
The return value of the call to \code{\link[=file.copy]{file.copy()}} that
wrote the shell file.
}
\description{
This function also does a \code{chmod +x}
to enable execute permissions.
}
\examples{
\dontrun{
test_with_dir("Quarantine side effects.", {
# Write shell.sh to your working directory.
# Read the high-performance computing chapter
# (https://ropenscilabs.github.io/drake-manual/hpc.html)
# to learn how it is used
# in Makefile parallelism.
shell_file()
})
}
}
\seealso{
\code{\link[=make]{make()}}, \code{\link[=max_useful_jobs]{max_useful_jobs()}},
\code{\link[=parallelism_choices]{parallelism_choices()}}, \code{\link[=drake_hpc_template_file]{drake_hpc_template_file()}},
\code{\link[=drake_example]{drake_example()}}, \code{\link[=drake_examples]{drake_examples()}}
}
