% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{short_hash}
\alias{short_hash}
\title{\code{drake} now only uses one hash algorithm per cache.
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}}
\usage{
short_hash(cache = drake::get_cache(verbose = verbose), verbose = 1L)
}
\arguments{
\item{cache}{drake cache. See \code{\link[=new_cache]{new_cache()}}.
If supplied, \code{path} is ignored.}

\item{verbose}{Deprecated on 2019-09-11.}
}
\value{
A character vector naming a hash algorithm.
}
\description{
Deprecated on 2018-12-12.
}
\keyword{internal}
