% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache.R
\name{short_hash}
\alias{short_hash}
\title{Function short_hash}
\usage{
short_hash(cache)
}
\arguments{
\item{cache}{drake cache}
}
\description{
Get the short hash algorithm of a drake cache.
}
\details{
See \code{?\link{default_long_hash_algo}()}
}
\examples{
\dontrun{
load_basic_example()
config <- make(my_plan, return_config = TRUE)
cache <- config$cache
short_hash(cache)
}
}
\seealso{
\code{\link{default_short_hash_algo}},
\code{\link{default_long_hash_algo}}
}
