% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/console.R
\name{show_source}
\alias{show_source}
\title{Show how a target/import was produced.}
\usage{
show_source(target, config, character_only = FALSE)
}
\arguments{
\item{target}{symbol denoting the target or import
or a character vector if character_only is \code{TRUE}.}

\item{config}{a \code{\link[=drake_config]{drake_config()}} list}

\item{character_only}{logical, whether to interpret
\code{target} as a symbol (\code{FALSE}) or character vector
(\code{TRUE}).}
}
\description{
Show the command that produced a target
or indicate that the object or file was imported.
}
\examples{
\dontrun{
plan <- drake_plan(x = rnorm(15))
make(plan)
config <- drake_config(plan)
show_source(x, config)
show_source(rnorm, config)
}
}
