% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hooks.R
\name{silencer_hook}
\alias{silencer_hook}
\title{Function silencer_hook}
\usage{
silencer_hook(code)
}
\arguments{
\item{code}{code to run to build the target.}
}
\description{
an example \code{hook} argument to
\code{make()} that redirects output and error messages
}
\examples{
\dontrun{
silencer_hook({
  cat(1234)
  stop(5678)
})
x <- workplan(loud = cat(1234), bad = stop(5678))
make(x, hook = silencer_hook)
}
}
\seealso{
\code{\link{make}}, \code{\link{message_sink_hook}},
\code{\link{output_sink_hook}}
}
