% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hooks.R
\name{silencer_hook}
\alias{silencer_hook}
\title{Function \code{silencer_hook}}
\usage{
silencer_hook(code)
}
\arguments{
\item{code}{code to run to build the target.}
}
\value{
A function that you can supply to the \code{hook} argument
of \code{\link{make}()}.
}
\description{
an example \code{hook} argument to
\code{make()} that redirects output and error messages
}
\examples{
\dontrun{
test_with_dir("Quarantine side effects.", {
# Test out the silencer hook on its own.
try(
  silencer_hook({
    cat(1234)
    stop(5678)
  }),
  silent = FALSE
)
# Make a new workflow plan.
x <- plan_drake(loud = cat(1234), bad = stop(5678))
# Test out the silencer hook on a drake project.
# All output should be suppressed.
try(make(x, hook = silencer_hook), silent = FALSE)
})
}
}
\seealso{
\code{\link{make}}, \code{\link{message_sink_hook}},
\code{\link{output_sink_hook}}
}
