% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache.R
\name{status}
\alias{status}
\title{Function \command{status}}
\usage{
status(path = getwd(), search = FALSE)
}
\arguments{
\item{path}{Root directory of the drake project,
or if \command{search} is \command{TRUE}, either the
project root or a subdirectory of the project.}

\item{search}{If \command{TRUE}, search parent directories
to find the nearest drake cache. Otherwise, look in the
current working directory only.}
}
\value{
data frame containing the build status
of the last session
}
\description{
Get the build status of the last call to 
\command{\link{run}()}: which objects were skipped, built,
and imported.
}
\seealso{
\command{\link{session}},
\command{\link{built}}, \command{\link{imported}},
\command{\link{readd}}, 
\command{\link{plan}}, \command{\link{run}}
}

