% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{store_drake_config}
\alias{store_drake_config}
\title{Store an internal configuration list
from \code{\link{drake_config}()}.}
\usage{
store_drake_config(config)
}
\arguments{
\item{config}{Internal configuration list}
}
\value{
Nothing.
}
\description{
Exported for demonstration and tinkering purposes
only. Not meant to be called by the user.
}
\examples{
\dontrun{
test_with_dir("Quarantine side effects.", {
load_basic_example()
config <- drake_config(my_plan)
store_drake_config(config = config)
read_drake_config()
})
}
}
\keyword{internal}
