% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/defunct.R
\name{summaries}
\alias{summaries}
\title{summaries \ifelse{html}{\figure{lifecycle-defunct.svg}{options: alt='Defunct lifecycle'}}{\strong{Defunct}}}
\usage{
summaries(...)
}
\arguments{
\item{...}{Arguments}
}
\description{
2019-02-15
}
\keyword{internal}
