% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate.R
\name{summaries}
\alias{summaries}
\title{Function \code{summaries}}
\usage{
summaries(commands, analyses, datasets, gather = rep("list",
  dim(commands)[1]))
}
\arguments{
\item{commands}{Data frame output of the \code{\link{workflow}} function.}

\item{analyses}{Data frame of \code{skipthat} commands to generate analyses.}

\item{datasets}{Data frame of \code{skipthat} commands to generate datasets.}

\item{gather}{Character vector, names of functions to gather the summaries.
If not \code{NULL}, length must be the number of rows in the \code{workflow}
argument.}
}
\value{
preprocessed data frame of \code{skipthat} summary commands
}
\description{
Produces a data frame of \code{skipthat} summary commands.
}

