% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate.R
\name{summaries}
\alias{summaries}
\title{Function \code{summaries}}
\usage{
summaries(plan, analyses, datasets, gather = rep("list", nrow(plan)))
}
\arguments{
\item{plan}{workflow plan data frame with commands for the summaries.
Use the \code{..analysis..} and \code{..dataset..} wildcards
just like the \code{..dataset..} wildcard in \code{\link{analyses}()}.}

\item{analyses}{workflow plan data frame of analysis instructions}

\item{datasets}{workflow plan data frame with instructions to make
or import the datasets.}

\item{gather}{Character vector, names of functions to gather the
summaries. If not \code{NULL}, the length must be the number of
rows in the \code{plan}. See the \code{\link{gather}()} function
for more.}
}
\value{
an evaluated workflow plan data frame of instructions
for computing summaries of analyses and datasets.
analyses of multiple datasets in multiple ways.
}
\description{
Generate a workflow plan data frame for summarizing
multiple analyses of multiple datasets multiple ways.
}
\examples{
# Create the part of the workflow plan data frame for the datasets.
datasets <- workplan(
  small = simulate(5),
  large = simulate(50))
# Create a template workflow plan containing the analysis methods.
methods <- workplan(
  regression1 = reg1(..dataset..),
  regression2 = reg2(..dataset..))
# Generate the part of the workflow plan to analyze the datasets.
analyses <- analyses(methods, datasets = datasets)
# Create a template workflow plan dataset with the
# types of summaries you want.
summary_types <- workplan(
  summ = summary(..analysis..),
  coef = coefficients(..analysis..))
# Evaluate the appropriate wildcards to encode the summary targets.
summaries(summary_types, analyses, datasets, gather = NULL)
summaries(summary_types, analyses, datasets)
summaries(summary_types, analyses, datasets, gather = "list")
summs <- summaries(
  summary_types, analyses, datasets, gather = c("list", "rbind"))
# For the final workflow plan, row bind the pieces together.
my_plan <- rbind(datasets, analyses, summs)
my_plan
}
\seealso{
\code{\link{analyses}}, \code{\link{make}}, \code{\link{workplan}}
}
