% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate.R
\name{summaries}
\alias{summaries}
\title{Function \code{summaries}}
\usage{
summaries(plan, analyses, datasets, gather = rep("list", dim(plan)[1]))
}
\arguments{
\item{plan}{data frame of instructions for summaries (wildcards unevaluated)}

\item{analyses}{data frame of analysis instructions}

\item{datasets}{data frame of dataset instructions}

\item{gather}{Character vector, names of functions to gather the summaries.
If not \code{NULL}, length must be the number of rows in the \code{plan}
argument.}
}
\value{
portion of a \code{drake} plan for summarizing multiple
analyses of multiple datasets in multiple ways
}
\description{
Generate a portion of a \code{drake} plan
for summarizing multiple anlayses of multiple datasets in multiple ways.
}
\seealso{
\code{\link{analyses}},
 \code{\link{run}}, \code{\link{plan}}
}

