% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate.R
\name{summaries}
\alias{summaries}
\title{Function \code{summaries}}
\usage{
summaries(plan, analyses, datasets, gather = rep("list", nrow(plan)))
}
\arguments{
\item{plan}{workflow plan data frame with commands for the summaries.
Use the \code{..analysis..} and \code{..dataset..} wildcards 
just like the \code{..dataset..} wildcard in \code{\link{analyses}()}.}

\item{analyses}{workflow plan data frame of analysis instructions}

\item{datasets}{workflow plan data frame with instructions to make
or import the datasets.}

\item{gather}{Character vector, names of functions to gather the 
summaries. If not \code{NULL}, the length must be the number of 
rows in the \code{plan}. See the \code{\link{gather}()} function
for more.}
}
\value{
an evaluated workflow plan data frame of instructions
for computing summaries of analyses and datasets.
analyses of multiple datasets in multiple ways.
}
\description{
Generate a worklfow plan data frame for summarizing
multiple analyses of multiple datasets multiple ways.
}
\seealso{
\code{\link{analyses}}, \code{\link{run}}, \code{\link{plan}}
}

