% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecate.R
\name{summaries}
\alias{summaries}
\title{Deprecated function \code{summaries}}
\usage{
summaries(plan, analyses, datasets, gather = rep("list", nrow(plan)))
}
\arguments{
\item{plan}{Same as for \code{\link{plan_summaries}()}.}

\item{analyses}{Same as for \code{\link{plan_summaries}()}.}

\item{datasets}{Same as for \code{\link{plan_summaries}()}.}

\item{gather}{Same as for \code{\link{plan_summaries}()}.}
}
\value{
Same as for \code{\link{plan_summaries}()}.
}
\description{
Use \code{\link{summaries}()} instead
}
\details{
Deprecated on 2017-11-12.
}
\examples{
# See ?drake_session for examples.
}
\seealso{
\code{\link{summaries}}
}
\keyword{internal}
