% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api-plan.R
\name{target}
\alias{target}
\title{Define custom columns in a \code{\link[=drake_plan]{drake_plan()}}.}
\usage{
target(command = NULL, ...)
}
\arguments{
\item{command}{The command to build the target.}

\item{...}{Named arguments specifying non-standard
fields of the workflow plan.}
}
\value{
A one-row workflow plan data frame with the named
arguments as columns.
}
\description{
The \code{target()} function lets you define
custom columns in a workflow plan data frame, both
inside and outside calls to \code{\link[=drake_plan]{drake_plan()}}.
}
\details{
Tidy evaluation is applied to the arguments,
and the \code{!!} operator is evaluated immediately
for expressions and language objects.
}
\examples{
# Use target() to create your own custom columns in a drake plan.
# See ?triggers for more on triggers.
plan <- drake_plan(
  website_data = target(
    download_data("www.your_url.com"),
    trigger = "always",
    custom_column = 5
  ),
  analysis = analyze(website_data)
)
plan
# make(plan) # nolint
# Call target() inside or outside drake_plan().
target(
  download_data("www.your_url.com"),
  trigger = "always",
  custom_column = 5
)
}
\seealso{
\code{\link[=drake_plan]{drake_plan()}}, \code{\link[=make]{make()}}
}
