% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api-plan.R
\name{target}
\alias{target}
\title{Define custom columns in a \code{\link[=drake_plan]{drake_plan()}}.}
\usage{
target(command = NULL, ...)
}
\arguments{
\item{command}{The command to build the target.}

\item{...}{Named arguments specifying non-standard
fields of the workflow plan.}
}
\value{
A one-row workflow plan data frame with the named
arguments as columns.
}
\description{
Not a user-side function. Please use from within
\code{\link[=drake_plan]{drake_plan()}} only.
}
\examples{
# Use target() to create your own custom columns in a drake plan.
# See ?triggers for more on triggers.
drake_plan(
  website_data = target(
    download_data("www.your_url.com"),
    trigger = "always",
    custom_column = 5
  ),
  analysis = analyze(website_data)
)
}
\seealso{
\code{\link[=drake_plan]{drake_plan()}}, \code{\link[=make]{make()}}
}
\keyword{internal}
