% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{this_cache}
\alias{this_cache}
\title{Get the cache at the exact file path specified.
\lifecycle{deprecated}}
\usage{
this_cache(path = NULL, force = FALSE, verbose = 1L,
  fetch_cache = NULL, console_log_file = NULL)
}
\arguments{
\item{path}{File path of the cache.}

\item{force}{Deprecated.}

\item{verbose}{Integer, control printing to the console/terminal.
\itemize{
\item \code{0}: print nothing.
\item \code{1}: print targets, retries, and failures.
\item \code{2}: also show a spinner when preprocessing tasks are underway.
}}

\item{fetch_cache}{Deprecated.}

\item{console_log_file}{Optional character scalar of a file name or
connection object (such as \code{stdout()}) to dump maximally verbose
log information for \code{\link[=make]{make()}} and other functions (all functions that
accept a \code{config} argument, plus \code{drake_config()}).
If you choose to use a text file as the console log,
it will persist over multiple function calls
until you delete it manually.
Fields in each row the log file, from left to right:
- The node name (short host name) of the
computer (from \code{Sys.info()["nodename"]}).
- The process ID (from \code{Sys.getpid()}).
- A timestamp with the date and time (in microseconds).
- A brief description of what \code{drake} was doing.\code{ The fields are separated by pipe symbols (}"|"`).}
}
\value{
A drake/storr cache at the specified path, if it exists.
}
\description{
This function does not apply to
in-memory caches such as \code{storr_environment()}.
}
\keyword{internal}
