% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{this_cache}
\alias{this_cache}
\title{Get the cache at the exact file path specified.
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}}
\usage{
this_cache(
  path = NULL,
  force = FALSE,
  verbose = 1L,
  fetch_cache = NULL,
  console_log_file = NULL
)
}
\arguments{
\item{path}{File path of the cache.}

\item{force}{Deprecated.}

\item{verbose}{Deprecated on 2019-09-11.}

\item{fetch_cache}{Deprecated.}

\item{console_log_file}{Deprecated in favor of \code{log_make}.}
}
\value{
A drake/storr cache at the specified path, if it exists.
}
\description{
This function does not apply to
in-memory caches such as \code{storr_environment()}.
}
\keyword{internal}
