% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache.R
\name{this_cache}
\alias{this_cache}
\title{Function this_cache}
\usage{
this_cache(path = NULL)
}
\arguments{
\item{path}{file path of the cache}
}
\description{
Get a specific drake cache
at the exact specified file path
}
\examples{
\dontrun{
x <- this_cache() # Does not exist yet
load_basic_example()
make(my_plan)
y <- this_cache()
z <- this_cache(".drake") # same as above
manual <- new_cache("manual_cache")
manual2 <- get_cache("manual_cache") # same as above
}
}
