% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timestamp.R
\name{time_stamps}
\alias{time_stamps}
\title{Internal function time_stamps}
\usage{
time_stamps(config)
}
\arguments{
\item{config}{Internal master configuration list
produced by \code{\link{drake_config}}.}
}
\value{
nothing
}
\description{
Write dummy timestamps for Makefiles
}
\details{
Used internally with
\code{\link{make}(..., parallelism = "Makefile")}
}
\examples{
\dontrun{
load_basic_example()
config <- drake_config(my_plan) # Master internal configuration list
time_stamps(config)
# Now look in '.drake/ts' for dummy timestamp files.
} 
}
\keyword{internal}
