% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph.R
\name{tracked}
\alias{tracked}
\title{Function \code{tracked}}
\usage{
tracked(plan = workplan(), targets = drake::possible_targets(plan),
  envir = parent.frame(), jobs = 1, verbose = TRUE, trigger = "any")
}
\arguments{
\item{plan}{workflow plan data frame, same as for function
\code{\link{make}()}.}

\item{targets}{names of targets to build, same as for function
\code{\link{make}()}.}

\item{envir}{environment to import from, same as for function
\code{\link{make}()}.}

\item{jobs}{number of jobs to accelerate the construction
of the dependency graph. A light \code{mclapply}-based
parallelism is used if your operating system is not Windows.}

\item{verbose}{logical, whether to print
progress messages to the console.}

\item{trigger}{Name of the trigger to apply to all targets.
Ignored if \code{plan} has a \code{trigger} column.
Must be in \code{\link{triggers}()}.
See \code{?triggers} for explanations of the choices.}
}
\description{
Print out which objects, functions, files, targets, etc.
are reproducibly tracked.
}
\examples{
\dontrun{
load_basic_example()
tracked(my_plan)
}
}
