% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dependencies.R
\name{tracked}
\alias{tracked}
\title{List the targets and imports
that are reproducibly tracked.}
\usage{
tracked(config)
}
\arguments{
\item{config}{An output list from \code{\link[=drake_config]{drake_config()}}.}
}
\value{
A character vector with the names of reproducibly-tracked targets.
}
\description{
In other words, list all the ordinances
in your project's dependency network.
}
\examples{
\dontrun{
test_with_dir("Quarantine side effects.", {
load_mtcars_example() # Load the canonical example for drake.
# List all the targets/imports that are reproducibly tracked.
config <- drake_config(my_plan)
tracked(config)
})
}
}
