% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph.R
\name{tracked}
\alias{tracked}
\title{Function \code{tracked}}
\usage{
tracked(plan, targets = plan$target, envir = parent.frame())
}
\arguments{
\item{plan}{workflow plan data frame, same as for function
\code{\link{make}()}.}

\item{targets}{names of targets to bulid, same as for function
\code{\link{make}()}.}

\item{envir}{environment to import from, same as for function
\code{\link{make}()}.}
}
\description{
Print out which objects, functions, files, targets, etc.
are reproducibly tracked.
}
