% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api-transform.R
\name{transform_plan}
\alias{transform_plan}
\title{Transform a plan}
\usage{
transform_plan(plan, trace = FALSE)
}
\arguments{
\item{plan}{Workflow plan data frame with a column for targets,
a column for commands, and a column for transformations.}

\item{trace}{Logical, whether to add columns to show
what happened during target transformations, e.g.
\code{drake_plan(x = target(..., transform = ...), transform = TRUE)}.}
}
\value{
A transformed workflow plan data frame
}
\description{
Take an existing \code{drake} plan with a \code{transform} column,
and apply transformations that expand and gather targets.
Use to generate large plans. \code{transform_plan()} is useful
if you generated multiple plans with \code{drake_plan(transform = FALSE)}
and and want to combine and transform them later.
}
\examples{
plan1 <- drake_plan(
  analysis = target(
    analyze_data("source"),
    transform = cross(source = c(source1, source2))
  ),
  transform = FALSE
)
plan2 <- drake_plan(
  summarize = target(
    summarize_analyses(analysis),
    transform = summarize()
  ),
  transform = FALSE
)
plan <- bind_plans(plan1, plan2)
plan
transform_plan(plan)
transform_plan(plan, trace = TRUE)
}
\seealso{
\code{\link[=drake_plan]{drake_plan()}}
}
