% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{triggers}
\alias{triggers}
\title{List the old drake triggers.
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}}
\usage{
triggers()
}
\value{
A character vector with the names of the old triggers.
}
\description{
Triggers are target-level rules
that tell \code{\link[=make]{make()}} how to know if a target
is outdated or up to date.
}
\details{
Deprecated on 2018-07-22.
}
\keyword{internal}
