% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecate.R
\name{triggers}
\alias{triggers}
\title{Deprecated. List the old drake triggers.}
\usage{
triggers()
}
\value{
A character vector with the names of the old triggers.
}
\description{
Triggers are target-level rules
that tell \code{\link[=make]{make()}} how to know if a target
is outdated or up to date.
}
\details{
Deprecated on 2018-07-22.
}
\examples{
# Deprecated. See the trigger() function instead (singular).
}
\keyword{internal}
