% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache_types.R
\name{type_of_cache}
\alias{type_of_cache}
\title{Experimental function \code{type_of_cache}}
\usage{
type_of_cache(path = drake::default_cache_path())
}
\arguments{
\item{path}{Path to the file system cache.}
}
\value{
Type of the cache at the specified path.
}
\description{
Try to get the type of a drake
file system cache. Only works on known caches
with known file systems.
}
\details{
Experimental function for a possible
new future features.
}
\examples{
\dontrun{
test_with_dir("Quarantine side effects.", {
load_basic_example() # Load drake's canonical example.
make(my_plan) # Run the project, build the targets.
# Get the type of the cache
# located in the (default) '.drake' folder.
type_of_cache(".drake") # NULL, signaling "storr_rds"
})
}
}
