% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drake_plan.R
\name{type_sum.expr_list}
\alias{type_sum.expr_list}
\title{Type summary printing
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#questioning}{\figure{lifecycle-questioning.svg}{options: alt='[Questioning]'}}}{\strong{[Questioning]}}}
\usage{
type_sum.expr_list(x)
}
\arguments{
\item{x}{List of language objects.}
}
\description{
Ensures \verb{<expr>} is printed at the top
of any \code{drake} plan column that is a list of language objects
(e.g. \code{plan$command}).
}
\keyword{internal}
