% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api-plan.R
\name{type_sum.expr_list}
\alias{type_sum.expr_list}
\title{Type summary printing}
\usage{
type_sum.expr_list(x)
}
\arguments{
\item{x}{List of language objects.}
}
\description{
Ensures \code{<expr>} is printed at the top
of any \code{drake} plan column that is a list of language objects
(e.g. \code{plan$command}).
}
\keyword{internal}
