% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plan.R
\name{workflow}
\alias{workflow}
\title{Function \code{workflow}}
\usage{
workflow(..., list = character(0), file_targets = FALSE,
  strings_in_dots = c("filenames", "literals"))
}
\arguments{
\item{...}{same as for \code{drake::\link{plan}()}}

\item{list}{same as for \code{drake::\link{plan}()}}

\item{file_targets}{same as for \code{drake::\link{plan}()}}

\item{strings_in_dots}{same as for \code{drake::\link{plan}()}}
}
\value{
data frame of targets and command
}
\description{
An alternative to \code{drake::\link{plan}()}
that does not conflict with \code{future::plan()}.
}
\examples{
workflow(small = simulate(5), large = simulate(50))
workflow(list = c(x = "1 + 1", y = "sqrt(x)"))
workflow(data = readRDS("my_data.rds"))
workflow(my_file.rds = saveRDS(1+1, "my_file.rds"), file_targets = TRUE,
  strings_in_dots = "literals")
}
\seealso{
\code{\link{check}}, \code{\link{make}},
\code{\link{plan}}
}
