% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecate.R
\name{workplan}
\alias{workplan}
\title{Deprecated function \code{workplan}}
\usage{
workplan(..., list = character(0), file_targets = FALSE,
  strings_in_dots = c("filenames", "literals"))
}
\arguments{
\item{...}{same as for \code{drake::\link{plan_drake}()}}

\item{list}{same as for \code{drake::\link{plan_drake}()}}

\item{file_targets}{same as for \code{drake::\link{plan_drake}()}}

\item{strings_in_dots}{same as for \code{drake::\link{plan_drake}()}}
}
\value{
A data frame of targets and commands.
}
\description{
Turns a named collection of command/target pairs into
a workflow plan data frame for \code{\link{make}} and
\code{\link{check}}.
}
\details{
Deprecated on 2017-11-29
}
\examples{
# See ?plan_drake for examples.
}
\keyword{internal}
